% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ord-accessors.r
\name{accessors}
\alias{accessors}
\alias{recover_rows}
\alias{recover_cols}
\alias{recover_factor}
\alias{recover_rows.default}
\alias{recover_cols.default}
\alias{recover_rows.data.frame}
\alias{recover_cols.data.frame}
\alias{get_rows}
\alias{get_cols}
\alias{get_factor}
\alias{as.matrix.tbl_ord}
\alias{recover_inertia}
\alias{recover_inertia.default}
\alias{recover_coord}
\alias{recover_coord.default}
\alias{recover_coord.data.frame}
\alias{get_coord}
\alias{dim.tbl_ord}
\alias{dim_rows}
\alias{dim_cols}
\alias{dim_factor}
\title{Access factors, coordinates, and metadata from ordination objects}
\usage{
recover_rows(x)

recover_cols(x)

recover_factor(x, .matrix)

\method{recover_rows}{default}(x)

\method{recover_cols}{default}(x)

\method{recover_rows}{data.frame}(x)

\method{recover_cols}{data.frame}(x)

get_rows(x, .supplement = TRUE, align = TRUE)

get_cols(x, .supplement = TRUE, align = TRUE)

get_factor(x, .matrix, .supplement = TRUE, align = TRUE)

\method{as.matrix}{tbl_ord}(x, ..., .matrix, .supplement = TRUE, align = TRUE)

recover_inertia(x)

\method{recover_inertia}{default}(x)

recover_coord(x)

\method{recover_coord}{default}(x)

\method{recover_coord}{data.frame}(x)

get_coord(x, align = TRUE)

\method{dim}{tbl_ord}(x)

dim_rows(x)

dim_cols(x)

dim_factor(x, .matrix)
}
\arguments{
\item{x}{A \link{tbl_ord}, or an ordination object coercible to one.}

\item{.matrix}{A character string matching one of several indicators for one
or both matrices in a matrix decomposition used for ordination. The
standard values are \code{"rows"}, \code{"cols"}, and \code{"dims"} (for both).}

\item{.supplement}{Logical; whether to include}

\item{align}{Logical; whether to align the matrix factors and coordinates
according to an \code{"align"} matrix attribute.}

\item{...}{Additional arguments from \code{\link[base:matrix]{base::as.matrix()}}; ignored.}
}
\description{
These functions return information about the matrix
factorization underlying an ordination.
}
\details{
The \verb{recover_*()} functions extract one or both of the row and column matrix
factors that constitute the original ordination. These are interpreted as the
case scores (rows) and the variable loadings (columns). The \verb{get_*()}
functions optionally (and by default) apply any alignment stored as the
\code{"align"} attribute (see \link{alignment}). Only the \verb{recover_*()} functions are
generics that require methods for each ordination class.

\code{get_coord()} retrieves the names of the coordinates shared by the matrix
factors on which the original data were ordinated, and \code{dim.tbl_ord()}
retrieves their number, the rank of the ordination. The outer dimensions of
the matrix decomposition are returned by \code{dim_rows()} and \code{dim_cols()}.
}
\examples{
# Multidimensional scaling of country differences and regression of attributes
# Reproduce Exhibit 4.2 in Greenacre (2010)
data(country_differences)
country_differences \%>\%
  cmdscale_ord(k = 2) \%>\%
  as_tbl_ord() \%>\%
  print() -> differences_cmds
differences_plot <- differences_cmds \%>\%
  ggbiplot(aes(x = 1, y = 2, label = .name)) +
  geom_cols_text()
differences_plot
# Reproduce Exhibit 4.5 in Greenacre (2010)
data(country_attributes)
lm(country_attributes ~ get_rows(differences_cmds)) \%>\%
  as_tbl_ord() \%>\%
  print() -> attributes_fit
differences_plot +
  geom_cols_vector(data = attributes_fit) +
  geom_cols_text_radiate(data = attributes_fit, hjust = .3)
# Confer PCA weights and inertia between cases and variables
data(country_attributes)
country_attributes \%>\%
  prcomp() \%>\%
  as_tbl_ord() \%>\%
  print() -> attributes_pca
# form biplot
attributes_pca \%>\%
  ggbiplot(
    aes(label = .name),
    sec.axes = "cols", scale.factor = dim_cols(attributes_pca)
  ) +
  theme_bw() + theme(panel.grid = element_blank()) +
  geom_rows_text(aes(x = PC1, y = PC2), color = "seagreen") +
  geom_cols_vector(aes(x = PC1, y = PC2), color = "darkred") +
  geom_cols_text(
    aes(x = PC1, y = PC2),
    color = "darkred", hjust = "outward", vjust = "outward"
  )
# covariance biplot
attributes_pca \%>\%
  confer_inertia(0) \%>\%
  ggbiplot(
    aes(label = .name),
    sec.axes = "rows", scale.factor = dim_rows(attributes_pca)
  ) +
  theme_bw() + theme(panel.grid = element_blank()) +
  geom_rows_text(aes(x = PC1, y = PC2), color = "seagreen") +
  geom_cols_vector(aes(x = PC1, y = PC2), color = "darkred") +
  geom_cols_text_repel(aes(x = PC1, y = PC2), color = "darkred")
}
