% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ord-augmentation.r
\name{augmentation}
\alias{augmentation}
\alias{augmentation_rows}
\alias{augmentation_cols}
\alias{augmentation_factor}
\alias{augmentation.tbl_ord}
\alias{augmentation_coord}
\alias{augment.tbl_ord}
\alias{augment_rows}
\alias{augment_cols}
\alias{tidy.tbl_ord}
\title{Augment metadata on the factors and coordinates of tbl_ords}
\usage{
augmentation_rows(x)

augmentation_cols(x)

augmentation_factor(x, .matrix)

augmentation.tbl_ord(x, .matrix)

augmentation_coord(x)

\method{augment}{tbl_ord}(x, data, .matrix = "dims", ...)

augment_rows(x)

augment_cols(x)

\method{tidy}{tbl_ord}(x, ...)
}
\arguments{
\item{x}{A '\link{tbl_ord}' object, or an ordination object coercible to one.}

\item{.matrix}{A character string matching one of several indicators for one
or both matrices in a matrix decomposition used for ordination. The
standard values are \code{"rows"}, \code{"cols"}, and \code{"dims"} (for both).}

\item{data}{Passed to \code{\link[generics:augment]{generics::augment()}}; currently ignored.}

\item{...}{Additional arguments allowed by generics; ignored.}
}
\description{
These functions return data associated with the cases,
variables, and coordinates of an ordination object, and attach it to the
object.
}
\details{
The \verb{augmentation_*()} methods produce \link[tibble:tibble]{tibble}s of values
associated with the rows, columns, and coordinates of a '\link{tbl_ord}' object.
The first field of each tibble is \code{.name}, which contains the case, variable,
or coordinate names. Additional fields contain information about the cases,
variables, or coordinates extracted from the original ordination object.

The \verb{augment_*()} functions return the ordination with either or both factors
annotated with the result of \verb{augmentation_*()}. In this way \verb{augment_*()}
works like \code{\link[generics:augment]{generics::augment()}} by extracting information for a tidy summary
of the components, but it differs in returning an annotated tbl_ord rather
than a \link[tibble:tbl_df-class]{'tbl_df'}. The advantage of implementing separate
methods for the different components is that more information contained in
the original object becomes accessible to the user.

Three generics popularized by the \strong{ggplot2} and \strong{broom} packages make use
of these functions:
\itemize{
\item The \code{\link[generics:augment]{generics::augment()}} method adds information about
the rows and columns while maintaining 'tbl_ord' class structure.
\item The \code{\link[generics:tidy]{generics::tidy()}} method summarizes information about
model components, which here are the artificial coordinates
created by ordinations.
\item The \code{\link[ggplot2:fortify]{ggplot2::fortify()}} method (\code{\link[=fortify.tbl_ord]{fortify.tbl_ord()}}) augments and
collapses row and/or column data into a single tibble.
}
}
