% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibble-annotation.r, R/bibble-cmds.r,
%   R/bibble-lm.r, R/bibble-lpca.r
\name{bibble-annotation}
\alias{bibble-annotation}
\alias{u_annot}
\alias{v_annot}
\alias{factor_annot}
\alias{coord_annot}
\alias{u_annot.cmds}
\alias{v_annot.cmds}
\alias{coord_annot.cmds}
\alias{u_annot.lm}
\alias{v_annot.lm}
\alias{coord_annot.lm}
\alias{u_annot.mlm}
\alias{v_annot.mlm}
\alias{coord_annot.mlm}
\alias{u_annot.lpca}
\alias{v_annot.lpca}
\alias{coord_annot.lpca}
\alias{u_annot.lsvd}
\alias{v_annot.lsvd}
\alias{coord_annot.lsvd}
\title{Recover annotation for subjects, variables, or coordinates}
\usage{
u_annot(x)

v_annot(x)

factor_annot(x, .matrix)

coord_annot(x)

\method{u_annot}{cmds}(x)

\method{v_annot}{cmds}(x)

\method{coord_annot}{cmds}(x)

\method{u_annot}{lm}(x)

\method{v_annot}{lm}(x)

\method{coord_annot}{lm}(x)

\method{u_annot}{mlm}(x)

\method{v_annot}{mlm}(x)

\method{coord_annot}{mlm}(x)

\method{u_annot}{lpca}(x)

\method{v_annot}{lpca}(x)

\method{coord_annot}{lpca}(x)

\method{u_annot}{lsvd}(x)

\method{v_annot}{lsvd}(x)

\method{coord_annot}{lsvd}(x)
}
\arguments{
\item{x}{A bibble, or an object convertible to one.}

\item{.matrix}{A character string matching one of several indicators for one
or both matrices in a matrix decomposition used for ordination. The
standard values are \code{"u"}, \code{"v"}, and \code{"uv"}.}
}
\description{
Recover annotation for subjects, variables, or coordinates
}
