% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibble-augmentation.r
\name{bibble-augmentation}
\alias{bibble-augmentation}
\alias{augment_u}
\alias{augment_v}
\alias{augment_factor}
\alias{augment.bbl}
\alias{augment_coord}
\alias{annotate_u}
\alias{annotate_v}
\alias{u_annot}
\alias{v_annot}
\alias{set_u_annot}
\alias{set_v_annot}
\title{Augment information for the factors and coordinates of an ordination object}
\usage{
augment_u(x)

augment_v(x)

augment_factor(x, .matrix)

\method{augment}{bbl}(x, .matrix)

augment_coord(x)

annotate_u(x)

annotate_v(x)

u_annot(x)

v_annot(x)

set_u_annot(x, annot)

set_v_annot(x, annot)
}
\arguments{
\item{x}{A bibble, or an ordination object coercible to one.}

\item{.matrix}{A character string matching one of several indicators for one
or both matrices in a matrix decomposition used for ordination. The
standard values are \code{"u"}, \code{"v"}, and \code{"uv"}.}

\item{annot}{A \code{\link[tibble]{tibble}} having the same number of rows as
\code{x}.}
}
\description{
These functions return data associated with the cases, variables, and
coordinates of an ordination.
}
\details{
The \code{augment_*} functions produce \link[tibble]{tibble}s of data 
augmented to the cases, variables, and coordinates. The first field of each
tibble is \code{.name} and contains the case, variable, and coordinate names,
respectively.

The \code{augment_*} functions return data augmented from two sources:
\enumerate{
  \item Information about the cases, variables, or coordinates contained
        in the original ordination object. In this sense \code{augment_*}
        works like \code{\link[broom]{augment}} to extract all such
        information that can be included in a tidy summary of the components.
        The advantage of implementing separate methods for the different
        components is that more information contained in the original object
        becomes accessible to the user.
  \item Additional information about the components manually added to the
        bibble using \strong{\link[dplyr]{dplyr}} verbs adapted to bibbles.
        \emph{These are not yet implemented.}
}
Once the \strong{dplyr} verbs are implemented, each \code{augment_*} function
will retrieve both sources and combines them using 
\code{link[dplyr]{bind_cols}}.
}
