% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-cmds.r
\name{bibble-cmds}
\alias{bibble-cmds}
\alias{as_bibble.cmds}
\alias{get_u.cmds}
\alias{get_v.cmds}
\alias{get_coord.cmds}
\alias{u_annot.cmds}
\alias{v_annot.cmds}
\alias{coord_annot.cmds}
\alias{negate_to.cmds}
\alias{permute_to.cmds}
\alias{rotate_to.cmds}
\title{Bibble functionality for classical multidimensional scaling objects}
\usage{
\method{as_bibble}{cmds}(x)

\method{get_u}{cmds}(x)

\method{get_v}{cmds}(x)

\method{get_coord}{cmds}(x)

\method{u_annot}{cmds}(x)

\method{v_annot}{cmds}(x)

\method{coord_annot}{cmds}(x)

\method{negate_to}{cmds}(x, y, ..., .matrix)

\method{permute_to}{cmds}(x, y, ..., .matrix)

\method{rotate_to}{cmds}(x, y, ..., .matrix)
}
\arguments{
\item{x}{An ordination object.}

\item{y}{A matrix or an ordination object.}

\item{...}{Additional elements of a custom bibble.}

\item{.matrix}{A character string matching one of several indicators for one
or both matrices in a matrix decomposition used for ordination. The
standard values are \code{"u"}, \code{"v"}, and \code{"uv"}.}
}
\description{
These methods extract data from, and attribute new data to, objects of class 
\code{"cmds"}. (This is a class introduced in this package to identify 
objects returned by \code{\link[stats]{cmdscale}}, which is masked by a
wrapper that adds the class attribute.)
}
\examples{
# reproduce Exhibit 4.5 in Greenacre (2010)

data(country_differences, country_attributes)
# format both tibbles as matrices
x1 <- as.matrix(dplyr::select(country_differences, -Countries))
rownames(x1) <- dplyr::pull(country_differences, Countries)
x2 <- as.matrix(dplyr::select(country_attributes, -Countries))
rownames(x2) <- dplyr::pull(country_differences, Countries)

# multidimensional scaling setup
(m <- cmdscale(x1, k = 2))
(b <- as_bibble(m))
(d <- fortify(b))

# basic multidimensional scaling biplot
gg <- ggbiplot(b, aes(x = PCo1, y = PCo2)) +
  geom_v_text(aes(label = .name))
gg
# default aesthetic assignments
ggbiplot(b) +
  geom_v_text(aes(label = .name))
# numeric aesthetic assignments
ggbiplot(b, aes(x = 2, y = 1)) +
  geom_v_text(aes(label = .name))
# regress the attributes on the plotting dimensions and annotate the biplot
fit <- as_bibble(lm(x2 ~ get_u(b)))
gg +
  geom_v_axis(data = fit, aes(label = .name))
gg +
  geom_v_axis(data = fortify(fit, include = "all"), aes(label = .name))
gg +
  geom_v_axis(data = fortify(fit, "v"), aes(label = .name))

# must add aesthetic `.matrix = NULL` to use new data without `.matrix` column
# (this should be remedied)
dat <- fortify(fit, "v")
dat$.matrix <- NULL
gg +
  geom_v_vector(data = fit) +
  geom_text(data = dat, aes(label = .name, .matrix = NULL))
}
