% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-bibble.r
\name{bibble-formatting}
\alias{bibble-formatting}
\alias{format.bbl}
\alias{print.bbl}
\title{Formatting and printing methods for bibbles}
\usage{
\method{format}{bbl}(x, ..., n = NULL, width = NULL, n_extra = NULL)

\method{print}{bbl}(x, ..., n = NULL, width = NULL, n_extra = NULL)
}
\arguments{
\item{x}{An ordination object.}

\item{...}{Additional arguments.}

\item{n}{Number of rows to show. If \code{NULL}, the default, will print
all rows if less than option \code{tibble.print_max}. Otherwise, will
print \code{tibble.print_min} rows.}

\item{width}{Width of text output to generate. This defaults to NULL, which
means use \code{getOption("tibble.width")} or (if also NULL)
\code{getOption("width")}; the latter displays only the columns that
fit on one screen. You can also set \code{options(tibble.width = Inf)} to
override this default and always print all columns.}

\item{n_extra}{Number of extra columns to print abbreviated information for,
if the width is too small for the entire tibble. If \code{NULL}, the
default, will print information about at most \code{tibble.max_extra_cols}
extra columns.}
}
\description{
The \code{format} and \code{print} methods for class \code{"bbl"} are adapted
from those for class \code{"tbl_df"} in the \strong{\link[tibble]{tibble}} 
package and for class \code{"tbl_graph"} in the 
\strong{\link[tidygraph]{tidygraph}} package.
}
\details{
The \code{format} function is tedius but cannot be easily modularized without
invoking \code{get_*()} and \code{*_annot()} multiple times, thereby
significantly reducing performance.
}
