% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lpca.r
\name{bibble-lpca}
\alias{bibble-lpca}
\alias{as_bibble.lpca}
\alias{get_u.lpca}
\alias{get_v.lpca}
\alias{get_coord.lpca}
\alias{u_annot.lpca}
\alias{v_annot.lpca}
\alias{coord_annot.lpca}
\alias{negate_to.lpca}
\alias{as_bibble.lsvd}
\alias{get_u.lsvd}
\alias{get_v.lsvd}
\alias{get_coord.lsvd}
\alias{u_annot.lsvd}
\alias{v_annot.lsvd}
\alias{coord_annot.lsvd}
\alias{reconstruct.lsvd}
\alias{negate_to.lsvd}
\title{Bibble functionality for logistic principal components analysis and singular 
value decomposition objects}
\usage{
\method{as_bibble}{lpca}(x)

\method{get_u}{lpca}(x)

\method{get_v}{lpca}(x)

\method{get_coord}{lpca}(x)

\method{u_annot}{lpca}(x)

\method{v_annot}{lpca}(x)

\method{coord_annot}{lpca}(x)

\method{negate_to}{lpca}(x, y, ..., .matrix)

\method{as_bibble}{lsvd}(x)

\method{get_u}{lsvd}(x)

\method{get_v}{lsvd}(x)

\method{get_coord}{lsvd}(x)

\method{u_annot}{lsvd}(x)

\method{v_annot}{lsvd}(x)

\method{coord_annot}{lsvd}(x)

\method{reconstruct}{lsvd}(x)

\method{negate_to}{lsvd}(x, y, ..., .matrix)
}
\arguments{
\item{x}{An ordination object.}

\item{y}{A matrix or an ordination object.}

\item{...}{Additional elements of a custom bibble.}

\item{.matrix}{A character string matching one of several indicators for one
or both matrices in a matrix decomposition used for ordination. The
standard values are \code{"u"}, \code{"v"}, and \code{"uv"}.}
}
\description{
These methods extract data from, and attribute new data to, objects of class 
\code{"lpca"} and \code{"lsvd"} (see
\code{?logisticPCA::`logisticPCA-package`}). (This package masks the
signature functions of \strong{\link[logisticPCA]{logisticPCA}} with wrappers
that add row and column names from the input matrix to the output matrices.)
}
\examples{

data(finches)

finches_mat <- as.matrix(dplyr::select(finches, -Island))
rownames(finches_mat) <- finches$Island

finches_lsvd <- as_bibble(logisticSVD(finches_mat))
# ILLUSTRATE ROLE-REVERSAL BETWEEN U AND V
ggbiplot(finches_lsvd, aes(x = LSC1, y = LSC2)) +
  geom_u_vector(aes(x = LSC1 * .01, y = LSC2 * .01)) +
  geom_u_text(aes(x = LSC1 * .01, y = LSC2 * .01, label = .name), size = 3) +
  geom_v_label(aes(label = .name), size = 3, alpha = .5)

finches_lpca <- as_bibble(logisticPCA(finches_mat))
# DISTINGUISH SEGMENT SIZE FROM TEXT SIZE (EXTRA `cex` PARAMETER?)
ggbiplot(finches_lpca, aes(x = LPC1, y = LPC2)) +
  geom_u_axis(aes(x = LPC1 * .02, y = LPC2 * .02, label = .name), size = 3) +
  geom_v_label(aes(label = .name), size = 3, alpha = .5)
}
