% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibble.r
\name{bibble}
\alias{bibble}
\alias{as_bibble}
\alias{make_bibble}
\alias{is_bibble}
\alias{is.bibble}
\title{A unified ordination object class}
\usage{
as_bibble(x)

make_bibble(u = NULL, v = NULL, ...)

is_bibble(x)

is.bibble(x)
}
\arguments{
\item{x}{An ordination object.}

\item{u, v}{Matrices to be used as factors of a bibble.}

\item{...}{Additional elements of a custom bibble.}
}
\description{
The \code{bbl} ("bibble") class wraps around a range of ordination classes, 
making available a suite of ordination tools that specialize to each original
object class, including \code{print}, \code{fortify}, \code\link{{ggbiplot}},
and several biplot stat and geom layers.
}
\details{
No default method is provided for \code{as_bibble}, despite most defined 
methods being equivalent (simply adding \code{"bbl"} to the vector of object 
classes). This prevents objects for which other methods are not defined from 
being re-classed as bibbles.

The function \code{make_bibble} creates a bibble structured as a list of two 
matrices, \code{u} and \code{v}, which must have the same number of columns 
and the same column names.

\code{is_bibble} checks an object \code{x} for the \code{"bbl"} class and for
consistency between \code{get_coord(x)} and the columns of \code{get_u(x)} 
and \code{get_v(x)}, using the functions at \code{\link{bibble-factors}}.
}
