% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ord-classes.r
\name{classes}
\alias{classes}
\alias{eigen_ord}
\alias{svd_ord}
\title{Classes for workhorse matrix decompositions}
\usage{
eigen_ord(x, EISPACK = FALSE)

svd_ord(x, LINPACK = FALSE)
}
\arguments{
\item{x}{a numeric or complex matrix whose spectral decomposition is to
    be computed.  Logical matrices are coerced to numeric.}

\item{EISPACK}{logical. Defunct and ignored.}

\item{LINPACK}{logical.  Defunct and ignored.}
}
\description{
...
}
