% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{countries}
\alias{countries}
\alias{country_differences}
\alias{country_attributes}
\title{Student ratings of similarities and attributes of 13 countries}
\format{
\link[base:matrix]{Matrices}.
}
\source{
Greenacre (2010), pp. 43--44,46.
}
\usage{
data(country_differences)

data(country_attributes)
}
\description{
Greenacre (2010) used these ratings to illustrate
multidimensional scaling. The uses shown there are reproduced in an
example.
}
\examples{
# Multidimensional scaling of country differences and regression of attributes
# Reproduce Exhibit 4.2 in Greenacre (2010)
data(country_differences)
country_differences \%>\%
  cmdscale_ord(k = 2) \%>\%
  as_tbl_ord() \%>\%
  print() -> differences_cmds
differences_plot <- differences_cmds \%>\%
  ggbiplot(aes(x = 1, y = 2, label = .name)) +
  geom_cols_text()
differences_plot
# Reproduce Exhibit 4.5 in Greenacre (2010)
data(country_attributes)
lm(country_attributes ~ get_rows(differences_cmds)) \%>\%
  as_tbl_ord() \%>\%
  print() -> attributes_fit
differences_plot +
  geom_cols_vector(data = attributes_fit) +
  geom_cols_text_radiate(data = attributes_fit, hjust = .3)
# Confer PCA weights and inertia between cases and variables
data(country_attributes)
country_attributes \%>\%
  prcomp() \%>\%
  as_tbl_ord() \%>\%
  print() -> attributes_pca
# form biplot
attributes_pca \%>\%
  ggbiplot(
    aes(label = .name),
    sec.axes = "cols", scale.factor = dim_cols(attributes_pca)
  ) +
  theme_bw() + theme(panel.grid = element_blank()) +
  geom_rows_text(aes(x = PC1, y = PC2), color = "seagreen") +
  geom_cols_vector(aes(x = PC1, y = PC2), color = "darkred") +
  geom_cols_text(
    aes(x = PC1, y = PC2),
    color = "darkred", hjust = "outward", vjust = "outward"
  )
# covariance biplot
attributes_pca \%>\%
  confer_inertia(0) \%>\%
  ggbiplot(
    aes(label = .name),
    sec.axes = "rows", scale.factor = dim_rows(attributes_pca)
  ) +
  theme_bw() + theme(panel.grid = element_blank()) +
  geom_rows_text(aes(x = PC1, y = PC2), color = "seagreen") +
  geom_cols_vector(aes(x = PC1, y = PC2), color = "darkred") +
  geom_cols_text_repel(aes(x = PC1, y = PC2), color = "darkred")
# Compare an MDS to a PCA
data(country_differences)
country_differences \%>\%
  cmdscale_ord(k = 2) \%>\%
  as_tbl_ord() \%>\%
  print() -> differences_cmds
country_differences \%>\%
  prcomp() \%>\%
  as_tbl_ord() \%>\%
  print() -> differences_pca
# negate PCA axis to match MDS axis
plot(gridExtra::arrangeGrob(grobs = list(
  ggbiplot(differences_cmds) + geom_rows_point(),
  ggbiplot(differences_pca) + scale_y_reverse() + geom_rows_point()
), ncol = 2))
}
\references{
Greenacre MJ (2010) \emph{Biplots in Practice}. Fundacion BBVA, ISBN:
978-84-923846.
\url{https://www.fbbva.es/microsite/multivariate-statistics/biplots.html}
}
\keyword{datasets}
