% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.r
\name{formatting}
\alias{formatting}
\alias{format.tbl_ord}
\alias{print.tbl_ord}
\title{Formatting and printing methods for \code{tbl_ord}s}
\usage{
\method{format}{tbl_ord}(x, ..., n = NULL, width = NULL,
  n_extra = NULL)

\method{print}{tbl_ord}(x, ..., n = NULL, width = NULL,
  n_extra = NULL)
}
\arguments{
\item{x}{An ordination object.}

\item{...}{Additional arguments.}

\item{n}{Number of rows to show. If \code{NULL}, the default, will print
all rows if less than option \code{tibble.print_max}. Otherwise, will
print \code{tibble.print_min} rows.}

\item{width}{Width of text output to generate. This defaults to NULL, which
means use \code{getOption("tibble.width")} or (if also NULL)
\code{getOption("width")}; the latter displays only the columns that
fit on one screen. You can also set \code{options(tibble.width = Inf)} to
override this default and always print all columns.}

\item{n_extra}{Number of extra columns to print abbreviated information for,
if the width is too small for the entire tibble. If \code{NULL}, the
default, will print information about at most \code{tibble.max_extra_cols}
extra columns.}
}
\description{
The \code{format} and \code{print} methods for class \code{"tbl_ord"} are
adapted from those for class \code{"tbl_df"} in
\strong{\link[tibble]{tibble}} and for class \code{"tbl_graph"} in
\strong{\link[tidygraph]{tidygraph}}.
}
\details{
Note: The \code{format} function is tedius but cannot be easily modularized
without invoking \code{get_*}, \code{annotation_*}, and \code{augmentation_*}
multiple times, thereby significantly reducing performance.
}
