% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify.r
\name{fortification}
\alias{fortification}
\alias{fortify.tbl_ord}
\alias{fortify_u}
\alias{fortify_v}
\alias{tidy.tbl_ord}
\title{Fortify a \code{tbl_ord} for plotting}
\usage{
\method{fortify}{tbl_ord}(model, data, ..., .matrix = "uv", include = "all")

fortify_u(model, include = "all")

fortify_v(model, include = "all")

\method{tidy}{tbl_ord}(x, ..., .matrix = "uv", include = "all")
}
\arguments{
\item{model, x}{A \code{\link{tbl_ord}}.}

\item{data}{Ignored.}

\item{...}{Additional arguments received from \code{fortify} or \code{tidy}; 
ignored.}

\item{.matrix}{A character string matching one of several indicators for one
or both matrices in a matrix decomposition used for ordination. The
standard values are \code{"u"}, \code{"v"}, and \code{"uv"}.}

\item{include}{Character matched to \code{"coordinates"}, \code{"shared"}, or
\code{"all"}; whether the fortified data frame should include only the 
ordination coordinates or also augmented case and variable data, and, if
the latter, whether only shared fields or all from both.}
}
\description{
These methods of \code{\link[ggplot2]{fortify}} and \code{\link[broom]{tidy}}
convert a \code{\link{tbl_ord}} to a \code{\link{tibble}}.
}
