% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify.r
\name{fortify}
\alias{fortify}
\alias{fortify.tbl_ord}
\alias{fortify_coord}
\alias{fortify_rows}
\alias{fortify_cols}
\title{Fortify a tbl_ord for plotting}
\usage{
\method{fortify}{tbl_ord}(
  model,
  data,
  ...,
  .matrix = "dims",
  .supplement = TRUE,
  include = "all"
)

fortify_coord(model)

fortify_rows(model, include = "all")

fortify_cols(model, include = "all")
}
\arguments{
\item{model}{A '\link{tbl_ord}' object.}

\item{data}{Ignored.}

\item{...}{Additional arguments received from \code{fortify()}; ignored.}

\item{.matrix}{A character string matching one of several indicators for one
or both matrices in a matrix decomposition used for ordination. The
standard values are \code{"rows"}, \code{"cols"}, and \code{"dims"} (for both).}

\item{.supplement}{Logical; whether to include
\link[=supplementation]{supplementary} points.}

\item{include}{Character matched to \code{"coordinates"}, \code{"shared"}, or \code{"all"};
whether the fortified data frame should include only the ordination
coordinates or also augmented case and variable data, and, if the latter,
whether only shared fields or all from both.}
}
\description{
This \code{\link[ggplot2:fortify]{ggplot2::fortify()}} method converts a 'tbl_ord' object to
a \link[tibble:tbl_df-class]{'tbl_df'} object.
}
\details{
The \code{fortify()} method for 'tbl_ord' objects produces the same
output, a \link[tibble:tibble]{tibble} obtained by binding columns obtained via
\code{\link[=accessors]{get_*()}}, \code{\link[=annotation]{annotation_*()}}, and
\code{\link[=augmentation]{augmentation_*}}, respectively.

In the special case \code{.matrix = "coordinates"}, the tibble consists instead of
the artificial coordinates and associated metadata, often including the
proportion of variance explained or captured by each. This can be used to
produce scree plots, for example.

If any augmented variables are included (i.e. unless \code{include = "coordinates"}), then the tibble is assigned a \code{"coordinates"} attribute
whose value is obtained via \code{\link[=get_coord]{get_coord()}}. (That this attribute will not be
printed with the tibble.) This facilitates some downstream functionality that
relies on more than those coordinates used as position aesthetics in a
biplot, in particular \code{\link[=stat_spantree]{stat_spantree()}}.
}
