% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-bagplot.r
\name{geom_bagplot}
\alias{geom_bagplot}
\title{Bagplots}
\usage{
geom_bagplot(
  mapping = NULL,
  data = NULL,
  stat = "bagplot",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this layer.
When using a \verb{geom_*()} function to construct a layer, the \code{stat}
argument can be used the override the default coupling between geoms and
stats. The \code{stat} argument accepts the following:
\itemize{
\item A \code{Stat} ggproto subclass, for example \code{StatCount}.
\item A string naming the stat. To give the stat as a string, strip the
function name of the \code{stat_} prefix. For example, to use \code{stat_count()},
give the stat as \code{"count"}.
\item For more information and other ways to specify the stat, see the
\link[ggplot2:layer_stats]{layer stat} documentation.
}}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{...}{Additional arguments passed to \code{\link[ggplot2:layer]{ggplot2::layer()}}.}

\item{na.rm}{Passed to \code{\link[ggplot2:layer]{ggplot2::layer()}}.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\value{
A ggproto \link[ggplot2:layer]{layer}.
}
\description{
Render bagplots from tagged data comprising medians, hulls,
contours, and outlier specifications.
}
\details{
\code{geom_bagplot()} is designed to pair with \code{\link[=stat_bagplot]{stat_bagplot()}}.

\strong{WARNING:}
The trade-off between precision and runtime is far greater for depth
estimation than for density estimation. At the resolution of the default
\eqn{100 \times 100} grid, basic examples may vary widely starting from
different random seeds.
}
\section{Biplot layers}{

\code{\link[=ggbiplot]{ggbiplot()}} uses \code{\link[ggplot2:fortify]{ggplot2::fortify()}} internally to produce a single data
frame with a \code{.matrix} column distinguishing the subjects (\code{"rows"}) and
variables (\code{"cols"}). The stat layers \code{stat_rows()} and \code{stat_cols()} simply
filter the data frame to one of these two.

The geom layers \verb{geom_rows_*()} and \verb{geom_cols_*()} call the corresponding
stat in order to render plot elements for the corresponding factor matrix.
\verb{geom_dims_*()} selects a default matrix based on common practice, e.g.
points for rows and arrows for columns.
}

\section{Aesthetics}{

\code{geom_bagplot()} understands the following aesthetics (required aesthetics
are in bold):
\itemize{
\item \strong{\code{x}}, \strong{\code{y}}
\item \strong{\code{component}}
\item \code{linewidth}, \code{linetype},
\code{colour}, \code{fill}, \code{alpha}
\item \code{shape}, \code{stroke}, \code{size},
\code{median_colour}, \code{median_fill}, \code{median_alpha}
\item \code{fence_linewidth}, \code{fence_linetype},
\code{fence_colour}, \code{fence_fill}, \code{fence_alpha}
\item \code{outlier_shape}, \code{outlier_stroke}, \code{outlier_size},
\code{outlier_colour}, \code{outlier_fill}, \code{outlier_alpha}
\item \code{group}
}
}

\examples{
# Motor Trends base plot with factorized cylinder counts
p <- mtcars |> 
  transform(cyl = factor(cyl)) |> 
  ggplot(aes(x = wt, y = disp)) +
  theme_bw()
# basic bagplot
# FIXME: Fence defaults to dotted rather than invisible `linetype`.
p + geom_bagplot()
# group by cylinder count
# FIXME: Component aesthetic data values aren't mapped to color values.
set.seed(526822L)
p + geom_bagplot(
  fraction = 0.4, coef = 1.2,
  aes(
    fill = cyl,
    linetype = cyl,# fence_linetype = cyl,
    median_colour = cyl, outlier_colour = cyl
  )
)
# fixed custom aesthetics (bag and fence shrunk to expose outliers)
set.seed(009003L)
p + geom_bagplot(
  fraction = 0.4, coef = 1.2,
  aes(fill = cyl), shape = "plus", outlier_shape = "asterisk"
)
}
\seealso{
Other geom layers: 
\code{\link{geom_axis}()},
\code{\link{geom_interpolation}()},
\code{\link{geom_isoline}()},
\code{\link{geom_lineranges}()},
\code{\link{geom_origin}()},
\code{\link{geom_rule}()},
\code{\link{geom_text_radiate}()},
\code{\link{geom_vector}()}
}
\concept{geom layers}
