% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-pin.r
\name{geom_pin}
\alias{geom_pin}
\title{Pins emanating from the abscissa}
\usage{
geom_pin(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  length = unit(0.2, "npc"),
  lineend = "round",
  pin_labels = TRUE,
  orientation = NA,
  ...,
  parse = FALSE,
  check_overlap = FALSE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this layer.
When using a \verb{geom_*()} function to construct a layer, the \code{stat}
argument can be used to override the default coupling between geoms and
stats. The \code{stat} argument accepts the following:
\itemize{
\item A \code{Stat} ggproto subclass, for example \code{StatCount}.
\item A string naming the stat. To give the stat as a string, strip the
function name of the \code{stat_} prefix. For example, to use \code{stat_count()},
give the stat as \code{"count"}.
\item For more information and other ways to specify the stat, see the
\link[ggplot2:layer_stats]{layer stat} documentation.
}}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{length}{A \code{\link[grid:unit]{grid::unit()}} object that sets the length of the pins. Use
scale expansion to avoid overplotting of data.}

\item{lineend}{Line end style (round, butt, square).}

\item{pin_labels}{Logical; whether to append labels to the pins.}

\item{orientation}{The orientation of the layer. The default (\code{NA})
automatically determines the orientation from the aesthetic mapping. In the
rare event that this fails it can be given explicitly by setting \code{orientation}
to either \code{"x"} or \code{"y"}. See the \emph{Orientation} section for more detail.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:layer]{ggplot2::layer()}}.}

\item{parse}{If \code{TRUE}, the labels will be parsed into expressions and
displayed as described in \code{?plotmath}.}

\item{check_overlap}{If \code{TRUE}, text that overlaps previous text in the
same layer will not be plotted. \code{check_overlap} happens at draw time and in
the order of the data. Therefore data should be arranged by the label
column before calling \code{geom_text()}. Note that this argument is not
supported by \code{geom_label()}.}

\item{na.rm}{Passed to \code{\link[ggplot2:layer]{ggplot2::layer()}}.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display. To include legend keys for all levels, even
when no data exists, use \code{TRUE}.  If \code{NA}, all levels are shown in legend,
but unobserved levels are omitted.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:annotation_borders]{annotation_borders()}}.}
}
\value{
A ggproto \link[ggplot2:layer]{layer}.
}
\description{
\code{geom_pin()} renders segments orthogonal to the abscissa,
optionally with outward-facing labels.
}
\section{Biplot layers}{

\code{\link[=ggbiplot]{ggbiplot()}} uses \code{\link[ggplot2:fortify]{ggplot2::fortify()}} internally to produce a single data
frame with a \code{.matrix} column distinguishing the subjects (\code{"rows"}) and
variables (\code{"cols"}). The stat layers \code{stat_rows()} and \code{stat_cols()} simply
filter the data frame to one of these two.

The geom layers \verb{geom_rows_*()} and \verb{geom_cols_*()} call the corresponding
stat in order to render plot elements for the corresponding factor matrix.
\verb{geom_dims_*()} selects a default matrix based on common practice, e.g.
points for rows and arrows for columns.
}

\section{Aesthetics}{

\code{geom_pin()} understands the following aesthetics (required aesthetics
are in bold):
\itemize{
\item \strong{\code{x}}
\item \strong{\code{y}}
\item \code{linetype}
\item \code{linewidth}
\item \code{colour}
\item \code{alpha}
\item \code{label}
\item \code{size}, \code{angle},
\code{family}, \code{fontface},
\code{label_colour}, \code{label_alpha}
\item \code{group}
}
}

\examples{
iris_pca <- prcomp(iris[, seq(4)], rank. = 1)
iris_rows <- 
  iris_pca$x |> as.data.frame() |> 
  tibble::rownames_to_column(var = "id") |> 
  transform(species = iris$Species)
iris_cols <- 
  iris_pca$rotation |> as.data.frame() |> 
  tibble::rownames_to_column(var = "measurement")
ggplot(iris_rows, aes(x = PC1)) +
  geom_density(aes(color = species)) +
  geom_pin(
    aes(x = PC1 * 10, label = measurement), ymax = 1,
    data = iris_cols,
    check_overlap = TRUE
  ) +
  scale_x_continuous(sec.axis = ~ . / 10)
# TODO: Expand plot to include pins. (Transform data in `$setup_data()`?)
ggplot(iris_rows, aes(x = PC1)) +
  geom_density(aes(color = species)) +
  geom_pin(
    aes(x = PC1 * 10, label = measurement), ymax = -1,
    data = iris_cols,
    check_overlap = TRUE
  ) +
  scale_x_continuous(sec.axis = ~ . / 10)
}
\seealso{
Other geom layers: 
\code{\link{geom_interpolation}()},
\code{\link{geom_origin}()}
}
\concept{geom layers}
