% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend-draw.r
\name{ggbiplot-draw-key}
\alias{ggbiplot-draw-key}
\alias{draw_key_line}
\alias{draw_key_crosslines}
\alias{draw_key_crosspoint}
\title{Biplot key drawing functions}
\usage{
draw_key_line(data, params, size)

draw_key_crosslines(data, params, size)

draw_key_crosspoint(data, params, size)
}
\arguments{
\item{data}{A single row data frame containing the scaled aesthetics to
display in this key}

\item{params}{A list of additional parameters supplied to the geom.}

\item{size}{Width and height of key in mm.}
}
\value{
A grid grob.
}
\description{
Biplot key drawing functions
}
\seealso{
\code{\link[ggplot2]{draw_key}}
}
