% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-layers.r
\docType{data}
\name{StatU}
\alias{StatU}
\alias{StatV}
\alias{stat_u}
\alias{stat_v}
\title{Render plot elements for one matrix of an ordination}
\format{An object of class \code{StatU} (inherits from \code{StatIdentity}, \code{Stat}, \code{ggproto}, \code{gg}) of length 2.}
\usage{
stat_u(mapping = NULL, data = data, geom = "point",
  position = "identity", ..., show.legend = NA, inherit.aes = TRUE)

stat_v(mapping = NULL, data = data, geom = "axis",
  position = "identity", ..., show.legend = NA, inherit.aes = TRUE)
}
\description{
Render plot elements for one matrix of an ordination
}
\section{Biplot layers}{

\code{\link{ggbiplot}} uses \code{\link[ggplot2]{fortify}} internally to 
produce a single data frame with a \code{.matrix} column distinguishing the 
subjects (\code{"u"}) and variables (\code{"v"}). The stat layers 
\code{stat_u} and \code{stat_v} simply filter the data frame to one of these 
two.

The geom layers \code{geom_u_*} and \code{geom_v_*} call the corresponding
stat in order to render plot elements for the corresponding matrix \eqn{U} or
\eqn{V}. \code{geom_biplot_*} selects a default matrix based on common
practice, e.g. \eqn{U} for points and \eqn{V} for arrows.
}

\keyword{datasets}
