% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-polygon.r
\name{ggbiplot-polygon}
\alias{ggbiplot-polygon}
\alias{geom_u_polygon}
\alias{geom_v_polygon}
\alias{geom_biplot_polygon}
\title{Render polygons around ordinates}
\usage{
geom_u_polygon(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)

geom_v_polygon(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)

geom_biplot_polygon(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", .matrix = "u", ..., na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame.}, and
will be used as the layer data.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Additional arguments passed to \code{\link[ggplot2]{layer}}.}

\item{na.rm}{Passed to \code{\link[ggplot2]{layer}}s.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}

\item{.matrix}{A character string matching one of several indicators for one
or both matrices in a matrix decomposition used for ordination. The
standard values are \code{"u"}, \code{"v"}, and \code{"uv"}.}
}
\description{
\code{geom_*_polygon} renders polygons around the convex hulls of the
positions of the subjects or vectors.
}
\section{Biplot layers}{

\code{\link{ggbiplot}} uses \code{\link[ggplot2]{fortify}} internally to 
produce a single data frame with a \code{.matrix} column distinguishing the 
subjects (\code{"u"}) and variables (\code{"v"}). The stat layers 
\code{stat_u} and \code{stat_v} simply filter the data frame to one of these 
two.

The geom layers \code{geom_u_*} and \code{geom_v_*} call the corresponding
stat in order to render plot elements for the corresponding matrix \eqn{U} or
\eqn{V}. \code{geom_biplot_*} selects a default matrix based on common
practice, e.g. \eqn{U} for points and \eqn{V} for arrows.
}

\section{Aesthetics}{
 \code{geom_*_polygon} understands the following
  aesthetics (required aesthetics are in bold):
\itemize{
  \item \strong{\code{x}}
  \item \strong{\code{y}}
  \item \code{alpha}
  \item \code{colour}
  \item \code{fill}
  \item \code{linetype}
  \item \code{size}
  \item \code{group}
}
}

\examples{
# reproduce Exhibit 7.1 in Greenacre (2010)

# manually construct singular value decomposition
x <- as.matrix(USArrests[, -3])
r <- (1 / sum(x)) * x \%*\% matrix(1, ncol(x))
c <- (1 / sum(x)) * t(x) \%*\% matrix(1, nrow(x))
l <- log(x)
y <- (diag(nrow(x)) - matrix(1, nrow(x)) \%*\% t(r)) \%*\%
  l \%*\%
  t(diag(ncol(x)) - matrix(1, ncol(x)) \%*\% t(c))
d_r <- diag(1 / nrow(x), nrow(x))
d_c <- diag(1 / ncol(x), ncol(x))
s <- sqrt(d_r) \%*\% y \%*\% sqrt(d_c)
dimnames(s) <- dimnames(x)
d <- svd(s)

# manually compute principal and standard coordinates
f <- diag(1 / sqrt(diag(d_r))) \%*\% d$u \%*\% diag(d$d)
g <- diag(1 / sqrt(diag(d_c))) \%*\% d$v \%*\% diag(d$d)
phi <- diag(1 / sqrt(diag(d_r))) \%*\% d$u
gamma <- diag(1 / sqrt(diag(d_c))) \%*\% d$v
sc <- 5
plot(rbind(f, gamma / sc), pch = NA)
points(f, pch = 16, col = "darkgreen")
points(gamma / sc, pch = 17, col = "brown4")
lines(rbind(gamma, gamma[1, ]) / sc, lty = 3, lwd = 2, col = "brown4")

# weighted SVD as 'tbl_ord'
(b <- augment(as_tbl_ord(d)))
get_conference(b)

b <- confer_inertia(b, 0)
ggbiplot(b, aes(label = .name)) +
  theme_bw() +
  geom_u_text(color = "darkgreen", size = 3) +
  geom_v_polygon(fill = NA, linetype = "dashed", color = "brown4") +
  geom_v_text(color = "brown4", fontface = "bold")

# using a secondary axis
b <- confer_inertia(b, 1)
ggbiplot(b, aes(label = .name)) +
  geom_u_text(color = "darkgreen", size = 3) +
  geom_v_polygon(fill = NA, linetype = "dashed", color = "brown4") +
  geom_v_text(color = "brown4", fontface = "bold")
ggbiplot(b, aes(label = .name), sec.axes = "v") +
  geom_u_text(color = "darkgreen", size = 3) +
  geom_v_polygon(fill = NA, linetype = "dashed", color = "brown4") +
  geom_v_text(color = "brown4", fontface = "bold")
ggbiplot(b, aes(label = .name), sec.axes = "u") +
  geom_u_text(color = "darkgreen", size = 3) +
  geom_v_polygon(fill = NA, linetype = "dashed", color = "brown4") +
  geom_v_text(color = "brown4", fontface = "bold")
ggbiplot(b, aes(label = .name), sec.axes = "u", scale.factor = 10) +
  geom_u_text(color = "darkgreen", size = 3) +
  geom_v_polygon(fill = NA, linetype = "dashed", color = "brown4") +
  geom_v_text(color = "brown4", fontface = "bold")
# NOTE: SECONDARY AXES ARE ADDED INTERNALLY USING `scale_*_continuous()`;
# THEY WILL BE REPLACED BY DOWNSTREAM `x` AND `y` SCALES.
}
