% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-text-radiate.r
\docType{data}
\name{ggbiplot-text-radiate}
\alias{ggbiplot-text-radiate}
\alias{GeomTextRadiate}
\alias{geom_u_text_radiate}
\alias{geom_v_text_radiate}
\alias{geom_biplot_text_radiate}
\title{Render text at ordinates radiating out from the origin}
\format{An object of class \code{GeomTextRadiate} (inherits from \code{GeomText}, \code{Geom}, \code{ggproto}, \code{gg}) of length 2.}
\usage{
geom_u_text_radiate(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., parse = FALSE, check_overlap = FALSE,
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE)

geom_v_text_radiate(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., parse = FALSE, check_overlap = FALSE,
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE)

geom_biplot_text_radiate(mapping = NULL, data = NULL,
  stat = "identity", position = "identity", .matrix = "v", ...,
  parse = FALSE, check_overlap = FALSE, na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:layer]{ggplot2::layer()}}.}

\item{parse}{See \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}}.}

\item{check_overlap}{See \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}}.}

\item{na.rm}{Passed to \code{\link[ggplot2:layer]{ggplot2::layer()}}.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}

\item{.matrix}{A character string matching one of several indicators for one
or both matrices in a matrix decomposition used for ordination. The
standard values are \code{"u"}, \code{"v"}, and \code{"uv"}.}
}
\description{
\code{geom_*_text_radiate()} is adapted from \code{\link[ggbiplot:ggbiplot]{ggbiplot::ggbiplot()}}.
It renders labels at specified positions and angles that radiate out from
the origin.
}
\section{Biplot layers}{

\code{\link[=ggbiplot]{ggbiplot()}} uses \code{\link[ggplot2:fortify]{ggplot2::fortify()}} internally to produce a single data
frame with a \code{.matrix} column distinguishing the subjects (\code{"u"}) and
variables (\code{"v"}). The stat layers \code{stat_u()} and \code{stat_v()} simply filter
the data frame to one of these two.

The geom layers \code{geom_u_*()} and \code{geom_v_*()} call the corresponding stat in
order to render plot elements for the corresponding matrix \eqn{U} or
\eqn{V}. \code{geom_biplot_*()} selects a default matrix based on common practice,
e.g. \eqn{U} for points and \eqn{V} for arrows.
}

\section{Aesthetics}{

\code{geom_*_text_radiate()} understands the following aesthetics (required
aesthetics are in bold):
\itemize{
\item \strong{\code{x}}
\item \strong{\code{y}}
\item \strong{\code{label}}
\item \code{alpha}
\item \code{angle}
\item \code{colour}
\item \code{family}
\item \code{fontface}
\item \code{hjust}
\item \code{lineheight}
\item \code{size}
\item \code{vjust}
\item \code{group}
}
}

\examples{
# Generalized multiple linear regression on marine ecosystem data
# Reproduce Exhibit 2.5 in Greenacre (2010)
bioenv_std <- dplyr::mutate_if(bioenv, is.numeric, ~ as.vector(scale(.)))
resp_std <- as.matrix(dplyr::select(bioenv_std, a:e))
pred_std <- as.matrix(dplyr::select(bioenv_std, x = Depth, y = Pollution))
lm(resp_std ~ pred_std + 0) \%>\%
  as_tbl_ord() \%>\%
  print() -> bioenv_std_mlm
ggbiplot(bioenv_std_mlm, aes(label = .name)) +
  theme_bw() +
  geom_u_text(color = "darkgreen") +
  geom_v_vector(color = "brown4") +
  geom_v_text_radiate(color = "brown4")
# Reproduce Exhibit 3.2 in Greenacre (2010)
bioenv \%>\%
  dplyr::mutate_at(dplyr::vars(a:e), ~ . ^ (1/4)) \%>\%
  dplyr::mutate_at(dplyr::vars(Pollution:Depth), ~ as.vector(scale(.))) \%>\%
  print() -> bioenv_4rt
resp_4rt <- as.matrix(dplyr::select(bioenv_4rt, a:e))
pred_4rt <- as.matrix(dplyr::select(bioenv_4rt, x = Depth, y = Pollution))
lm(resp_4rt ~ pred_4rt + 0) \%>\%
  as_tbl_ord() \%>\%
  print() -> bioenv_4rt_mlm
ggbiplot(bioenv_4rt_mlm, aes(x = x, y = y, label = .name)) +
  theme_bw() +
  geom_u_text(color = "darkgreen") +
  geom_v_vector(color = "brown4") +
  geom_v_text_radiate(color = "brown4")
}
\keyword{datasets}
