% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-vector.r
\docType{data}
\name{GeomVector}
\alias{GeomVector}
\alias{geom_u_vector}
\alias{geom_v_vector}
\alias{geom_biplot_vector}
\alias{GeomAxis}
\alias{geom_u_axis}
\alias{geom_v_axis}
\alias{geom_biplot_axis}
\title{Render axes vectors, with or without angled labels}
\format{An object of class \code{GeomVector} (inherits from \code{GeomSegment}, \code{Geom}, \code{ggproto}, \code{gg}) of length 4.}
\usage{
geom_u_vector(mapping = NULL, data = NULL, position = "identity",
  arrow = default_arrow, ..., na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)

geom_v_vector(mapping = NULL, data = NULL, position = "identity",
  arrow = default_arrow, ..., na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)

geom_biplot_vector(mapping = NULL, data = NULL, position = "identity",
  .matrix = "v", arrow = default_arrow, ..., na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE)

geom_u_axis(mapping = NULL, data = NULL, position = "identity",
  arrow = default_arrow, ..., na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)

geom_v_axis(mapping = NULL, data = NULL, position = "identity",
  arrow = default_arrow, ..., na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)

geom_biplot_axis(mapping = NULL, data = NULL, position = "identity",
  .matrix = "v", arrow = default_arrow, ..., na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame.}, and
will be used as the layer data.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}
}
\description{
\code{geom_*_vector} renders arrows from the origin to the position of each
subject or vector. \code{geom_*_axis} is adapted from
\strong{link[ggbiplot]{ggbiplot}}. In addition to the arrows, it renders
labels that expand outward from the arrowheads at the same angles.
}
\section{Biplot layers}{

\code{\link{ggbiplot}} uses \code{\link[ggplot2]{fortify}} internally to 
produce a single data frame with a \code{.matrix} column distinguishing the 
subjects (\code{"u"}) and variables (\code{"v"}). The stat layers 
\code{stat_u} and \code{stat_v} simply filter the data frame to one of these 
two.

The geom layers \code{geom_u_*} and \code{geom_v_*} call the corresponding
stat in order to render plot elements for the corresponding matrix \eqn{U} or
\eqn{V}. \code{geom_biplot_*} selects a default matrix based on common
practice, e.g. \eqn{U} for points and \eqn{V} for arrows.
}

\section{Aesthetics}{

\code{geom_*_vector} and \code{geom_*_axis} understand the following
aesthetics (required aesthetics are in bold):
\itemize{
  \item \strong{\code{x}}
  \item \strong{\code{y}}
  \item \strong{\code{label}} (\code{geom_*_axis} only)
  \item \code{alpha}
  \item \code{colour}
  \item \code{fill}
  \item \code{linetype}
  \item \code{size}
  \item \code{group}
}
}

\keyword{datasets}
