% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{glass}
\alias{glass}
\title{Glass composition data from archaeological sites in Israel}
\format{
A \link[tibble:tibble]{tibble}.
}
\source{
Baxter & Freestone (2006).
}
\usage{
data(glass)
}
\description{
Sites, artifacts, and amounts of several components of the sand.
}
\keyword{datasets}
