% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ord-gof.r
\name{goodness-of-fit}
\alias{goodness-of-fit}
\alias{gof}
\alias{ord_quality}
\alias{ord_adequacy}
\alias{ord_predictivity}
\title{Measures of goodness of fit of ordination models}
\usage{
ord_quality(x)

ord_adequacy(x, .matrix)

ord_predictivity(x, .matrix)
}
\arguments{
\item{x}{A 'tbl_ord' object.}

\item{.matrix}{A character string partially matched (lowercase) to several
indicators for one or both matrices in a matrix decomposition used for
ordination. The standard values are \code{"rows"}, \code{"cols"}, and \code{"dims"} (for
both).}
}
\value{
A vector, matrix, or list of matrices of numeric goodness-of-fit
statistics.
}
\description{
Compute the quality, adequacy, and predictivity of a 'tbl_ord'
object from the retrieved matrix factors.
}
\details{
Gower, Gardner--Lubbe, & le Roux (2011) detail several measures of fit for
biplots, most prominently
\itemize{
\item the \emph{quality} of the \eqn{r}-dimensional biplot,
measured as the proportion of variance in the plot,
calculated as the quotient of the traces of
\eqn{\Lambda_r = {D_r}^2} and of \eqn{\Lambda = D^2}.
\item the \emph{adequacy} of the representation of the \eqn{j}-th row
(respectively, column) in the \eqn{r}-dimensional biplot,
calculated as the \eqn{j}-th diagonal element of
\eqn{U_r\ {U_r}^\top} (respectively, \eqn{V_r\ {V_r}^\top}),
understood as the fidelity of the projections
of the standard coordinates.
\item the \emph{predictivity} of the \eqn{j}-th row
(respectively, column) in the \eqn{r}-dimensional biplot,
measured as the quotient of the \eqn{j}-th diagonal elements of
\eqn{U_r\ \Lambda_r\ {U_r}^\top} and of \eqn{U\ \Lambda\ U^\top}
(respectively, of \eqn{V_r\ \Lambda_r\ {V_r}^\top}
and of \eqn{V\ \Lambda\ V^\top}),
understood as the fidelity of the projections
of the principal coordinates.
}
These can be calculated directly from any SVD or EVD and interpreted for any
technique based on them. In some cases they may also be calculated for
supplementary elements.
}
\examples{
# log-ratio analysis of Apollonia glass composition data
glass_apollonia <- subset(
  glass,
  Site == "Apollonia",
  select = c("SiO2", "Na2O", "CaO", "Al2O3", "MgO", "K2O")
)
glass_lra <- lra(glass_apollonia, weighted = FALSE)

# quality (cumulative proportion of inertia included)
ord_quality(glass_lra)
# adequacy (fidelity of projections to standard coordinates)
ord_adequacy(glass_lra, "rows")
ord_adequacy(glass_lra, "cols")
# adequacy (fidelity of projections to principal coordinates)
ord_predictivity(glass_lra, "dims")

# principal components analysis of setosa iris data
iris_pca <- princomp(iris3[, , "Setosa"])

# quality (cumulative proportion of inertia included)
ord_quality(iris_pca)
# adequacy (fidelity of projections to standard coordinates)
# FIXME: Adequacy returns an unnamed matrix on empty elements.
ord_adequacy(iris_pca, "f")
ord_adequacy(iris_pca, "g")
# adequacy (fidelity of projections to principal coordinates)
# FIXME: Predictability throws error on empty elements.
ord_predictivity(iris_pca, "both")
}
\references{
Gower JC, Gardner--Lubbe S, & le Roux NJ (2011) \emph{Understanding Biplots}.
Wiley, ISBN: 978-0-470-01255-0. \url{https://www.wiley.com/go/biplots}
}
