% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{issp_women}
\alias{issp_women}
\title{Original respondent-level data from Spanish ISSP sample}
\format{
A \link[base:matrix]{matrix}.
}
\source{
Greenacre (2010).
}
\description{
Responses to several questions about working women and family.
}
\examples{
# Confer inertia between rows and columns of CA on Spanish ISSP sample
data(issp_women)
issp_women \%>\%
  ca::ca() \%>\%
  as_tbl_ord() \%>\%
  augment() \%>\%
  print() -> issp_ca
# Reproduce Exhibit 9.2 in Greenacre (2010)...
issp_ca \%>\%
  confer_inertia(c(1, 1)) \%>\%
  ggbiplot(aes(label = .name)) +
  theme_bw() +
  geom_u_point(color = "darkgreen") +
  geom_u_text_repel(color = "darkgreen") +
  geom_v_point(color = "brown", shape = 17) +
  geom_v_text_repel(color = "brown")
# Reproduce Exhibit 9.3 in Greenacre (2010)...
issp_ca \%>\%
  confer_inertia("rowprincipal") \%>\%
  ggbiplot(aes(label = .name)) +
  theme_bw() +
  geom_u_point(color = "darkgreen") +
  geom_v_point(color = "brown", shape = 17) +
  geom_v_text_repel(color = "brown")
# Reproduce Exhibits 9.4 and 9.6 in Greenacre (2010)...
}
\keyword{datasets}
