% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.r
\name{lda-ord}
\alias{lda-ord}
\alias{lda_ord}
\title{Augmented implementation of linear discriminant analysis}
\usage{
lda_ord(x, ...)
}
\arguments{
\item{x}{
    (required if no formula is given as the principal argument.)
    a matrix or data frame or Matrix containing the explanatory variables.
  }

\item{...}{
    arguments passed to or from other methods.
}

\item{axes.scale}{Character string indicating how to left-transform the
\code{scaling} value when rendering biplots using \code{\link[=ggbiplot]{ggbiplot()}}.}
}
\description{
This function replicates \code{\link[MASS:lda]{MASS::lda()}} with options to retain
elements useful to the \link{tbl_ord} class and biplot calculations.
}
\examples{
# Linear discriminant analysis of Reaven & Miller diabetes data
head(heplots::Diabetes)
(diabetes_lda <- MASS::lda(group ~ ., heplots::Diabetes))
dim(supplementation_u(diabetes_lda))
as_tbl_ord(diabetes_lda) \%>\%
  augment() \%>\%
  mutate_u(discriminant = ifelse(! .supplement, "centroid", "case")) \%>\%
  print() -> diabetes_lda
ggbiplot(diabetes_lda) +
  theme_bw() +
  geom_u_point(aes(color = .grouping, size = discriminant), alpha = .5) +
  geom_v_axis(color = "#888888") +
  geom_v_text_radiate(aes(label = .name)) +
  scale_color_brewer(type = "qual", palette = 2) +
  ggtitle(
    "Column-standardized LDA biplot of Reaven & Miller diabetes data"
  )
}
