% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ade4.r
\name{methods-ade4}
\alias{methods-ade4}
\alias{as_tbl_ord.nipals}
\alias{recover_rows.nipals}
\alias{recover_cols.nipals}
\alias{recover_inertia.nipals}
\alias{recover_coord.nipals}
\alias{recover_conference.nipals}
\alias{augmentation_rows.nipals}
\alias{augmentation_cols.nipals}
\alias{augmentation_coord.nipals}
\title{Functionality for non-linear iterative PLS ('nipals') objects}
\usage{
\method{as_tbl_ord}{nipals}(x)

\method{recover_rows}{nipals}(x)

\method{recover_cols}{nipals}(x)

\method{recover_inertia}{nipals}(x)

\method{recover_coord}{nipals}(x)

\method{recover_conference}{nipals}(x)

\method{augmentation_rows}{nipals}(x)

\method{augmentation_cols}{nipals}(x)

\method{augmentation_coord}{nipals}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"nipals"} as returned by \code{\link[ade4:nipals]{ade4::nipals()}}.
}
