% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ca-ca.r
\name{methods-ca}
\alias{methods-ca}
\alias{as_tbl_ord.ca}
\alias{recover_rows.ca}
\alias{recover_cols.ca}
\alias{recover_inertia.ca}
\alias{recover_conference.ca}
\alias{recover_coord.ca}
\alias{augmentation_rows.ca}
\alias{augmentation_cols.ca}
\alias{augmentation_coord.ca}
\title{Functionality for correspondence analysis ('ca') objects}
\usage{
\method{as_tbl_ord}{ca}(x)

\method{recover_rows}{ca}(x)

\method{recover_cols}{ca}(x)

\method{recover_inertia}{ca}(x)

\method{recover_conference}{ca}(x)

\method{recover_coord}{ca}(x)

\method{augmentation_rows}{ca}(x)

\method{augmentation_cols}{ca}(x)

\method{augmentation_coord}{ca}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"ca"} from the \strong{\link[ca:ca]{ca}} package.
}
\examples{
# table of hair and eye color data collapsed by sex
class(HairEyeColor)
haireye <- as.data.frame(rowSums(HairEyeColor, dims = 2L))
print(haireye)
# use correspondence analysis to construct row and column profiles
haireye \%>\%
  ca::ca() \%>\%
  as_tbl_ord() \%>\%
  print() -> haireye_ca
# summarize ordination
glance(haireye_ca)
# recover row and column profiles
get_rows(haireye_ca)
get_cols(haireye_ca)
# augment profiles with names, masses, distances, and inertias
(haireye_ca <- augment_ord(haireye_ca))
# summarize artificial coordinates
tidy(haireye_ca)
# scree plot of inertia
tidy(haireye_ca) \%>\%
  ggplot(aes(x = .name, y = .inertia)) +
  theme_bw() +
  geom_col() +
  labs(x = "", y = "Inertia")
# fortification adds all above columns
fortify(haireye_ca)
# symmetric biplot
haireye_ca \%>\%
  confer_inertia("symmetric") \%>\%
  ggbiplot(aes(label = .name, color = .matrix, shape = .matrix)) +
  scale_color_manual(values = c("navyblue", "brown")) +
  scale_shape_manual(values = c(17L, 16L)) +
  theme_bw() + theme_biplot() +
  geom_origin() +
  geom_rows_point(aes(size = .inertia)) +
  geom_rows_text_repel(show.legend = FALSE) +
  geom_cols_point(aes(size = .inertia)) +
  geom_cols_text_repel(show.legend = FALSE) +
  scale_size_area() +
  labs(color = "Dimension", shape = "Dimension", size = "Inertia")
# symmetric map biplot (both dimensions in principal coordinates)
haireye_ca \%>\%
  confer_inertia(c(1, 1)) \%>\%
  ggbiplot(aes(label = .name, color = .matrix, shape = .matrix)) +
  scale_color_manual(values = c("navyblue", "brown")) +
  scale_shape_manual(values = c(17L, 16L)) +
  theme_bw() + theme_biplot() +
  geom_origin() +
  geom_rows_point(aes(size = .inertia)) +
  geom_rows_text_repel(show.legend = FALSE) +
  geom_cols_point(aes(size = .inertia)) +
  geom_cols_text_repel(show.legend = FALSE) +
  scale_size_area() +
  labs(color = "Dimension", shape = "Dimension", size = "Inertia")
}
