% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-stats-factanal.r
\name{methods-factanal}
\alias{methods-factanal}
\alias{as_tbl_ord.factanal}
\alias{recover_rows.factanal}
\alias{recover_cols.factanal}
\alias{recover_inertia.factanal}
\alias{recover_coord.factanal}
\alias{recover_conference.factanal}
\alias{supplementation_rows.factanal}
\alias{augmentation_rows.factanal}
\alias{augmentation_cols.factanal}
\alias{augmentation_coord.factanal}
\title{Functionality for factor analysis ('factanal') objects}
\usage{
\method{as_tbl_ord}{factanal}(x)

\method{recover_rows}{factanal}(x)

\method{recover_cols}{factanal}(x)

\method{recover_inertia}{factanal}(x)

\method{recover_coord}{factanal}(x)

\method{recover_conference}{factanal}(x)

\method{supplementation_rows}{factanal}(x)

\method{augmentation_rows}{factanal}(x)

\method{augmentation_cols}{factanal}(x)

\method{augmentation_coord}{factanal}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"factanal"} as returned by \code{\link[stats:factanal]{stats::factanal()}}.
}
\details{
Most \strong{ordr} methods will only work if the 'factanal' object includes a
\verb{$scores} matrix, hence if a data matrix is provided.

\strong{NB:} \code{factanal()} normalizes the columns of \code{x} to unit length before it
calculates their loadings, so, for consistency with the behavior of other
methods (e.g. those of \code{\link[=lda_ord]{lda_ord()}}), inertia is assumed to lie with the rows.
However, the scores (retrieved by \code{recover_rows()}) are not factors of a
matrix decomposition and are better understood as predictions like those of
\code{lda_ord()}, so redistributing inertia to the loadings should be done
cautiously.
}
\examples{
# data frame of Swiss fertility and socioeconomic indicators
class(swiss)
head(swiss)
# perform factor analysis
swiss_fa <- factanal(~ ., factors = 2L, data = swiss)
}
