% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lra.r
\name{methods-lra}
\alias{methods-lra}
\alias{as_tbl_ord.lra}
\alias{recover_rows.lra}
\alias{recover_cols.lra}
\alias{recover_inertia.lra}
\alias{recover_coord.lra}
\alias{recover_conference.lra}
\alias{augmentation_rows.lra}
\alias{augmentation_cols.lra}
\alias{augmentation_coord.lra}
\title{Functionality for log-ratio analysis ('lra') objects}
\usage{
\method{as_tbl_ord}{lra}(x)

\method{recover_rows}{lra}(x)

\method{recover_cols}{lra}(x)

\method{recover_inertia}{lra}(x)

\method{recover_coord}{lra}(x)

\method{recover_conference}{lra}(x)

\method{augmentation_rows}{lra}(x)

\method{augmentation_cols}{lra}(x)

\method{augmentation_coord}{lra}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"lra"}, a class introduced in this package to organize
the singular value decomposition of a double-centered log-transformed data
matrix output by \code{\link[=lra]{lra()}}.
}
\examples{
# data frame of violent crime arrests in the United States
class(USArrests)
head(USArrests)
# get state abbreviation data
state <- data.frame(
  .name = state.name,
  .abb = state.abb
)
# compute (non-compositional, unweighted) log-ratio analysis
USArrests \%>\%
  subset(select = -UrbanPop) \%>\%
  lra() \%>\%
  as_tbl_ord() \%>\%
  print() -> arrests_lra
# summarize ordination
glance(arrests_lra)
# augment log-ratio profiles with names and join state abbreviations
arrests_lra \%>\%
  augment_ord() \%>\%
  left_join_rows(state, by = ".name") \%>\%
  print() -> arrests_lra
# recover state and arrest profiles
head(get_rows(arrests_lra))
get_cols(arrests_lra)
# summarize principal components
tidy(arrests_lra)
# scree plot of inertia
tidy(arrests_lra) \%>\%
  ggplot(aes(x = .name, y = .inertia)) +
  theme_bw() +
  geom_col() +
  labs(x = "", y = "Inertia")
# scree plot of proportion of variance (inertia)
tidy(arrests_lra) \%>\%
  ggplot(aes(x = .name, y = .prop_var)) +
  theme_bw() +
  scale_y_continuous(labels = scales::percent) +
  geom_col() +
  labs(x = "", y = "Proportion of inertia")
# fortification adds all above columns
fortify(arrests_lra)
# initially, inertia is conferred on neither factor
get_conference(arrests_lra)
# row-principal biplot with coordinate-wise standard deviations
arrests_lra \%>\%
  confer_inertia(1) \%>\%
  ggbiplot(aes(color = .matrix), sec.axes = "cols", scale.factor = 1/20) +
  scale_color_manual(values = c("brown4", "darkgreen")) +
  theme_bw() +
  geom_rows_text(aes(label = .abb), size = 3, alpha = 2/3) +
  geom_cols_polygon(fill = NA, linetype = "dashed") +
  geom_cols_text(aes(label = .name), fontface = "bold") +
  ggtitle(
    "Log-ratio analysis of violent crime arrest rates",
    "United States, 1973"
  ) +
  guides(color = FALSE, size = FALSE)
}
