% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-nipals.r
\name{methods-nipals}
\alias{methods-nipals}
\alias{as_tbl_ord.nipals_ord}
\alias{recover_rows.nipals_ord}
\alias{recover_cols.nipals_ord}
\alias{recover_inertia.nipals_ord}
\alias{recover_coord.nipals_ord}
\alias{recover_conference.nipals_ord}
\alias{augmentation_rows.nipals_ord}
\alias{augmentation_cols.nipals_ord}
\alias{augmentation_coord.nipals_ord}
\title{Functionality for non-linear iterative PLS ('nipals') objects}
\usage{
\method{as_tbl_ord}{nipals_ord}(x)

\method{recover_rows}{nipals_ord}(x)

\method{recover_cols}{nipals_ord}(x)

\method{recover_inertia}{nipals_ord}(x)

\method{recover_coord}{nipals_ord}(x)

\method{recover_conference}{nipals_ord}(x)

\method{augmentation_rows}{nipals_ord}(x)

\method{augmentation_cols}{nipals_ord}(x)

\method{augmentation_coord}{nipals_ord}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"nipals_ord"}. This is a class introduced in this package
to identify objects returned by \code{\link[=nipals_ord]{nipals_ord()}}, which wraps
\code{\link[nipals:nipals]{nipals::nipals()}}.
}
