% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-pma-cca.r
\name{methods-pma}
\alias{methods-pma}
\alias{as_tbl_ord.CCA}
\alias{recover_coord.CCA}
\alias{recover_rows.CCA}
\alias{recover_cols.CCA}
\alias{recover_inertia.CCA}
\alias{recover_conference.CCA}
\alias{augmentation_rows.CCA}
\alias{augmentation_cols.CCA}
\alias{augmentation_coord.CCA}
\title{Functionality for penalized multivariate analysis ('SPC', 'CCA',
'MultiCCA') objects}
\usage{
\method{as_tbl_ord}{CCA}(x)

\method{recover_coord}{CCA}(x)

\method{recover_rows}{CCA}(x)

\method{recover_cols}{CCA}(x)

\method{recover_inertia}{CCA}(x)

\method{recover_conference}{CCA}(x)

\method{augmentation_rows}{CCA}(x)

\method{augmentation_cols}{CCA}(x)

\method{augmentation_coord}{CCA}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class 'SPC', 'CCA', and 'MultiCCA' from the
\strong{\link[PMA:PMA-package]{PMA}} package.
}
\details{
Witten, Tibshirani, and Hastie (2009) provide a theoretical basis and
computational algorithm for penalized matrix decomposition that specializes
to sparse PCA and to sparse CCA. Their R package \strong{PMA} implements these
specializations as well as one to sparse multiple CCA.

\strong{Note:} These draft methods produce the biplot of Greenacre (1984), which
are advised against by ter Braak (1990).
}
\examples{
# data frame of life-cycle savings across countries
class(LifeCycleSavings)
head(LifeCycleSavings)
# canonical correlation analysis of age distributions and financial factors
savings_cca <- PMA::CCA(
  LifeCycleSavings[, c(2L, 3L)],
  LifeCycleSavings[, c(1L, 4L, 5L)],
  K = 2L, penaltyx = .7, penaltyz = .7,
  xnames = names(LifeCycleSavings)[c(2L, 3L)],
  znames = names(LifeCycleSavings)[c(1L, 4L, 5L)]
)
# wrap as a 'tbl_ord' object
(savings_cca <- as_tbl_ord(savings_cca))
# summarize ordination
glance(savings_cca)
# recover canonical variates
get_rows(savings_cca)
get_cols(savings_cca)
# augment canonical variates with variable names
(savings_cca <- augment_ord(savings_cca))
# summarize canonical correlations
tidy(savings_cca)
# scree plot of canonical correlations
tidy(savings_cca) \%>\%
  ggplot(aes(x = .name, y = .inertia)) +
  geom_col() +
  labs(x = "", y = "Inertia (squared canonical correlations)")
# fortification binds tibbles of canonical variates
fortify(savings_cca)
# column-standard biplot of canonical variates
savings_cca \%>\%
  confer_inertia("rows") \%>\%
  ggbiplot(aes(label = .name, color = .matrix)) +
  theme_bw() +
  geom_origin() +
  geom_unit_circle(linetype = "dotted") +
  geom_cols_vector() +
  geom_cols_text_radiate() +
  geom_rows_point() +
  geom_rows_text_repel() +
  expand_limits(x = c(-1, 1), y = c(-1, 1))
}
\references{
Witten DM, Tibshirani R, & Hastie T (2009) "A penalized matrix decomposition,
with applications to sparse principal components and canonical correlation
analysis". \emph{Biostatistics} \strong{10}(3): 515--534.
\doi{10.1093/biostatistics/kxp008}

Greenacre MJ (1984) \emph{Theory and applications of correspondence analysis}.
London: Academic Press, ISBN 0-12-299050-1.
\url{http://www.carme-n.org/?sec=books5}

ter Braak CJF (1990) "Interpreting canonical correlation analysis through
biplots of structure correlations and weights". \emph{Psychometrika} 55(3),
519--531. \doi{10.1007/BF02294765}
}
