% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-princomp.r
\name{methods-princomp}
\alias{methods-princomp}
\alias{as_tbl_ord.princomp}
\alias{recover_rows.princomp}
\alias{recover_cols.princomp}
\alias{recover_inertia.princomp}
\alias{recover_coord.princomp}
\alias{recover_conference.princomp}
\alias{augmentation_rows.princomp}
\alias{augmentation_cols.princomp}
\alias{augmentation_coord.princomp}
\title{Functionality for principal components analysis ('princomp') objects}
\usage{
\method{as_tbl_ord}{princomp}(x)

\method{recover_rows}{princomp}(x)

\method{recover_cols}{princomp}(x)

\method{recover_inertia}{princomp}(x)

\method{recover_coord}{princomp}(x)

\method{recover_conference}{princomp}(x)

\method{augmentation_rows}{princomp}(x)

\method{augmentation_cols}{princomp}(x)

\method{augmentation_coord}{princomp}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"princomp"} as returned by \code{\link[stats:princomp]{stats::princomp()}}.
}
\examples{
# Scaled PCA of Anderson Iris data with centroid stars
iris[, -5] \%>\%
  princomp(cor = TRUE) \%>\%
  as_tbl_ord() \%>\%
  mutate_rows(species = iris$Species) \%>\%
  mutate_cols(measure = gsub("\\\\.", " ", tolower(names(iris)[-5]))) \%>\%
  print() -> iris_pca
iris_pca \%>\%
  ggbiplot(aes(color = species), sec.axes = "cols", scale.factor = 3) +
  theme_bw() +
  scale_color_brewer(type = "qual", palette = 2) +
  stat_rows_star(alpha = .5, fun.center = "mean") +
  stat_rows_center(size = 3, fun.center = "mean") +
  geom_rows_point(alpha = .5) +
  geom_cols_vector(color = "#444444") +
  ggtitle(
    "Row-principal PCA biplot of Anderson iris data",
    "Segments connect each observation to its within-species centroid"
  )
# Unscaled PCA and two-scaled biplot of Anderson iris data
iris[, -5] \%>\%
  princomp() \%>\%
  as_tbl_ord() \%>\%
  confer_inertia(1) \%>\%
  mutate_rows(species = iris$Species) \%>\%
  mutate_cols(measure = gsub("\\\\.", " ", tolower(names(iris)[-5]))) \%>\%
  print() -> iris_pca
iris_pca \%>\%
  ggbiplot(aes(color = species), sec.axes = "cols", scale.factor = 3) +
  theme_bw() +
  scale_color_brewer(type = "qual", palette = 2) +
  geom_rows_point() +
  geom_cols_vector(color = "#444444") +
  geom_cols_text_radiate(aes(label = measure), color = "#444444") +
  ggtitle(
    "Row-principal unscaled PCA biplot of Anderson iris data",
    "Variable loadings scaled to secondary axes"
  ) +
  expand_limits(y = c(-1, 3.5))
}
\author{
Emily Paul
}
