% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-svd.r
\name{methods-svd}
\alias{methods-svd}
\alias{as_tbl_ord.svd_ord}
\alias{recover_rows.svd_ord}
\alias{recover_cols.svd_ord}
\alias{recover_inertia.svd_ord}
\alias{recover_coord.svd_ord}
\alias{recover_conference.svd_ord}
\alias{augmentation_rows.svd_ord}
\alias{augmentation_cols.svd_ord}
\alias{augmentation_coord.svd_ord}
\title{Functionality for singular value decompositions}
\usage{
\method{as_tbl_ord}{svd_ord}(x)

\method{recover_rows}{svd_ord}(x)

\method{recover_cols}{svd_ord}(x)

\method{recover_inertia}{svd_ord}(x)

\method{recover_coord}{svd_ord}(x)

\method{recover_conference}{svd_ord}(x)

\method{augmentation_rows}{svd_ord}(x)

\method{augmentation_cols}{svd_ord}(x)

\method{augmentation_coord}{svd_ord}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"svd_ord"} returned by \code{\link[=svd_ord]{svd_ord()}}.
}
