% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-nipals.r
\name{nipals-ord}
\alias{nipals-ord}
\title{Augmented implementation of non-linear iterative PLS}
\arguments{
\item{ret.cmeans, ret.csd}{Logical; whether to retain as attributes the column
means and standard deviations used to scale the matrix before running the
nipals algorithm.}
}
\description{
This function replicates \code{\link[ade4:nipals]{ade4::nipals()}} with options to retain
elements useful to the \link{tbl_ord} class and biplot calculations.
}
