% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinate.r
\name{ordinate}
\alias{ordinate}
\alias{ordinate.default}
\alias{ordinate.array}
\alias{ordinate.table}
\alias{ordinate.data.frame}
\alias{ordinate.dist}
\title{Fit an ordination model to a data object}
\usage{
ordinate(x, model, cols, augment, ...)

\method{ordinate}{default}(x, model, ...)

\method{ordinate}{array}(x, model, ...)

\method{ordinate}{table}(x, model, ...)

\method{ordinate}{data.frame}(x, model, cols, augment, ...)

\method{ordinate}{dist}(x, model, ...)
}
\arguments{
\item{x}{A data object to be passed to the \code{model}, such as an \link{array},
\link{table}, \link{data.frame}, or \link[stats:dist]{stats::dist}.}

\item{model}{An ordination function whose output is coercible to class
'\link{tbl_ord}', or a symbol or character string (handled by \code{\link[=match.fun]{match.fun()}}).
Alternatively, a formula \code{~ fun(., ...)} where \code{fun} is such a function and
other arguments are explicit, which will be evaluated with \code{x} in place of
\code{.}.}

\item{cols}{<\code{\link[tidyr:tidyr_tidy_select]{tidy-select}}> If \code{x} is a data
frame, columns to pass to \code{model}. If missing, all columns are used.}

\item{augment}{<\code{\link[=tidyr_tidy_select]{tidy-select}}> If \code{x} is a data frame,
columns to augment to the row data of the ordination. If missing, all
columns not included in \code{cols} will be augmented.}

\item{...}{Additional arguments passed to \code{model}.}
}
\value{
An augmented tbl_ord.
}
\description{
This is a convenience function to fit an ordination model to a
specified subset of columns of a data frame and augment the tbl_ord object
with both its intrinsic diagnostics (via \verb{[augment_ord()]}) and any
additional columns of the data frame (via \verb{[cbind_rows()]}).
}
\details{
(Write this up.)
}
\examples{
# LRA of arrest data
ordinate(USArrests, cols = c(Murder, Rape, Assault), lra)

# CMDS of inter-city distance data
ordinate(UScitiesD, cmdscale_ord, k = 3L)

# PCA of iris data
ordinate(iris, princomp, cols = -Species, augment = c(Sepal.Width, Species))
ordinate(iris, cols = 1:4, ~ prcomp(., center = TRUE, scale. = TRUE))

# CA of hair & eye color data
haireye <- as.data.frame(rowSums(HairEyeColor, dims = 2L))
ordinate(haireye, MASS::corresp, cols = everything())

# FA of Swiss social data
ordinate(swiss, model = factanal, factors = 2L, scores = "Bartlett")

# LDA of iris data
ordinate(iris, ~ lda_ord(.[, 1:4], .[, 5], ret.x = TRUE))

# CCA of savings data
ordinate(
  LifeCycleSavings[, c("pop15", "pop75")],
  # second data set must be handled as an additional parameter to `model`
  y = LifeCycleSavings[, c("sr", "dpi", "ddpi")],
  model = cancor_ord, scores = TRUE
)
}
