% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordr.r
\docType{package}
\name{ordr}
\alias{ordr}
\alias{ordr-package}
\title{\strong{ordr} package}
\description{
This is a \strong{\link[tidyverse:tidyverse]{tidyverse}} extension for
handling, manipulating, and visualizing ordination models with consistent
conventions and in a tidy workflow.
}
\details{
This package is designed to integrate ordination analysis and biplot
visualization into a \strong{tidyverse} workflow. It is inspired in particular by
the extensions \strong{ggbiplot} and \strong{tidygraph}.

The package consists in several modules:
\itemize{
\item the \code{"tbl_ord"} class, a wrapper for various ordination object classes
\item extracting \link{augmentation} for the factors of an ordination
\item using \link{dplyr-verbs} to add \link{annotation} to the factors of an ordination
\item manipulating the coordinates of an ordination via \link{alignment} with other
objects and \link{conference} of inertia
\item methods of the above generics for several widely-used ordination object
classes
\item convenient \link{formatting} of ordination objects that respects the above
manipulations
\item \link{ggbiplot}, a \strong{ggplot2} extension for rendering biplots, including
several plot layers, integrated with the \code{"tbl_ord"} class
}
}
