% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{ratios}
\alias{ratios}
\title{Ingredient ratios for several basic recipes}
\format{
A \link[tibble:tibble]{tibble}.
}
\source{
Ruhlman (2009).
}
\usage{
data(ratios)
}
\description{
Ratios of ingredients for several basic recipes presented in a
popular cooking book.
}
\examples{
# Log-ratio analysis of shared ingredients in Ruhlman's recipe ratios
data(ratios)
ratios \%>\%
  dplyr::select(chapter, recipe, flour, egg, butter) \%>\%
  dplyr::filter(flour > 0 & egg > 0 & butter > 0) \%>\%
  dplyr::group_by(flour, egg, butter) \%>\%
  dplyr::summarize(chapter = unique(chapter),
                   recipes = stringr::str_c(recipe, collapse = " or ")) \%>\%
  dplyr::ungroup() \%>\%
  print() -> sub_ratios
sub_ratios \%>\%
  dplyr::select(-chapter, -recipes) \%>\%
  lra(compositional = TRUE) \%>\%
  as_tbl_ord() \%>\%
  augment() \%>\%
  bind_cols_rows(dplyr::select(sub_ratios, chapter, recipes)) \%>\%
  print() -> lra_ratios
lra_ratios \%>\%
  confer_inertia("rows") \%>\%
  ggbiplot(sec.axes = "v", scale.factor = .05) +
  geom_rows_text(aes(label = recipes, color = chapter)) +
  geom_cols_vector() +
  geom_cols_text(aes(label = .name), hjust = "outward", vjust = "outward") +
  expand_limits(x = c(-.2, .2)) +
  ggtitle("Log-ratio biplot of flour-butter-egg ratios in baking recipes") +
  expand_limits(x = c(-.35, .35))
# Correspondence analysis of Ruhlman's recipe ratios
data(ratios)
ratios \%>\%
  dplyr::select(-chapter) \%>\%
  tibble::column_to_rownames("recipe") \%>\%
  dplyr::select(-chocolate) \%>\%
  ca::ca() \%>\%
  as_tbl_ord() \%>\%
  augment() \%>\%
  mutate_rows(recipe = .name, chapter = ratios$chapter) \%>\%
  mutate_cols(ingredient = .name) \%>\%
  print() -> ratios_ca
ratios_ca \%>\%
  confer_inertia("rowprincipal") \%>\%
  ggbiplot() +
  theme_bw() +
  geom_cols_vector(alpha = .5) +
  stat_cols_chull(geom = "text_radiate", aes(label = ingredient), alpha = .5) +
  geom_rows_text(aes(label = recipe, color = chapter), size = 4) +
  ggtitle("Row-principal CA biplot of Ruhlman recipe ratios")
}
\references{
Ruhlman M (2009) \emph{Ratio: The Simple Codes Behind the Craft of Everyday
Cooking}. Scribner, ISBN-13: 978-1416571728.
}
\keyword{datasets}
