% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.r
\name{reconstruct}
\alias{reconstruct}
\alias{reconstruct.list}
\title{Reconstruct approximations to original data from ordination factors}
\usage{
reconstruct(x)

\method{reconstruct}{list}(x)
}
\arguments{
\item{x}{A bibble, or an object convertible to one.}
}
\description{
Reconstruct approximations to original data from ordination factors
}
