% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{simple_example}
\alias{simple_example}
\title{A simple illustrative biplot example from Greenacre (2010)}
\format{A \link{bibble}.}
\source{
Greenacre M.J. (2010). \emph{Biplots in Practice}, pp. 16--20. 
  Fundacion BBVA.
}
\description{
This is a simple example of a biplot, a \eqn{5\times 4} matrix that
decomposes into the product of a \eqn{5\times 2} matrix and a \eqn{2\times 4}
matrix.
}
\keyword{datasets}
