% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-spantree.r
\docType{data}
\name{stat-biplot-spantree}
\alias{stat-biplot-spantree}
\alias{StatSpantree}
\alias{stat_spantree}
\alias{StatRowsSpantree}
\alias{StatColsSpantree}
\alias{stat_rows_spantree}
\alias{stat_cols_spantree}
\title{Calculate a minimum spanning tree among cases or variables}
\format{
An object of class \code{StatSpantree} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.

An object of class \code{StatRowsSpantree} (inherits from \code{StatSpantree}, \code{Stat}, \code{ggproto}, \code{gg}) of length 2.

An object of class \code{StatColsSpantree} (inherits from \code{StatSpantree}, \code{Stat}, \code{ggproto}, \code{gg}) of length 2.
}
\usage{
stat_spantree(
  mapping = NULL,
  data = NULL,
  geom = "segment",
  position = "identity",
  method = "euclidean",
  show.legend = NA,
  inherit.aes = TRUE,
  check.aes = TRUE,
  ...
)

stat_rows_spantree(
  mapping = NULL,
  data = NULL,
  geom = "segment",
  position = "identity",
  method = "euclidean",
  show.legend = NA,
  inherit.aes = TRUE,
  check.aes = TRUE,
  ...
)

stat_cols_spantree(
  mapping = NULL,
  data = NULL,
  geom = "segment",
  position = "identity",
  method = "euclidean",
  show.legend = NA,
  inherit.aes = TRUE,
  check.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{The geometric object to use display the data}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{method}{Passed to \code{\link[stats:dist]{stats::dist()}}.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{check.aes}{If \code{TRUE}, the default, will check that
supplied parameters and aesthetics are understood by the \code{geom} or
\code{stat}. Use \code{FALSE} to suppress the checks.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:layer]{ggplot2::layer()}}.}
}
\description{
This stat layer identifies the \eqn{n-1} pairs among \eqn{n}
points that form a minimum spanning tree, then calculates the segments
between these poirs in the two dimensions \code{x} and \code{y}.
}
\details{
A minimum spanning tree (MST) on the point cloud \eqn{X} is a minimal
connected graph on \eqn{X} with the smallest possible sum of distances (or
dissimilarities) between linked points. These layers call \code{\link[stats:dist]{stats::dist()}} to
calculate a distance/dissimilarity object and \code{\link[vegan:spantree]{vegan::spantree()}} to
calculate the MST. The result is formatted with position aesthetics readable
by \code{\link[ggplot2:geom_segment]{ggplot2::geom_segment()}}.

If any aesthetics of the form \verb{.coord[0-9]+} are detected, then the lot of
them are used to calculate distances/dissimilarities. These should not be
assigend manually but generated using the convenience function \code{\link[=ord_aes]{ord_aes()}}
(see the examples). Otherwise, \code{x} and \code{y} are used. Either way, \code{x} and \code{y}
provide the position aesthetics.

An MST calculated on \code{x} and \code{y} reflects the distances among the points in
\eqn{X} in the reduced-dimension plane of the biplot. In contrast, one
calculated on the full set of coordinates reflects distances in
higher-dimensional space. Plotting this high-dimensional MST on the
2-dimensional biplot provides a visual cue as to how faithfully two
dimensions can encapsulate the "true" distances between points (Jolliffe,
2002).
}
\section{Biplot layers}{

\code{\link[=ggbiplot]{ggbiplot()}} uses \code{\link[ggplot2:fortify]{ggplot2::fortify()}} internally to produce a single data
frame with a \code{.matrix} column distinguishing the subjects (\code{"rows"}) and
variables (\code{"cols"}). The stat layers \code{stat_rows()} and \code{stat_cols()} simply
filter the data frame to one of these two.

The geom layers \verb{geom_rows_*()} and \verb{geom_cols_*()} call the corresponding
stat in order to render plot elements for the corresponding factor matrix.
\verb{geom_biplot_*()} selects a default matrix based on common practice, e.g.
points for rows and arrows for columns.
}

\examples{
# Compare MDS to the spanning tree for road distances between European cities
eurodist \%>\%
  cmdscale_ord(k = 11) \%>\%
  as_tbl_ord() \%>\%
  augment() \%>\%
  # align to conventional compass orientation
  negate_ord(2) \%>\%
  print() -> euro_mds
euro_mds \%>\%
  ggbiplot() +
  stat_cols_spantree(
    ord_aes(euro_mds), check.aes = FALSE,
    alpha = .5, linetype = "dotted"
  ) +
  geom_cols_text(aes(label = .name), size = 3) +
  ggtitle("MDS biplot of road distances between European cities")
}
\references{
Jolliffe IT (2002) \emph{Principal Component Analysis}, Second Edition.
Springer Series in Statistics, ISSN 0172-7397. \doi{10.1007/b98835}
\url{https://www.springer.com/gp/book/9780387954424}
}
\keyword{datasets}
