% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-rule.r
\name{stat_rule}
\alias{stat_rule}
\alias{minpp}
\alias{maxpp}
\alias{minabspp}
\title{Construct limited rulers offset from the origin}
\usage{
stat_rule(
  mapping = NULL,
  data = NULL,
  geom = "axis",
  position = "identity",
  .referent = NULL,
  referent = NULL,
  fun.min = minpp,
  fun.max = maxpp,
  fun.offset = minabspp,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)

minpp(x, p = 0.1)

maxpp(x, p = 0.1)

minabspp(x, p = 0.1)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{The geometric object to use to display the data for this layer.
When using a \verb{stat_*()} function to construct a layer, the \code{geom} argument
can be used to override the default coupling between stats and geoms. The
\code{geom} argument accepts the following:
\itemize{
\item A \code{Geom} ggproto subclass, for example \code{GeomPoint}.
\item A string naming the geom. To give the geom as a string, strip the
function name of the \code{geom_} prefix. For example, to use \code{geom_point()},
give the geom as \code{"point"}.
\item For more information and other ways to specify the geom, see the
\link[ggplot2:layer_geoms]{layer geom} documentation.
}}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{.referent}{A character vector indicating the matrix factor(s) of an
ordination model to include in \code{referent}; should be a subset of \code{c("rows", "cols")}.}

\item{referent}{The point cloud to rule; a data frame with \code{x} and \code{y}
columns.}

\item{fun.min, fun.max}{Functions used to determine the limits of rules from
the projections of \code{referent} onto the axes.}

\item{fun.offset}{Function used to determine the directions and magnitudes of
the axis offsets from the projections of \code{referent} onto the normal vectors
of the axes.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:layer]{ggplot2::layer()}}.}

\item{x}{A numeric vector.}

\item{p}{A numeric value; the proportion of a range used as a buffer.}
}
\value{
A ggproto \link[ggplot2:layer]{layer}.
}
\description{
Determine axis limits and offset vectors from reference data.
}
\section{Biplot layers}{

\code{\link[=ggbiplot]{ggbiplot()}} uses \code{\link[ggplot2:fortify]{ggplot2::fortify()}} internally to produce a single data
frame with a \code{.matrix} column distinguishing the subjects (\code{"rows"}) and
variables (\code{"cols"}). The stat layers \code{stat_rows()} and \code{stat_cols()} simply
filter the data frame to one of these two.

The geom layers \verb{geom_rows_*()} and \verb{geom_cols_*()} call the corresponding
stat in order to render plot elements for the corresponding factor matrix.
\verb{geom_dims_*()} selects a default matrix based on common practice, e.g.
points for rows and arrows for columns.
}

\section{Computed variables}{
 These are calculated during the statistical
transformation and can be accessed with \link[ggplot2:aes_eval]{delayed evaluation}.
\describe{
\item{\verb{x,y}}{cartesian coordinates (if passed polar)}
\item{\verb{angle,radius}}{polar coordinates (if passed cartesian)}
\item{\verb{lower,upper}}{distances of rule endpoints from origin}
\item{\verb{yintercept,xintercept}}{intercepts of offset axis}
\item{\code{axis}}{unique axis identifier}
}
}

\examples{
# Freestone primary glass measurements
print(glass)

# default (standardized) linear discriminant analysis of sites on measurements
glass_lda <- MASS::lda(Site ~ SiO2 + Al2O3 + FeO + MgO + CaO, glass)
# bestow 'tbl_ord' class & augment observation, centroid, and variable fields
as_tbl_ord(glass_lda) \%>\%
  augment_ord() \%>\%
  mutate_rows(discriminant =
                ifelse(.element == "active", "centroid", "case")) \%>\%
  print() -> glass_lda
# row-standard biplot
glass_lda \%>\%
  confer_inertia(1) \%>\%
  ggbiplot() +
  theme_bw() + theme_biplot() +
  geom_rows_point(aes(shape = grouping, size = discriminant), alpha = .5) +
  stat_cols_rule(aes(label = name), color = "#888888", num = 8L,
                 fun.offset = \(x) minabspp(x, p = .2),
                 text_size = 2.5, label_dodge = .02) +
  scale_shape_manual(values = c(2L, 3L, 0L, 5L)) +
  ggtitle(
    "LDA of Freestone glass measurements",
    "Row-standard biplot of standardized LDA"
  )

# contribution LDA of sites on measurements
glass_lda <-
  lda_ord(Site ~ SiO2 + Al2O3 + FeO + MgO + CaO, glass,
          axes.scale = "contribution")
# bestow 'tbl_ord' class & augment observation, centroid, and variable fields
as_tbl_ord(glass_lda) \%>\%
  augment_ord() \%>\%
  mutate_rows(discriminant =
                ifelse(.element == "active", "centroid", "case")) \%>\%
  print() -> glass_lda
# symmetric biplot
glass_lda \%>\%
  confer_inertia(.5) \%>\%
  ggbiplot() +
  theme_bw() + theme_biplot() +
  geom_rows_point(aes(shape = grouping, alpha = discriminant)) +
  stat_cols_rule(aes(label = name), color = "#888888", num = 8L,
                 fun.offset = \(x) minabspp(x, p = .1),
                 text_size = 2.5, text_dodge = .025) +
  scale_shape_manual(values = c(16L, 17L, 15L, 18L)) +
  ggtitle(
    "LDA of Freestone glass measurements",
    "Symmetric biplot of contribution LDA"
  )
}
\seealso{
Other stat layers: 
\code{\link{stat_center}()},
\code{\link{stat_chull}()},
\code{\link{stat_cone}()},
\code{\link{stat_scale}()},
\code{\link{stat_spantree}()}
}
\concept{stat layers}
