% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.r
\name{theme_biplot}
\alias{theme_biplot}
\title{Biplot theme}
\usage{
theme_biplot()
}
\value{
A ggplot \link[ggplot2:theme]{theme}.
}
\description{
Omit coordinate visual aids from biplots.
}
\details{
Because the artificial axes often go uninterpreted, biplots may omit the
visual aids (tick marks and labels, grid lines) used to recover the
artificial coordinates of the row and column markers The biplot (partial)
theme removes these elements from the current theme. This can be especially
helpful when plotting \link[=geom_axis]{axes} or \link[=geom_isoline]{isolines}.
}
