% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-scaffold.r
\name{theme_scaffold}
\alias{theme_scaffold}
\alias{theme_biplot}
\title{Scaffolding theme}
\usage{
theme_scaffold()

theme_biplot()
}
\value{
A ggplot \link[ggplot2:theme]{theme}.
}
\description{
Omit cartesian coordinate visual aids.
}
\details{
Geometric data analysis concerns the intrinsic geometry of data.
Analyses often use artificial or arbitrary coordinate systems that carry no
useful interpretation but instead serve as scaffolding, especially for
graphical elements like \link[=geom_axis]{axes} that represent other variables
(Gardner, 2001). In such cases, the visual aids (tick marks and labels,
grid lines) used to recover the coordinates of the row and column markers
would add unnecessary clutter and should be omitted. This partial theme
updates the current theme by removing these elements. The biplot theme is
an alias included for convenience and backward compatibility.
}
\references{
Gardner S (2001) \emph{Extensions of biplot methodology to discriminant analysis
with applications of non-parametric principal components}. PhD thesis,
Stellenbosch University.
\url{https://scholar.sun.ac.za/items/279f7958-0b54-43f1-8c75-da652f65db3f}
}
