% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ord-accessors.r
\name{accessors}
\alias{accessors}
\alias{get_rows}
\alias{get_cols}
\alias{as.matrix.tbl_ord}
\alias{get_coord}
\alias{get_inertia}
\alias{dim.tbl_ord}
\title{Access factors, coordinates, and metadata from ordination objects}
\usage{
get_rows(x, elements = "all")

get_cols(x, elements = "all")

\method{as.matrix}{tbl_ord}(x, ..., .matrix, elements = "all")

get_coord(x)

get_inertia(x)

\method{dim}{tbl_ord}(x)
}
\arguments{
\item{x}{An object of class '\link{tbl_ord}'.}

\item{elements}{Character; which elements of each factor for which to render
graphical elements. One of \code{"all"} (the default), \code{"active"}, or
\code{"supplementary"}, with partial matching.}

\item{...}{Additional arguments from \code{\link[base:matrix]{base::as.matrix()}}; ignored.}

\item{.matrix}{A character string matching one of several indicators for one
or both matrices in a matrix decomposition used for ordination. The
standard values are \code{"rows"}, \code{"cols"}, and \code{"dims"} (for both).}
}
\description{
These functions return information about the matrix
factorization underlying an ordination.
}
\details{
The unexported \verb{recover_*()} functions extract one or both of the row and
column matrix factors that constitute the original ordination. These are
interpreted as the case scores (rows) and the variable loadings (columns).
The \verb{get_*()} functions optionally (and by default) include any supplemental
observations (see \link{supplementation}). The \verb{recover_*()} functions are
generics that require methods for each ordination class.

\code{get_coord()} retrieves the names of the coordinates shared by the matrix
factors on which the original data were ordinated, and \code{get_inertia()}
retrieves a vector of the inertia with these names. \code{dim()} retrieves the
dimensions of the row and column factors, which reflect the dimensions of the
matrix they reconstruct---\strong{not} the original data matrix. (This matters for
techniques that rely on eigendecomposition, for which the decomposed matrix
is square.)
}
\examples{
# example ordination: LRA of U.S. arrests data
arrests_lra <- ordinate(USArrests, cols = c(Murder, Rape, Assault), lra)
# extract matrix factors
as.matrix(arrests_lra, .matrix = "rows")
as.matrix(arrests_lra, .matrix = "cols")
# special named functions
get_rows(arrests_lra)
get_cols(arrests_lra)
# get names of artificial / latent coordinates
get_coord(arrests_lra)
# get distribution of inertia
get_inertia(arrests_lra)
# get dimensions of underlying matrix factorization (not of original data)
dim(arrests_lra)
}
