% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ord-annotation.r
\name{annotation}
\alias{annotation}
\alias{set_annotation_u}
\alias{set_annotation_v}
\alias{annotation_u}
\alias{annotation_v}
\title{Annotate the factors of a tbl_ord}
\usage{
set_annotation_u(x, annot)

set_annotation_v(x, annot)

annotation_u(x)

annotation_v(x)
}
\arguments{
\item{x}{A \link{tbl_ord}, or an ordination object coercible to one.}

\item{annot}{A \link[base:data.frame]{data.frame} having the same number of rows
as \code{get_*(x)}.}
}
\description{
These functions annotate the factors \eqn{U} and \eqn{V} of
\link{tbl_ord} objects with additional variables, and retrieve these
annotations.

The \verb{annotation_*()} and \verb{set_annotation_*()} functions assign and retrieve
values of the \code{"*_annotation"} attributes of \code{x}, which should have the same
number of rows as \code{get_*(x)}.
}
