% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmentation.r
\name{augmentation}
\alias{augmentation}
\alias{augmentation_u}
\alias{augmentation_v}
\alias{augmentation_factor}
\alias{augmentation.tbl_ord}
\alias{augmentation_coord}
\alias{augment_u}
\alias{augment_v}
\alias{augment_coord}
\title{Augment information for the factors and coordinates of ordination objects}
\usage{
augmentation_u(x)

augmentation_v(x)

augmentation_factor(x, .matrix)

augmentation.tbl_ord(x, .matrix)

augmentation_coord(x)

augment_u(x)

augment_v(x)

augment_coord(x)
}
\arguments{
\item{x}{A \code{tbl_ord}, or an ordination object coercible to one.}

\item{.matrix}{A character string matching one of several indicators for one
or both matrices in a matrix decomposition used for ordination. The
standard values are \code{"u"}, \code{"v"}, and \code{"uv"}.}
}
\description{
These functions return data associated with the cases, variables, and
coordinates of an ordination object.
}
\details{
The \code{augmentation_*} functions produce \link[tibble]{tibble}s of values
associated with the cases, variables, and coordinates of a \code{tbl_ord}
object. The first field of each tibble is \code{.name}, which contains the
case, variable, or coordinate names. Additional fields contain information
about the cases, variables, or coordinates extracted from the original
ordination object.

The \code{augment_*} functions return the ordination coordinates data
augmented with the result of \code{augmentation_*}. In this sense
\code{augment_*} works like \code{\link[broom]{augment}} in extracting all
information that can be included in a tidy summary of the components. The
advantage of implementing separate methods for the different components is
that more information contained in the original object becomes accessible to
the user.
}
