% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibble-annotation.r
\name{bibble-annotation}
\alias{bibble-annotation}
\alias{u_annot}
\alias{v_annot}
\alias{factor_annot}
\alias{coord_annot}
\title{Recover annotation for an ordination object}
\usage{
u_annot(x)

v_annot(x)

factor_annot(x, .matrix)

coord_annot(x)
}
\arguments{
\item{x}{A bibble, or an object convertible to one.}

\item{.matrix}{A character string matching one of several indicators for one
or both matrices in a matrix decomposition used for ordination. The
standard values are \code{"u"}, \code{"v"}, and \code{"uv"}.}
}
\description{
These functions return annotation associated with the subjects, variables,
and coordinates of an ordination.
}
\details{
The \code{*_annot} functions produce \link[tibble]{tibble}s of annotations 
for the subjects/scores, variables/loadings, and coordinates. The first field
of each tibble is \code{.name} and contains the subject, variable, and 
coordinate names, respectively.

The \code{*_annot} functions return annotation from two sources:
\enumerate{
  \item Information about the subjects, variables, or coordinates contained
        in the original ordination object. In this sense \code{*_annot} works
        like \code{\link[broom]{tidy}} to extract all such information that
        can be included in a tidy summary of the components. The advantage of
        implementing separate methods for the different components is that
        more information contained in the original object becomes accessible
        to the user.
  \item Additional annotation about the components manually added to the
        bibble using adapted \strong{\link[dplyr]{dplyr}} verbs.
        \strong{These are not yet implemented.}
}
Each \code{*_annot} function retrieves both sources and combines them using
\code{link[dplyr]{bind_cols}}.
}
