% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibble-factors.r
\name{bibble-factors}
\alias{bibble-factors}
\alias{get_u}
\alias{get_v}
\alias{get_factor}
\alias{get_coord}
\alias{u_annot}
\alias{v_annot}
\alias{factor_annot}
\alias{coord_annot}
\alias{rank}
\title{Extract the factors of an ordination object}
\usage{
get_u(x)

get_v(x)

get_factor(x, .matrix)

get_coord(x)

u_annot(x)

v_annot(x)

factor_annot(x, .matrix)

coord_annot(x)

rank(x)
}
\arguments{
\item{x}{A bibble, or an object convertible to one.}

\item{.matrix}{A character string matching one of several indicators for one
or both matrices in a matrix decomposition used for ordination. The
standard values are \code{"u"}, \code{"v"}, and \code{"uv"}.}
}
\description{
The \code{get_*} functions extract one or both of the matrix factors 
\eqn{U,V} that constitute an ordination. These are interpreted as the subject
scores (\eqn{U}) and the variable loadings (\eqn{V}). \code{get_coord} 
retrieves the names of the coordinates shared by \eqn{U} and \eqn{V}.
}
\details{
The \code{*_annot} functions produce \link[tibble]{tibble}s of annotations 
for the subjects/scores, variables/loadings, and coordinates. The first field
of each tibble is \code{.name} and contains the subject, variable, and 
coordinate names, respectively.

\code{rank} retrieves the number of coordinates on which the data have been 
ordinated. The existing \code{\link[base]{rank}} function is made the default
and a new method is introduced for bibbles.
}
