% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibble-factors.r, R/methods-cmds.r,
%   R/methods-lm.r, R/methods-lpca.r
\name{bibble-factors}
\alias{bibble-factors}
\alias{get_u}
\alias{get_v}
\alias{get_u.default}
\alias{get_v.default}
\alias{get_coord.default}
\alias{get_u.data.frame}
\alias{get_v.data.frame}
\alias{get_factor}
\alias{as.matrix.bbl}
\alias{get_coord}
\alias{dim.bbl}
\alias{get_u.cmds}
\alias{get_v.cmds}
\alias{get_coord.cmds}
\alias{get_u.lm}
\alias{get_v.lm}
\alias{get_coord.lm}
\alias{get_u.mlm}
\alias{get_v.mlm}
\alias{get_coord.mlm}
\alias{get_u.lpca}
\alias{get_v.lpca}
\alias{get_coord.lpca}
\alias{get_u.lsvd}
\alias{get_v.lsvd}
\alias{get_coord.lsvd}
\title{Extract the factors of an ordination object}
\usage{
get_u(x)

get_v(x)

\method{get_u}{default}(x)

\method{get_v}{default}(x)

\method{get_coord}{default}(x)

\method{get_u}{data.frame}(x)

\method{get_v}{data.frame}(x)

get_factor(x, .matrix)

\method{as.matrix}{bbl}(x, ..., .matrix)

get_coord(x)

\method{dim}{bbl}(x)

\method{get_u}{cmds}(x)

\method{get_v}{cmds}(x)

\method{get_coord}{cmds}(x)

\method{get_u}{lm}(x)

\method{get_v}{lm}(x)

\method{get_coord}{lm}(x)

\method{get_u}{mlm}(x)

\method{get_v}{mlm}(x)

\method{get_coord}{mlm}(x)

\method{get_u}{lpca}(x)

\method{get_v}{lpca}(x)

\method{get_coord}{lpca}(x)

\method{get_u}{lsvd}(x)

\method{get_v}{lsvd}(x)

\method{get_coord}{lsvd}(x)
}
\arguments{
\item{x}{A bibble, or an object convertible to one.}

\item{.matrix}{A character string matching one of several indicators for one
or both matrices in a matrix decomposition used for ordination. The
standard values are \code{"u"}, \code{"v"}, and \code{"uv"}.}

\item{...}{Additional arguments from \code{as.matrix}; ignored.}
}
\description{
These functions return information about the matrix factorization underlying
an ordination.
}
\details{
The \code{get_*} functions extract one or both of the matrix factors 
\eqn{U,V} that constitute an ordination. These are interpreted as the subject
scores (\eqn{U}) and the variable loadings (\eqn{V}). \code{get_coord} 
retrieves the names of the coordinates shared by \eqn{U} and \eqn{V}.

\code{dim.bbl} retrieves the number of coordinates on which the data have
been ordinated.
}
