% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ord-conference.r
\name{conference}
\alias{conference}
\alias{recover_conference}
\alias{recover_conference.default}
\alias{get_conference}
\alias{revert_conference}
\alias{confer_inertia}
\title{Confer inertia to factors of an ordination}
\usage{
recover_conference(x)

\method{recover_conference}{default}(x)

get_conference(x)

revert_conference(x)

confer_inertia(x, p)
}
\arguments{
\item{x}{A \link{tbl_ord} object.}

\item{p}{Numeric vector of length 1 or 2. If length 1, the proportion of the
inertia assigned to the cases, with the remainder \code{1 - p} assigned to the
variables. If length 2, the proportions of the inertia assigned to the
cases and to the variables, respectively.}
}
\description{
Re-distribute inertia between cases and variables in an
ordination.
}
\details{
The \emph{inertia} of a singular value decomposition consists in the squares of
the singular values, and represents the variance, likened to the physical
inertia, in the directions of the orthogonal singular vectors. Read more
about conferring inertia between cases and variables at
\url{https://stats.stackexchange.com/a/141755/68743}.

\emph{Note: In case the \code{"inertia"} attribute is a rectangular matrix, one may
only be able to confer it entirely to the cases (\code{p = 1}) or entirely to the
variables (\code{p = 0}).}
}
