% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conference.r
\name{conference}
\alias{conference}
\alias{confer_inertia}
\title{Re-distribute inertia between cases and variables in an ordination}
\usage{
confer_inertia(x, p)
}
\arguments{
\item{x}{A \code{tbl_ord}.}

\item{p}{Numeric vector of length 1 or 2. If length 1, the proportion of the 
inertia assigned to the cases, with the remainder \code{1 - p} assigned to 
the variables. If length 2, the proportions of the inertia assigned to the 
cases and to the variables, respectively.}
}
\description{
Note: In case \code{inertia} is a rectangular matrix, may only be able to 
confer it entirely to the cases (\code{p = 1}) or entirely to the variables
(\code{p = 0}).
}
