% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{countries}
\alias{countries}
\alias{country_differences}
\alias{country_attributes}
\title{Student ratings of similarities and attributes of 13 countries}
\format{
\link[base:matrix]{Matrices}.
}
\source{
Greenacre (2010), pp. 43--44,46.
}
\usage{
country_differences

data(country_differences)

data(country_attributes)
}
\description{
Greenacre (2010) used these ratings to illustrate
multidimensional scaling. The uses shown there are reproduced in an
example.
}
\examples{
# Multidimensional scaling of country differences and regression of attributes
# Reproduce Exhibit 4.2 in Greenacre (2010)
country_differences \%>\%
  cmdscale(k = 2) \%>\%
  as_tbl_ord() \%>\%
  print() -> differences_cmds
differences_plot <- differences_cmds \%>\%
  ggbiplot(aes(x = 1, y = 2, label = .name)) +
  geom_v_text()
differences_plot
# Reproduce Exhibit 4.5 in Greenacre (2010)
lm(country_attributes ~ get_u(differences_cmds)) \%>\%
  as_tbl_ord() \%>\%
  print() -> attributes_fit
differences_plot +
  geom_v_vector(data = attributes_fit) +
  geom_v_text_radiate(data = attributes_fit, hjust = .3)
# Confer PCA weights and inertia between cases and variables
country_attributes \%>\%
  prcomp() \%>\%
  as_tbl_ord() \%>\%
  print() -> attributes_pca
# form biplot
attributes_pca \%>\%
  ggbiplot(
    aes(label = .name),
    sec.axes = "v", scale.factor = dim_v(attributes_pca)
  ) +
  theme_bw() + theme(panel.grid = element_blank()) +
  geom_u_text(aes(x = PC1, y = PC2), color = "seagreen") +
  geom_v_vector(aes(x = PC1, y = PC2), color = "darkred") +
  geom_v_text(
    aes(x = PC1, y = PC2),
    color = "darkred", hjust = "outward", vjust = "outward"
  )
# covariance biplot
attributes_pca \%>\%
  confer_inertia(0) \%>\%
  ggbiplot(
    aes(label = .name),
    sec.axes = "u", scale.factor = dim_u(attributes_pca)
  ) +
  theme_bw() + theme(panel.grid = element_blank()) +
  geom_u_text(aes(x = PC1, y = PC2), color = "seagreen") +
  geom_v_vector(aes(x = PC1, y = PC2), color = "darkred") +
  geom_v_text_repel(aes(x = PC1, y = PC2), color = "darkred")
# Align an MDS ordination to a PCA ordination
country_differences \%>\%
  cmdscale(k = 2) \%>\%
  as_tbl_ord() \%>\%
  print() -> differences_cmds
country_differences \%>\%
  prcomp() \%>\%
  as_tbl_ord() \%>\%
  print() -> differences_pca
# negate PCA cases to MDS coordinates
differences_cmds \%>\%
  negate_to(differences_pca, "u") \%>\%
  print() -> differences_align_u
plot(gridExtra::arrangeGrob(grobs = list(
  ggbiplot(differences_cmds) + geom_u_point(),
  ggbiplot(differences_pca) + geom_u_point(),
  ggbiplot(differences_align_u) + geom_u_point()
), ncol = 3))
# negate PCA variables to MDS coordinates
differences_cmds \%>\%
  negate_to(differences_pca, "v") \%>\%
  print() -> differences_align_v
plot(gridExtra::arrangeGrob(grobs = list(
  ggbiplot(differences_cmds) + geom_v_vector(),
  ggbiplot(differences_pca) + geom_v_vector(),
  ggbiplot(differences_align_v) + geom_v_vector()
), ncol = 3))
}
\references{
Greenacre MJ (2010) \emph{Biplots in Practice}. Fundacion BBVA, ISBN:
978-84-923846.
\url{https://www.fbbva.es/microsite/multivariate-statistics/biplots.html}
}
\keyword{datasets}
