% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{countries}
\alias{countries}
\alias{country_differences}
\alias{country_attributes}
\title{Student ratings of similarities and attributes of 13 countries}
\format{\link[tibble]{tibble}s.}
\source{
Greenacre M.J. (2010). \emph{Biplots in Practice}, pp. 43--44,46. 
  Fundacion BBVA.
}
\usage{
country_differences

data(country_differences)
}
\description{
Greenacre (2010) used these ratings to illustrate multidimensional scaling. 
The uses shown there are reproduced in an example.
}
\keyword{datasets}
