% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-verbs.r
\name{dplyr-verbs}
\alias{dplyr-verbs}
\alias{pull_factor}
\alias{pull_u}
\alias{pull_v}
\alias{rename_u}
\alias{rename_v}
\alias{select_u}
\alias{select_v}
\alias{mutate_u}
\alias{mutate_v}
\alias{transmute_u}
\alias{transmute_v}
\alias{bind_cols_u}
\alias{bind_cols_v}
\title{*dplyr* verbs}
\usage{
pull_factor(.data, var = -1, .matrix)

pull_u(.data, var = -1)

pull_v(.data, var = -1)

rename_u(.data, ...)

rename_v(.data, ...)

select_u(.data, ...)

select_v(.data, ...)

mutate_u(.data, ...)

mutate_v(.data, ...)

transmute_u(.data, ...)

transmute_v(.data, ...)

bind_cols_u(.data, ...)

bind_cols_v(.data, ...)
}
\arguments{
\item{.data}{A \code{\link{tbl_ord}}.}

\item{var}{A variable specified as in \code{\link[dplyr]{pull}}.}

\item{.matrix}{A character string matching one of several indicators for one
or both matrices in a matrix decomposition used for ordination. The
standard values are \code{"u"}, \code{"v"}, and \code{"uv"}.}

\item{...}{Comma-separated unquoted expressions as in, e.g.,
\code{\link[dplyr]{select}}.}
}
\description{
These functions adapt \strong{\link[dplyr]{dplyr}} verbs to the factors of a 
\code{tbl_ord}. The raw verbs are not defined for \code{tbl_ord}s; instead, 
each verb has two analogues, corresponding to the two matrix factors. They 
each rely on a common workhorse function, which takes the composition of the 
\strong{dplyr} verb with \code{\link{fortify}}, applied to the factor, 
removes any variables corresponding to coordinates or produced by 
\code{augment_*} (see \code{\link{augmentation}}), and only then assigns it
as the new \code{"*_annot"} attribute of \code{.data}.
}
