% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{finches}
\alias{finches}
\title{Site-species occurrence data for Galapagos finches}
\format{A \link[base:data.frame]{data.frame}.}
\source{
These data are taken from Sanderson (2000). They are available in a
different format in the \strong{\link[cooccur:cooccur]{cooccur}} package.
}
\description{
Site-species occurrence data for Galapagos finches
}
\examples{

data(finches)

finches_mat <- t(finches)

finches_lsvd <- as_tbl_ord(logisticSVD(finches_mat))
ggbiplot(finches_lsvd, aes(x = LSC1, y = LSC2)) +
  geom_u_vector(aes(x = LSC1 * .01, y = LSC2 * .01)) +
  geom_u_text(aes(x = LSC1 * .01, y = LSC2 * .01, label = .name), size = 3) +
  geom_v_label(aes(label = .name), size = 3, alpha = .5)

finches_lpca <- as_tbl_ord(logisticPCA(finches_mat))
ggbiplot(finches_lpca, aes(x = LPC1, y = LPC2)) +
  geom_u_vector(aes(x = LPC1 * .02, y = LPC2 * .02)) +
  geom_u_text_radiate(aes(x = LPC1 * .02, y = LPC2 * .02, label = .name),
                      size = 3) +
  geom_v_label(aes(label = .name), size = 3, alpha = .5)

finches_clpca <- as_tbl_ord(convexLogisticPCA(finches_mat[, -13, drop = FALSE]))
ggbiplot(finches_clpca, aes(x = LPC1, y = LPC2)) +
  geom_u_vector(aes(x = LPC1 * .02, y = LPC2 * .02)) +
  geom_u_text_radiate(aes(x = LPC1 * .02, y = LPC2 * .02, label = .name),
                      size = 3) +
  geom_v_label(aes(label = .name), size = 3, alpha = .5)
}
\references{
Sanderson JG (2000) "Testing Ecological Patterns." \emph{American Scientist},
\strong{88}(4), 332--339. \doi{10.1511/2000.4.332}
}
\keyword{datasets}
