% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ord-format.r
\name{format}
\alias{format}
\alias{format.tbl_ord}
\alias{print.tbl_ord}
\title{Format a tbl_ord for printing}
\usage{
\method{format}{tbl_ord}(
  x,
  width = NULL,
  ...,
  n = NULL,
  max_extra_cols = NULL,
  max_footer_lines = NULL
)

\method{print}{tbl_ord}(
  x,
  width = NULL,
  ...,
  n = NULL,
  max_extra_cols = NULL,
  max_footer_lines = NULL
)
}
\arguments{
\item{x}{A \link{tbl_ord} object.}

\item{width}{Width of text output to generate. This defaults to \code{NULL}, which
means use \code{getOption("tibble.width")} or (if also \code{NULL})
\code{getOption("width")}; the latter displays only the columns that fit on one
screen. You can also set \code{options(tibble.width = Inf)} to override this
default and always print all columns, this may be slow for very wide tibbles.}

\item{...}{Additional arguments.}

\item{n}{Number of rows to show. If \code{NULL}, the default, will print all rows
if less than option \code{tibble.print_max}. Otherwise, will print
\code{tibble.print_min} rows.}
}
\description{
These methods of \code{\link[base:format]{base::format()}} and \code{\link[base:print]{base::print()}} render a
(usually more) tidy readout of a \link{tbl_ord} that is consistent across all
original ordination classes.
}
\details{
The \code{format} and \code{print} methods for class 'tbl_ord' are adapted from those
for class '\link[tibble:tbl_df-class]{tbl_df}' and for class
'\link[tidygraph:tbl_graph]{tbl_graph}'.

\strong{Note:} The \code{format()} function is tedius but cannot be easily modularized
without invoking \link{accessors}, \link{annotation}, and \link{augmentation} multiple
times, thereby significantly reducing performance.
}
