% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot-legend.r
\name{ggbiplot-draw-key}
\alias{ggbiplot-draw-key}
\alias{draw_key_line}
\alias{draw_key_crosslines}
\alias{draw_key_crosspoint}
\title{Biplot key drawing functions}
\usage{
draw_key_line(data, params, size)

draw_key_crosslines(data, params, size)

draw_key_crosspoint(data, params, size)
}
\arguments{
\item{data}{A single row data frame containing the scaled aesthetics to
display in this key}

\item{params}{A list of additional parameters supplied to the geom.}

\item{size}{Width and height of key in mm.}
}
\value{
A grid grob.
}
\description{
These key drawing functions supplement those built into
\strong{\link[ggplot2:ggplot2]{ggplot2}} for producing legends suitable to biplots.
}
\details{
\code{draw_key_line()} is a horizontal counterpart to \code{\link[ggplot2:draw_key_vline]{ggplot2::draw_key_vline()}}.
\code{draw_key_crosslines()} superimposes these two keys, and
\code{draw_key_crosspoint()} additionally superimposes an oversized
\code{\link[ggplot2:draw_key_point]{ggplot2::draw_key_point()}}.
}
\examples{
# Anderson iris sepal and petal data
head(iris)
# PCA ordination
iris[, -5] \%>\%
  prcomp(scale = TRUE) \%>\%
  as_tbl_ord() \%>\%
  # row-principal conference of intertia
  confer_inertia(1) \%>\%
  mutate_u(species = iris$Species) \%>\%
  mutate_v(measure = gsub("\\\\.", " ", tolower(names(iris)[-5]))) \%>\%
  print() -> iris_pca
# PCA biplot with species ranges
iris_pca \%>\%
  ggbiplot(aes(color = species)) +
  theme_bw() +
  scale_color_brewer(type = "qual", palette = 2) +
  # 
  geom_u_lineranges(fun.data = mean_sdl, size = .75) +
  geom_u_point(alpha = .5) +
  geom_v_vector(color = "#444444") +
  ggtitle(
    "Row-principal PCA biplot of Anderson iris data",
    "Ranges 2 sample standard deviations from centroids"
  )
# PCA biplot with species confidence intervals
iris_pca \%>\%
  ggbiplot(aes(color = species)) +
  theme_bw() +
  scale_color_brewer(type = "qual", palette = 2) +
  # 99\% nonparametric confidence intervals
  geom_u_lineranges(
    fun.data = mean_cl_boot, fun.args = list(conf.int = .99),
    size = .75
  ) +
  geom_u_point(alpha = .5) +
  geom_v_vector(color = "#444444") +
  ggtitle(
    "Row-principal PCA biplot of Anderson iris data",
    "99\% confidence intervals based on nonparametric bootstrap sampling"
  )
}
\seealso{
\link[ggplot2:draw_key]{ggplot2::draw_key}
}
