% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbiplot.r
\name{ggbiplot}
\alias{ggbiplot}
\title{Build a biplot object from ordination data wrapped as a bibble}
\usage{
ggbiplot(ordination = NULL, mapping = aes(x = 1, y = 2), ...)
}
\description{
\code{ggbiplot} produces a \strong{\link[ggplot2]{ggplot2}} object from a 
\code{"bbl"}-class ordination object. The baseline object is the default 
unadorned \code{\link[ggplot2]{ggplot}} object \code{p} with the following
differences:
\itemize{
  \item \code{p$mapping} is augmented with \code{.matrix = .matrix},
        which expects either \code{.matrix = "u"} or \code{.matrix = "v"}
        from the biplot layers.
  \item \code{p$coordinates} is set to \code{\link[ggplot2]{coord_fixed}}
        in order to faithfully render the geometry of an ordination.
  \item \code{p} is assigned the class \code{"ggbiplot"}. This serves no
        functional purpose currently.
}
Furthermore, the user may feed single integer values to the \code{x} and 
\code{y} aesthetics, which will be interpreted as the corresponding
coordinates in the ordination.
}
\section{Biplot layers}{

\code{\link{ggbiplot}} uses \code{\link[ggplot2]{fortify}} internally to 
produce a single data frame with a \code{.matrix} column distinguishing the 
subjects (\code{"u"}) and variables (\code{"v"}). The stat layers 
\code{stat_u} and \code{stat_v} simply filter the data frame to one of these 
two.

The geom layers \code{geom_u_*} and \code{geom_v_*} call the corresponding
stat in order to render plot elements for the corresponding matrix \eqn{U} or
\eqn{V}. \code{geom_biplot_*} selects a default matrix based on common
practice, e.g. \eqn{U} for points and \eqn{V} for arrows.
}

