% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbiplot.r
\name{ggbiplot}
\alias{ggbiplot}
\title{Build a biplot object from ordination data wrapped as a bibble}
\usage{
ggbiplot(ordination = NULL, mapping = aes(x = 1, y = 2), ...)
}
\arguments{
\item{ordination}{A \link{bibble}, i.e. an ordination object of class
\code{"bbl"}.}

\item{mapping}{List of default aesthetic mappings to use for the biplot. The 
default assigns the first two coordinates to the aesthetics \code{x} and 
\code{y}. Other assignments must be supplied in each layer added to the 
plot.}

\item{...}{Additional arguments passed to \code{\link[ggplot2]{ggplot}}.}
}
\description{
\code{ggbiplot} produces a \strong{\link[ggplot2]{ggplot2}} object from a 
\code{"bbl"}-class ordination object. The baseline object is the default 
unadorned \code{\link[ggplot2]{ggplot}} object \code{p} with the following
differences:
\itemize{
  \item \code{p$mapping} is augmented with \code{.matrix = .matrix},
        which expects either \code{.matrix = "u"} or \code{.matrix = "v"}
        from the biplot layers.
  \item \code{p$coordinates} is set to \code{\link[ggplot2]{coord_fixed}}
        in order to faithfully render the geometry of an ordination.
  \item \code{p} is assigned the class \code{"ggbiplot"}. This serves no
        functional purpose currently.
}
Furthermore, the user may feed single integer values to the \code{x} and 
\code{y} aesthetics, which will be interpreted as the corresponding
coordinates in the ordination.
}
\section{Biplot layers}{

\code{\link{ggbiplot}} uses \code{\link[ggplot2]{fortify}} internally to 
produce a single data frame with a \code{.matrix} column distinguishing the 
subjects (\code{"u"}) and variables (\code{"v"}). The stat layers 
\code{stat_u} and \code{stat_v} simply filter the data frame to one of these 
two.

The geom layers \code{geom_u_*} and \code{geom_v_*} call the corresponding
stat in order to render plot elements for the corresponding matrix \eqn{U} or
\eqn{V}. \code{geom_biplot_*} selects a default matrix based on common
practice, e.g. \eqn{U} for points and \eqn{V} for arrows.
}

\examples{
# reproduce Exhibit 2.3 in Greenacre (2010)

data(bioenv)
# standardize two predictors
bioenv <- dplyr::mutate(
  bioenv,
  x = (Depth - mean(Depth)) / sd(Depth),
  y = (Pollution - mean(Pollution)) / sd(Pollution)
)

# linear regression setup
(m <- lm(data = bioenv, formula = d ~ x + y))
(b <- as_bibble(m))
(d <- fortify(b))

# basic linear regression biplot
gg <- ggbiplot(b, aes(x = x, y = y))
gg +
  geom_u_point() +
  geom_v_vector()
# same biplot using default matrices for each layer
gg +
  geom_biplot_point() +
  geom_biplot_vector()
# same biplot with isolines for sole axis
gg +
  geom_u_point() +
  geom_v_vector() +
  geom_v_isolines(ids = 1, by = 5)

# shade points by expected species prevalence and darken outliers
ggbiplot(d, aes(x = x, y = y, color = .fitted, alpha = .resid^2)) +
  theme_bw() +
  scale_color_distiller(type = "div", palette = 1) +
  scale_alpha_continuous(range = c(1/3, 1)) +
  geom_u_point() +
  geom_v_vector() +
  geom_v_isolines(ids = 1, by = 5)
}
