% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot-gg.r
\name{ggbiplot}
\alias{ggbiplot}
\alias{ord_aes}
\title{Biplots following the grammar of graphics}
\usage{
ggbiplot(ordination = NULL, mapping = aes(x = 1, y = 2),
  sec.axes = NULL, scale.factor = NULL, scale_u = NULL,
  scale_v = NULL, ...)

ord_aes(ordination, ...)
}
\arguments{
\item{ordination}{A \link{tbl_ord}.}

\item{mapping}{List of default aesthetic mappings to use for the biplot. The
default assigns the first two coordinates to the aesthetics \code{x} and \code{y}.
Other assignments must be supplied in each layer added to the plot.}

\item{sec.axes}{Matrix factor character to specify a secondary set of axes.}

\item{scale.factor}{Numeric value used to scale the secondary axes against
the primary axes; ignored if \code{sec.axes} is not specified.}

\item{scale_u, scale_v}{Either the character name of a numeric variable in
\code{get_*(ordination)} or a numeric vector of length
\code{nrow(get_*(ordination))}, used to scale the coordinates of \eqn{U} or
\eqn{V}, respectively.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} or to
\code{\link[ggplot2:aes]{ggplot2::aes()}}.}
}
\description{
Build a biplot visualization from ordination data wrapped as a
\code{tbl_ord} object.
}
\details{
\code{ggbiplot()} produces a \link[ggplot2:ggplot]{ggplot} object from a \code{tbl_ord}
object. The baseline object is the default unadorned \code{"ggplot"}-class object
\code{p} with the following differences from what \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} returns:
\itemize{
\item \code{p$mapping} is augmented with \code{.matrix = .matrix}, which expects either
\code{.matrix = "u"} or \code{.matrix = "v"} from the biplot.
\item \code{p$coordinates} is defaulted to \link[ggplot2:coord_fixed]{ggplot2::coord_fixed} in order to
faithfully render the geometry of an ordination.
\item \code{p} is assigned the class \code{"ggbiplot"} in addition to \code{"ggplot"}. This
serves no functional purpose currently.
}

Furthermore, the user may feed single integer values to the \code{x} and \code{y}
aesthetics, which will be interpreted as the corresponding coordinates in the
ordination.

\code{ord_aes()} is a convenience function that generates a full-rank set of
coordinate aesthetics \code{.coord1}, \code{.coord2}, etc. mapped to the
shared coordinates of the ordination object, along with any additional
aesthetics that are processed internally by \code{\link[ggplot2:aes]{ggplot2::aes()}}.
}
\section{Biplot layers}{

\code{\link[=ggbiplot]{ggbiplot()}} uses \code{\link[ggplot2:fortify]{ggplot2::fortify()}} internally to produce a single data
frame with a \code{.matrix} column distinguishing the subjects (\code{"u"}) and
variables (\code{"v"}). The stat layers \code{stat_u()} and \code{stat_v()} simply filter
the data frame to one of these two.

The geom layers \code{geom_u_*()} and \code{geom_v_*()} call the corresponding stat in
order to render plot elements for the corresponding matrix \eqn{U} or
\eqn{V}. \code{geom_biplot_*()} selects a default matrix based on common practice,
e.g. \eqn{U} for points and \eqn{V} for arrows.
}

