% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot-gg.r
\name{ggbiplot}
\alias{ggbiplot}
\alias{ord_aes}
\title{Biplots following the grammar of graphics}
\usage{
ggbiplot(ordination = NULL, mapping = aes(x = 1, y = 2),
  axis.percents = TRUE, sec.axes = NULL, scale.factor = NULL,
  scale_u = NULL, scale_v = NULL, ...)

ord_aes(ordination, ...)
}
\arguments{
\item{ordination}{A \code{[tbl_ord]}.}

\item{mapping}{List of default aesthetic mappings to use for the biplot. The
default assigns the first two coordinates to the aesthetics \code{x} and \code{y}.
Other assignments must be supplied in each layer added to the plot.}

\item{axis.percents}{Whether to concatenate default axis labels with inertia
percentages.}

\item{sec.axes}{Matrix factor character to specify a secondary set of axes.}

\item{scale.factor}{Numeric value used to scale the secondary axes against
the primary axes; ignored if \code{sec.axes} is not specified.}

\item{scale_u, scale_v}{Either the character name of a numeric variable in
\code{get_*(ordination)} or a numeric vector of length
\code{nrow(get_*(ordination))}, used to scale the coordinates of \eqn{U} or
\eqn{V}, respectively.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} or to
\code{\link[ggplot2:aes]{ggplot2::aes()}}.}
}
\description{
Build a biplot visualization from ordination data wrapped as a
\code{tbl_ord} object.
}
\details{
\code{ggbiplot()} produces a \link[ggplot2:ggplot]{ggplot} object from a \code{tbl_ord}
object \code{ordination}. The baseline object is the default unadorned
\code{"ggplot"}-class object \code{p} with the following differences from what
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} returns:
\itemize{
\item \code{p$mapping} is augmented with \code{.matrix = .matrix}, which expects either
\code{.matrix = "u"} or \code{.matrix = "v"} from the biplot.
\item \code{p$coordinates} is defaulted to \code{\link[ggplot2:coord_equal]{ggplot2::coord_equal()}} in order to
faithfully render the geometry of an ordination.
\item When \code{x} or \code{y} are mapped to coordinates of \code{ordination}, and if
\code{axis.percents} is \code{TRUE}, \code{p$labels$x} or \code{p$labels$y} are defaulted to the
coordinate names concatenated with the percentages of \link[=conference]{inertia}
captured by the coordinates.
\item \code{p} is assigned the class \code{"ggbiplot"} in addition to \code{"ggplot"}. This
serves no functional purpose currently.
}

Furthermore, the user may feed single integer values to the \code{x} and \code{y}
aesthetics, which will be interpreted as the corresponding coordinates in the
ordination.

\code{ord_aes()} is a convenience function that generates a full-rank set of
coordinate aesthetics \code{.coord1}, \code{.coord2}, etc. mapped to the
shared coordinates of the ordination object, along with any additional
aesthetics that are processed internally by \code{\link[ggplot2:aes]{ggplot2::aes()}}.
}
\section{Biplot layers}{

\code{\link[=ggbiplot]{ggbiplot()}} uses \code{\link[ggplot2:fortify]{ggplot2::fortify()}} internally to produce a single data
frame with a \code{.matrix} column distinguishing the subjects (\code{"u"}) and
variables (\code{"v"}). The stat layers \code{stat_u()} and \code{stat_v()} simply filter
the data frame to one of these two.

The geom layers \code{geom_u_*()} and \code{geom_v_*()} call the corresponding stat in
order to render plot elements for the corresponding matrix \eqn{U} or
\eqn{V}. \code{geom_biplot_*()} selects a default matrix based on common practice,
e.g. \eqn{U} for points and \eqn{V} for arrows.
}

\examples{
# Regression analysis of Motor Trend design and performance data
mtcars \%>\%
  scale(scale = FALSE) \%>\%
  as.data.frame() \%>\%
  lm(formula = mpg ~ wt + cyl) \%>\%
  as_tbl_ord() \%>\%
  augment() \%>\%
  mutate_u(influence = .wt.res^2) \%>\%
  print() -> mtcars_lm
mtcars_lm \%>\%
  ggbiplot(aes(x = wt, y = cyl)) +
  geom_u_point(aes(color = influence)) +
  geom_v_vector() +
  # weight isolines
  geom_v_isolines(ids = 1, by = 5) +
  ggtitle(
    "Weight isolines with data colored by importance",
    "Regressing mpg onto weight and number of cylinders"
  )
# Unscaled PCA and two-scaled biplot of Anderson iris data
iris[, -5] \%>\%
  princomp() \%>\%
  as_tbl_ord() \%>\%
  confer_inertia(1) \%>\%
  mutate_u(species = iris$Species) \%>\%
  mutate_v(measure = gsub("\\\\.", " ", tolower(names(iris)[-5]))) \%>\%
  print() -> iris_pca
iris_pca \%>\%
  ggbiplot(aes(color = species), sec.axes = "v", scale.factor = 3) +
  theme_bw() +
  scale_color_brewer(type = "qual", palette = 2) +
  geom_u_point() +
  geom_v_vector(color = "#444444") +
  geom_v_text_radiate(aes(label = measure), color = "#444444") +
  ggtitle(
    "Row-principal unscaled PCA biplot of Anderson iris data",
    "Variable loadings scaled to secondary axes"
  )
# Logistic SVD of Sanderson finches data
finches \%>\% t() \%>\%
  logisticSVD() \%>\%
  as_tbl_ord() \%>\%
  augment() \%>\%
  print() -> finches_lsvd
finches_lsvd \%>\%
  ggbiplot(
    aes(x = LSC1, y = LSC2, label = .name),
    sec.axes = "v", scale.factor = 100
  ) +
  geom_u_vector(alpha = .5, color = "darkred") +
  geom_u_text_radiate(size = 3, color = "darkred") +
  geom_v_label(size = 3, alpha = .5, color = "royalblue3") +
  ggtitle(
    "Logistic SVD of the Galapagos island finches",
    "Islands (finches) scaled to the primary (secondary) axes"
  )
# Logistic PCA of Sanderson finches data
finches \%>\% t() \%>\%
  logisticPCA() \%>\%
  as_tbl_ord() \%>\%
  augment() \%>\%
  print() -> finches_lpca
finches_lpca \%>\%
  ggbiplot(
    aes(x = LPC1, y = LPC2, label = .name),
    sec.axes = "v", scale.factor = 50
  ) +
  geom_u_vector(alpha = .5, color = "darkred") +
  geom_u_text_radiate(size = 3, color = "darkred") +
  geom_v_label(size = 3, alpha = .5, color = "royalblue3") +
  ggtitle(
    "Logistic PCA of the Galapagos island finches",
    "Islands (finches) scaled to the primary (secondary) axes"
  )
# Convex logistic PCA of Sanderson finches data
finches \%>\% t() \%>\%
  # omit finches present on every island
  {.[, ! apply(., 2, all), drop = FALSE]} \%>\%
  convexLogisticPCA() \%>\%
  as_tbl_ord() \%>\%
  augment() \%>\%
  print() -> finches_clpca
finches_clpca \%>\%
  ggbiplot(
    aes(x = LPC1, y = LPC2, label = .name),
    sec.axes = "v", scale.factor = 50
  ) +
  geom_u_vector(alpha = .5, color = "darkred") +
  geom_u_text_radiate(size = 3, color = "darkred") +
  geom_v_label(size = 3, alpha = .5, color = "royalblue3") +
  ggtitle(
    "Convex logistic PCA of the Galapagos island finches",
    "Islands (finches) scaled to the primary (secondary) axes"
  )
}
\seealso{
\code{\link[ggplot2:ggplot2]{ggplot2::ggplot2()}}
}
