% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{glass}
\alias{glass}
\title{Glass composition data}
\format{
A \link[tibble:tibble]{tibble}.
}
\source{
Baxter & Freestone (2006).
}
\usage{
data(glass)
}
\description{
Sites, artifacts, and amounts of several components of the sand
from archaeological sites in Israel.
}
\references{
Baxter MJ & Freestone IC (2006) "Log-Ratio Compositional Data Analysis in
Archaeometry". \emph{Archaeometry}, \strong{48}(3): 511--531.
\doi{10.1111/j.1475-4754.2006.00270.x}
}
\keyword{datasets}
