% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.r
\name{lda-ord}
\alias{lda-ord}
\alias{lda_ord}
\title{Augmented implementation of linear discriminant analysis}
\usage{
lda_ord(x, ...)
}
\arguments{
\item{x}{
    (required if no formula is given as the principal argument.)
    a matrix or data frame or Matrix containing the explanatory variables.
  }

\item{...}{
    arguments passed to or from other methods.
}

\item{axes.scale}{Character string indicating how to left-transform the
\code{scaling} value when rendering biplots using \code{\link[=ggbiplot]{ggbiplot()}}.}
}
\description{
This function replicates \code{\link[MASS:lda]{MASS::lda()}} with options to retain
elements useful to the \link{tbl_ord} class and biplot calculations.
}
