% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-lra.r
\name{lra-ord}
\alias{lra-ord}
\alias{lra}
\title{Log-ratio analysis}
\usage{
lra(x, compositional = FALSE, weighted = TRUE)
}
\arguments{
\item{x}{A numeric matrix or rectangular data set.}

\item{compositional}{Logical; whether to normalize rows of \code{x} to sum to 1.}

\item{weighted}{Logical; whether to weight rows and columns by their sums.}
}
\value{
Given an \eqn{n * p} data matrix and setting \eqn{r=min(n,p)},
\code{lra()} returns a list of class \code{"lra"} containing three elements:
\itemize{
\item{sv}{The \eqn{r-1} singular values}
\item{row.coords}{The \eqn{n * (r-1)} matrix
of row standard coordinates.}
\item{column.coords}{The \eqn{p * (r-1)} matrix
of column standard coordinates.}
}
}
\description{
Represent log-ratios between variables based on their values on
a population of cases.
}
\details{
Log-ratio analysis (LRA) is based on a double-centering of log-transformed
data, usually weighted by row and column totals. The technique is suitable
for positive-valued variables on a common scale (e.g. percentages). The
distances between variables' coordinates (in the full-dimensional space) are
their pairwise log-ratios. The distances between cases' coordinates are
called their \emph{log-ratio distances}, and the total variance is the weighted
sum of their squares.

LRA is not implemented in standard R distributions but is a useful member of
the ordination toolkit. This is a minimal implementation following
Greenacre's (2010) exposition in Chapter 7.
}
\examples{
# U.S. 1973 violent crime arrests
head(USArrests)
# row and column subsets
state_examples <- c("Hawaii", "Mississippi", "North Dakota")
arrests <- c(1L, 2L, 4L)
# pairwise log-ratios of violent crime arrests for two states
arrest_pairs <- combn(arrests, 2L)
arrest_ratios <-
  USArrests[, arrest_pairs[1L, ]] / USArrests[, arrest_pairs[2L, ]]
colnames(arrest_ratios) <- paste(
  colnames(USArrests)[arrest_pairs[1L, ]], "/",
  colnames(USArrests)[arrest_pairs[2L, ]], sep = ""
)
arrest_logratios <- log(arrest_ratios)
arrest_logratios[state_examples, ]
# non-compositional log-ratio analysis
(arrests_lra <- lra(USArrests[, c(1, 2, 4)]))
(arrests_lra_ord <- augment_ord(as_tbl_ord(arrests_lra)))
# state abbreviations
state_abbr <- data.frame(state = state.name, abbr = state.abb)
# row-principal biplot
arrests_lra_ord \%>\%
  left_join_rows(state_abbr, by = c(".name" = "state")) \%>\%
  confer_inertia("rows") \%>\%
  ggbiplot(aes(color = .matrix), sec.axes = "cols", scale.factor = 1/20) +
  scale_color_manual(values = c("tomato4", "turquoise4")) +
  theme_bw() +
  geom_rows_text(aes(label = abbr), size = 3, alpha = .75) +
  geom_cols_polygon(fill = NA, linetype = "dashed") +
  geom_cols_text(aes(label = .name), fontface = "bold") +
  ggtitle(
    "Non-compositional LRA of violent crime arrest rates",
    "United States, 1973"
  ) +
  expand_limits(x = c(-.35)) +
  guides(color = "none")
# compositional log-ratio analysis
(arrests_lra <- lra(USArrests[, c(1, 2, 4)], compositional = TRUE))
(arrests_lra_ord <- augment_ord(as_tbl_ord(arrests_lra)))
# row-principal biplot
arrests_lra_ord \%>\%
  left_join_rows(state_abbr, by = c(".name" = "state")) \%>\%
  confer_inertia("rows") \%>\%
  ggbiplot(aes(color = .matrix), sec.axes = "cols", scale.factor = 1/20) +
  scale_color_manual(values = c("tomato4", "turquoise4")) +
  theme_bw() +
  geom_rows_text(aes(label = abbr), size = 3, alpha = .75) +
  geom_cols_polygon(fill = NA, linetype = "dashed") +
  geom_cols_text(aes(label = .name), fontface = "bold") +
  ggtitle(
    "Compositional LRA of violent crime arrest rates",
    "United States, 1973"
  ) +
  expand_limits(x = c(-.4)) +
  guides(color = "none")
}
\references{
Greenacre MJ (2010) \emph{Biplots in Practice}. Fundacion BBVA, ISBN:
978-84-923846.
\url{https://www.fbbva.es/microsite/multivariate-statistics/biplots.html}
}
