% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-lra.r
\name{lra-ord}
\alias{lra-ord}
\alias{lra}
\title{Log-ratio analysis}
\usage{
lra(x, compositional = FALSE, weighted = TRUE)
}
\arguments{
\item{x}{A numeric matrix or rectangular data set.}

\item{compositional}{Logical; whether to normalize rows of \code{x} to sum to 1.}

\item{weighted}{Logical; whether to weight rows and columns by their sums.}
}
\value{
Given an \eqn{n * p} data matrix and setting \eqn{r=min(n,p)},
\code{lra()} returns a list of class \code{"lra"} containing three elements:
\itemize{
\item{sv}{The \eqn{r-1} singular values}
\item{row.coords}{The \eqn{n * (r-1)} matrix
of row standard coordinates.}
\item{column.coords}{The \eqn{p * (r-1)} matrix
of column standard coordinates.}
}
}
\description{
Represent log-ratios between variables based on their values on
a population of cases.
}
\details{
Log-ratio analysis (LRA) is based on a double-centering of log-transformed
data, usually weighted by row and column totals. The technique is suitable
for positive-valued variables on a common scale (e.g. percentages). The
distances between variables' coordinates (in the full-dimensional space) are
their pairwise log-ratios. The distances between cases' coordinates are
called their \emph{log-ratio distances}, and the total variance is the weighted
sum of their squares.

LRA is not implemented in standard R distributions but is a useful member of
the ordination toolkit. This is a minimal implementation following
Greenacre's (2010) exposition in Chapter 7.
}
\examples{
# Log-ratio analysis of shared ingredients in Ruhlman's recipe ratios
data(ratios)
ratios \%>\%
  dplyr::select(chapter, recipe, flour, egg, butter) \%>\%
  dplyr::filter(flour > 0 & egg > 0 & butter > 0) \%>\%
  dplyr::group_by(flour, egg, butter) \%>\%
  dplyr::summarize(chapter = unique(chapter),
                   recipes = stringr::str_c(recipe, collapse = " or ")) \%>\%
  dplyr::ungroup() \%>\%
  print() -> sub_ratios
sub_ratios \%>\%
  dplyr::select(-chapter, -recipes) \%>\%
  lra(compositional = TRUE) \%>\%
  as_tbl_ord() \%>\%
  augment() \%>\%
  bind_cols_rows(dplyr::select(sub_ratios, chapter, recipes)) \%>\%
  print() -> lra_ratios
lra_ratios \%>\%
  confer_inertia("rows") \%>\%
  ggbiplot(sec.axes = "cols", scale.factor = .05) +
  geom_rows_text(aes(label = recipes, color = chapter)) +
  geom_cols_vector() +
  geom_cols_text(aes(label = .name), hjust = "outward", vjust = "outward") +
  expand_limits(x = c(-.2, .2)) +
  ggtitle("Log-ratio biplot of flour-butter-egg ratios in baking recipes") +
  expand_limits(x = c(-.35, .35))
# state abbreviations
state <- data.frame(
  .name = state.name,
  .abb = state.abb
)
# Log-ratio analysis of 1973 violent crime arrests in the United States
USArrests \%>\%
  subset(select = -UrbanPop) \%>\%
  lra() \%>\%
  print() -> arrests_lra
arrests_lra \%>\%
  as_tbl_ord() \%>\%
  augment() \%>\%
  left_join_rows(state, by = ".name") \%>\%
  print() -> arrests_lra
# Adapt Exhibit 7.1 in Greenacre (2010)
arrests_lra \%>\%
  confer_inertia(1) \%>\%
  ggbiplot(sec.axes = "cols", scale.factor = 1/20) +
  theme_bw() +
  geom_rows_text(
    aes(label = .abb),
    size = 3, color = "darkgreen", alpha = .5
  ) +
  geom_cols_polygon(fill = NA, linetype = "dashed", color = "brown4") +
  geom_cols_text(
    aes(label = .name),
    color = "brown4", fontface = "bold"
  ) +
  ggtitle(
    "Log-ratio analysis of violent crime arrest rates",
    "United States, 1973"
  ) +
  guides(color = FALSE, size = FALSE)
# Compare PCA to LRA on the Freestone primary class composition data
# following Baxter & Freestone (2006)
# (do not exclude compositional outliers)
data(glass)
levantine_glass <- glass \%>\%
  dplyr::filter(Site != "Banias") \%>\%
  dplyr::mutate(Type = dplyr::case_when(
    Site == "Dor" ~ "Levantine I",
    Site == "Apollonia" ~ "Levantine I",
    Site == "Bet Eli'ezer" ~ "Levantine II"
  ))
# scaled principal components analysis
levantine_glass \%>\%
  dplyr::select(SiO2, Al2O3, CaO, FeO, MgO) \%>\%
  princomp(cor = TRUE) \%>\%
  as_tbl_ord() \%>\%
  bind_cols_rows(dplyr::select(levantine_glass, Site, Type)) \%>\%
  print() -> pca_glass
ggbiplot(pca_glass) +
  geom_rows_point(aes(shape = Site, color = Type))
# completely compositional log-ratio analysis
levantine_glass \%>\%
  dplyr::select(SiO2, Al2O3, CaO, FeO, MgO) \%>\%
  lra(compositional = TRUE) \%>\%
  as_tbl_ord() \%>\%
  confer_inertia("rows") \%>\%
  bind_cols_rows(dplyr::select(levantine_glass, Site, Type)) \%>\%
  print() -> lra_glass
ggbiplot(lra_glass, sec.axes = "cols", scale.factor = .05) +
  geom_rows_point(aes(shape = Site, color = Type)) +
  geom_cols_vector() +
  geom_cols_text(aes(label = .name), hjust = "outward", vjust = "outward") +
  expand_limits(x = c(-.2, .2))
# completely compositional log-ratio analysis with FeO and MgO excluded
levantine_glass \%>\%
  dplyr::select(SiO2, Al2O3, CaO) \%>\%
  lra(compositional = TRUE) \%>\%
  as_tbl_ord() \%>\%
  confer_inertia("rows") \%>\%
  bind_cols_rows(dplyr::select(levantine_glass, Site, Type)) \%>\%
  print() -> lra_glass
ggbiplot(lra_glass, sec.axes = "cols", scale.factor = .05) +
  geom_rows_point(aes(shape = Site, color = Type)) +
  geom_cols_vector() +
  geom_cols_text(aes(label = .name), hjust = "outward", vjust = "outward") +
  expand_limits(x = c(NA, .2))
}
\references{
Greenacre MJ (2010) \emph{Biplots in Practice}. Fundacion BBVA, ISBN:
978-84-923846.
\url{https://www.fbbva.es/microsite/multivariate-statistics/biplots.html}
}
