% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masks.r
\name{masks}
\alias{masks}
\alias{cmdscale}
\alias{logisticSVD}
\alias{logisticPCA}
\alias{convexLogisticPCA}
\title{Masks of external ordination objects}
\usage{
cmdscale(d, k = 2, add = FALSE)

logisticSVD(x, k = 2, quiet = TRUE, max_iters = 1000,
  conv_criteria = 1e-05, random_start = FALSE, ...,
  partial_decomp = TRUE, main_effects = TRUE)

logisticPCA(x, k = 2, m = 4, quiet = TRUE, partial_decomp = FALSE,
  max_iters = 1000, conv_criteria = 1e-05, random_start = FALSE, ...,
  main_effects = TRUE)

convexLogisticPCA(x, k = 2, m = 4, quiet = TRUE,
  partial_decomp = FALSE, max_iters = 1000, conv_criteria = 1e-06,
  random_start = FALSE, ..., main_effects = TRUE, ss_factor = 4)
}
\arguments{
\item{d}{a distance structure such as that returned by \code{dist}
    or a full symmetric matrix containing the dissimilarities.}

\item{k}{the maximum dimension of the space which the data are to be
    represented in; must be in \eqn{\{1, 2, \ldots, n-1\}}{{1, 2, \\dots, n-1}}.}

\item{add}{logical indicating if an additive constant \eqn{c*} should
    be computed, and added to the non-diagonal dissimilarities such that
    the modified dissimilarities are Euclidean.}

\item{x}{matrix with all binary entries}

\item{quiet}{logical; whether the calculation should give feedback}

\item{max_iters}{number of maximum iterations}

\item{conv_criteria}{convergence criteria. The difference between average deviance
in successive iterations}

\item{random_start}{logical; whether to randomly inititalize the parameters. If \code{FALSE},
function will use an eigen-decomposition as starting value}

\item{...}{Additional parameters passed to original functions.}

\item{partial_decomp}{logical; if \code{TRUE}, the function uses the rARPACK package
to more quickly calculate the eigen-decomposition. This is usually faster than standard
eigen-decomponsition when \code{ncol(x) > 100} and \code{k} is small}

\item{main_effects}{logical; whether to include main effects in the model}

\item{m}{value to approximate the saturated model. If \code{m = 0}, m is solved for}

\item{ss_factor}{step size multiplier. Amount by which to multiply the step size. Quadratic
convergence rate can be proven for \code{ss_factor = 1}, but I have found higher values
sometimes work better. The default is \code{ss_factor = 4}.
If it is not converging, try \code{ss_factor = 1}.}
}
\description{
Some popular ordination functions in R return objects without a
specific class (e.g. \code{\link[stats:cmdscale]{stats::cmdscale()}}). Some others return output that
is not self-contained, for instance missing annotation from the original
dataset (e.g. \code{\link[logisticPCA:logisticPCA]{logisticPCA::logisticPCA()}}). The functions below wrap the
originals with the minimal overhead necessary for complete functionality
with the \code{"[tbl_ord]"} class.
}
