% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ca.r
\name{methods-ca}
\alias{methods-ca}
\alias{as_tbl_ord.ca}
\alias{reconstruct.ca}
\alias{recover_u.ca}
\alias{recover_v.ca}
\alias{recover_inertia.ca}
\alias{recover_conference.ca}
\alias{recover_coord.ca}
\alias{augmentation_u.ca}
\alias{augmentation_v.ca}
\alias{augmentation_coord.ca}
\title{Functionality for correspondence analysis ('ca') objects}
\usage{
\method{as_tbl_ord}{ca}(x)

\method{reconstruct}{ca}(x)

\method{recover_u}{ca}(x)

\method{recover_v}{ca}(x)

\method{recover_inertia}{ca}(x)

\method{recover_conference}{ca}(x)

\method{recover_coord}{ca}(x)

\method{augmentation_u}{ca}(x)

\method{augmentation_v}{ca}(x)

\method{augmentation_coord}{ca}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to, objects of class 
\code{"ca"} from the \strong{ca} package.
}
\examples{

data(smoke, package = "ca")

(m <- ca::ca(smoke))
(b <- as_tbl_ord(m))
(d <- fortify(b))

plot(m, map = "rowprincipal", col = c("green", "brown"))

# confer singular values to reproduce the row-principal biplot
b <- confer_inertia(b, "rowprincipal")
ggbiplot(b, aes(label = .name)) +
  theme_bw() +
  geom_u_point(color = "green") +
  geom_u_text_repel(color = "blue") +
  geom_v_point(color = "brown", shape = 17) +
  geom_v_text_repel(color = "brown")
}
