% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ca-ca.r
\name{methods-ca}
\alias{methods-ca}
\alias{as_tbl_ord.ca}
\title{Functionality for correspondence analysis ('ca') objects}
\usage{
\method{as_tbl_ord}{ca}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"ca"} from the \strong{\link[ca:ca]{ca}} package.
}
\examples{
# table of hair and eye color data collapsed by sex
class(HairEyeColor)
haireye <- as.data.frame(rowSums(HairEyeColor, dims = 2L))
print(haireye)
# use correspondence analysis to construct row and column profiles
haireye \%>\%
  ca::ca() \%>\%
  as_tbl_ord() \%>\%
  print() -> haireye_ca
# recover row and column profiles
get_rows(haireye_ca)
get_cols(haireye_ca)
# augment profiles with names, masses, distances, and inertias
(haireye_ca <- augment_ord(haireye_ca))
}
\seealso{
Other methods for singular value decomposition-based techniques: 
\code{\link{methods-cancor}},
\code{\link{methods-lda}},
\code{\link{methods-lra}},
\code{\link{methods-mjca}},
\code{\link{methods-prcomp}},
\code{\link{methods-princomp}},
\code{\link{methods-svd}}
}
\concept{methods for singular value decomposition-based techniques}
