% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-cancor.r, R/methods-stats-cancor.r
\name{methods-cancor}
\alias{methods-cancor}
\alias{as_tbl_ord.cancor_ord}
\alias{recover_rows.cancor_ord}
\alias{recover_cols.cancor_ord}
\alias{recover_inertia.cancor_ord}
\alias{recover_coord.cancor_ord}
\alias{recover_conference.cancor_ord}
\alias{augmentation_rows.cancor_ord}
\alias{augmentation_cols.cancor_ord}
\alias{augmentation_coord.cancor_ord}
\title{Functionality for canonical correlations}
\usage{
\method{as_tbl_ord}{cancor_ord}(x)

\method{recover_rows}{cancor_ord}(x)

\method{recover_cols}{cancor_ord}(x)

\method{recover_inertia}{cancor_ord}(x)

\method{recover_coord}{cancor_ord}(x)

\method{recover_conference}{cancor_ord}(x)

\method{augmentation_rows}{cancor_ord}(x)

\method{augmentation_cols}{cancor_ord}(x)

\method{augmentation_coord}{cancor_ord}(x)

\method{as_tbl_ord}{cancor_ord}(x)

\method{recover_rows}{cancor_ord}(x)

\method{recover_cols}{cancor_ord}(x)

\method{recover_inertia}{cancor_ord}(x)

\method{recover_coord}{cancor_ord}(x)

\method{recover_conference}{cancor_ord}(x)

\method{augmentation_rows}{cancor_ord}(x)

\method{augmentation_cols}{cancor_ord}(x)

\method{augmentation_coord}{cancor_ord}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"cancor_ord"}. This is a class introduced in this package
to identify objects returned by \code{\link[=cancor_ord]{cancor_ord()}}, which wraps
\code{\link[stats:cancor]{stats::cancor()}}.

These methods extract data from, and attribute new data to,
objects of class \code{"cancor_ord"}. This is a class introduced in this package
to identify objects returned by \code{\link[=cancor_ord]{cancor_ord()}}, which wraps
\code{\link[stats:cancor]{stats::cancor()}}.
}
\examples{
# data frame of life-cycle savings across countries
class(LifeCycleSavings)
head(LifeCycleSavings)
# canonical correlation analysis of age distributions and financial factors
savings_cancor <- cancor_ord(
  LifeCycleSavings[, c("pop15", "pop75")],
  LifeCycleSavings[, c("sr", "dpi", "ddpi")]
)
# wrap as a 'tbl_ord' object
(savings_cancor <- as_tbl_ord(savings_cancor))
# recover canonical weights
get_rows(savings_cancor)
get_cols(savings_cancor)
# augment canonical weights with row names and centers
(savings_cancor <- augment_ord(savings_cancor))
# row-standard biplot of structure correlations
savings_cancor \%>\%
  confer_inertia("cols") \%>\%
  ggbiplot(aes(label = .name, color = .matrix)) +
  theme_bw() +
  geom_origin() +
  geom_unit_circle(linetype = "dotted") +
  geom_rows_vector() +
  geom_rows_text_radiate() +
  geom_cols_point() +
  geom_cols_text_repel() +
  scale_color_brewer(limits = c("rows", "cols"), type = "qual") +
  expand_limits(x = c(-1, 1), y = c(-1, 1))
# column-standard biplot of structure correlations
savings_cancor \%>\%
  confer_inertia("rows") \%>\%
  ggbiplot(aes(label = .name, color = .matrix)) +
  theme_bw() +
  geom_origin() +
  geom_unit_circle(linetype = "dotted") +
  geom_cols_vector() +
  geom_cols_text_radiate() +
  geom_rows_point() +
  geom_rows_text_repel() +
  scale_color_brewer(limits = c("rows", "cols"), type = "qual") +
  expand_limits(x = c(-1, 1), y = c(-1, 1))
# symmetric biplot of structure correlations
savings_cancor \%>\%
  confer_inertia("symmetric") \%>\%
  ggbiplot(aes(label = .name, color = .matrix), elements = "active") +
  theme_bw() +
  geom_origin() +
  geom_unit_circle(linetype = "dotted") +
  geom_cols_vector() +
  geom_cols_text_radiate() +
  geom_rows_vector() +
  geom_rows_text_radiate() +
  scale_color_brewer(limits = c("rows", "cols"), type = "qual") +
  expand_limits(x = c(-1, 1), y = c(-1, 1))
# data frame of life-cycle savings across countries
class(LifeCycleSavings)
head(LifeCycleSavings)
# canonical correlation analysis of age distributions and financial factors
savings_cancor <- cancor_ord(
  LifeCycleSavings[, c("pop15", "pop75")],
  LifeCycleSavings[, c("sr", "dpi", "ddpi")]
)
# wrap as a 'tbl_ord' object
(savings_cancor <- as_tbl_ord(savings_cancor))
# recover canonical weights
get_rows(savings_cancor)
get_cols(savings_cancor)
# augment canonical weights with row names and centers
(savings_cancor <- augment_ord(savings_cancor))
# row-standard biplot of structure correlations
savings_cancor \%>\%
  confer_inertia("cols") \%>\%
  ggbiplot(aes(label = .name, color = .matrix)) +
  theme_bw() +
  geom_origin() +
  geom_unit_circle(linetype = "dotted") +
  geom_rows_vector() +
  geom_rows_text_radiate() +
  geom_cols_point() +
  geom_cols_text_repel() +
  scale_color_brewer(limits = c("rows", "cols"), type = "qual") +
  expand_limits(x = c(-1, 1), y = c(-1, 1))
# column-standard biplot of structure correlations
savings_cancor \%>\%
  confer_inertia("rows") \%>\%
  ggbiplot(aes(label = .name, color = .matrix)) +
  theme_bw() +
  geom_origin() +
  geom_unit_circle(linetype = "dotted") +
  geom_cols_vector() +
  geom_cols_text_radiate() +
  geom_rows_point() +
  geom_rows_text_repel() +
  scale_color_brewer(limits = c("rows", "cols"), type = "qual") +
  expand_limits(x = c(-1, 1), y = c(-1, 1))
# symmetric biplot of structure correlations
savings_cancor \%>\%
  confer_inertia("symmetric") \%>\%
  ggbiplot(aes(label = .name, color = .matrix), elements = "active") +
  theme_bw() +
  geom_origin() +
  geom_unit_circle(linetype = "dotted") +
  geom_cols_vector() +
  geom_cols_text_radiate() +
  geom_rows_vector() +
  geom_rows_text_radiate() +
  scale_color_brewer(limits = c("rows", "cols"), type = "qual") +
  expand_limits(x = c(-1, 1), y = c(-1, 1))
}
