% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-candisc-cancor.r
\name{methods-cancor}
\alias{methods-cancor}
\alias{as_tbl_ord.cancor}
\alias{recover_rows.cancor}
\alias{recover_cols.cancor}
\alias{recover_inertia.cancor}
\alias{recover_coord.cancor}
\alias{recover_conference.cancor}
\alias{augmentation_rows.cancor}
\alias{augmentation_cols.cancor}
\alias{augmentation_coord.cancor}
\alias{supplementation_rows.cancor}
\alias{supplementation_cols.cancor}
\title{Functionality for canonical correlation ('cancor') objects}
\usage{
\method{as_tbl_ord}{cancor}(x)

\method{recover_rows}{cancor}(x)

\method{recover_cols}{cancor}(x)

\method{recover_inertia}{cancor}(x)

\method{recover_coord}{cancor}(x)

\method{recover_conference}{cancor}(x)

\method{augmentation_rows}{cancor}(x)

\method{augmentation_cols}{cancor}(x)

\method{augmentation_coord}{cancor}(x)

\method{supplementation_rows}{cancor}(x)

\method{supplementation_cols}{cancor}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class 'cancor' from the \strong{\link[candisc:candisc-package]{candisc}}
package.
}
\details{
ter Braak (1990) recommends two families of biplots for the canonical
correspondence analysis of data matrices \eqn{X} and \eqn{Y}. For the first,
using structural correlations, either the interset correlations of \eqn{X}
(equivalent to principal coordinates) are plotted with the intraset
correlations of \eqn{Y} (standard coordinates) or vice-versa, so that their
product recovers the correlation matrix \eqn{XY'}. Where \code{cc} is the output
of \code{\link[candisc:cancor]{candisc::cancor()}} on centered and scaled data matrices, these are
obtained from \code{cc$structure}. For consistency with the canonical variate
scores \code{cc$scores} available as supplementary points, \strong{ordr} takes rows and
columns from the intraset correlations \code{cc$structure$X.xscores} and
\code{cc$structure$Y.yscores}, on which no intertia is conferred. ter Braak's
biplots can then be recovered by \link[=conference]{balancing} the inertia across
the two factors.
}
\examples{
# data frame of life-cycle savings across countries
class(LifeCycleSavings)
head(LifeCycleSavings)
# canonical correlation analysis of age distributions and financial factors
savings_cancor <- candisc::cancor(
  LifeCycleSavings[, c("pop15", "pop75")],
  LifeCycleSavings[, c("sr", "dpi", "ddpi")]
)
# wrap as a 'tbl_ord' object
(savings_cancor <- as_tbl_ord(savings_cancor))
# recover canonical weights
get_rows(savings_cancor, elements = "active")
get_cols(savings_cancor, elements = "active")
head(get_rows(savings_cancor))
head(get_cols(savings_cancor))
# augment canonical weights with row names and centers
(savings_cancor <- augment_ord(savings_cancor))
# row-standard biplot of structure correlations
savings_cancor \%>\%
  confer_inertia("cols") \%>\%
  ggbiplot(aes(label = .name, color = .matrix), elements = "active") +
  theme_bw() +
  geom_origin() +
  geom_unit_circle(linetype = "dotted") +
  geom_rows_vector() +
  geom_rows_text_radiate() +
  geom_cols_point() +
  geom_cols_text_repel() +
  scale_color_brewer(limits = c("rows", "cols"), type = "qual") +
  expand_limits(x = c(-1, 1), y = c(-1, 1))
# column-standard biplot of structure correlations
savings_cancor \%>\%
  confer_inertia("rows") \%>\%
  ggbiplot(aes(label = .name, color = .matrix), elements = "active") +
  theme_bw() +
  geom_origin() +
  geom_unit_circle(linetype = "dotted") +
  geom_cols_vector() +
  geom_cols_text_radiate() +
  geom_rows_point() +
  geom_rows_text_repel() +
  scale_color_brewer(limits = c("rows", "cols"), type = "qual") +
  expand_limits(x = c(-1, 1), y = c(-1, 1))
# symmetric biplot of structure correlations
savings_cancor \%>\%
  confer_inertia("symmetric") \%>\%
  ggbiplot(aes(label = .name, color = .matrix), elements = "active") +
  theme_bw() +
  geom_origin() +
  geom_unit_circle(linetype = "dotted") +
  geom_cols_vector() +
  geom_cols_text_radiate() +
  geom_rows_vector() +
  geom_rows_text_radiate() +
  scale_color_brewer(limits = c("rows", "cols"), type = "qual") +
  expand_limits(x = c(-1, 1), y = c(-1, 1))
}
\references{
ter Braak CJF (1990) "Interpreting canonical correlation analysis through
biplots of structure correlations and weights". \emph{Psychometrika} 55(3),
519--531. \doi{10.1007/BF02294765}
}
\seealso{
Other methods for singular value decomposition-based techniques: 
\code{\link{methods-ca}},
\code{\link{methods-lda}},
\code{\link{methods-lra}},
\code{\link{methods-mjca}},
\code{\link{methods-prcomp}},
\code{\link{methods-princomp}},
\code{\link{methods-svd}}
}
\concept{methods for singular value decomposition-based techniques}
