% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-stats-cancor.r
\name{methods-cancor}
\alias{methods-cancor}
\alias{as_tbl_ord.cancor_ord}
\alias{recover_rows.cancor_ord}
\alias{recover_cols.cancor_ord}
\alias{recover_inertia.cancor_ord}
\alias{recover_coord.cancor_ord}
\alias{recover_conference.cancor_ord}
\alias{augmentation_rows.cancor_ord}
\alias{augmentation_cols.cancor_ord}
\alias{augmentation_coord.cancor_ord}
\alias{supplementation_rows.cancor_ord}
\alias{supplementation_cols.cancor_ord}
\title{Functionality for canonical correlations}
\usage{
\method{as_tbl_ord}{cancor_ord}(x)

\method{recover_rows}{cancor_ord}(x)

\method{recover_cols}{cancor_ord}(x)

\method{recover_inertia}{cancor_ord}(x)

\method{recover_coord}{cancor_ord}(x)

\method{recover_conference}{cancor_ord}(x)

\method{augmentation_rows}{cancor_ord}(x)

\method{augmentation_cols}{cancor_ord}(x)

\method{augmentation_coord}{cancor_ord}(x)

\method{supplementation_rows}{cancor_ord}(x)

\method{supplementation_cols}{cancor_ord}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"cancor_ord"}. This is a class introduced in this package
to identify objects returned by \code{\link[=cancor_ord]{cancor_ord()}}, which wraps
\code{\link[stats:cancor]{stats::cancor()}}.
}
\details{
The canonical coefficients (loadings) are obtained directly from the
underlying singular value decomposition and constitute the active elements.
If scores are returned, then they are made available as supplementary
elements.

A biplot of the canonical coefficients can be interpreted as approximating
the \eqn{X}-\eqn{Y} covariance matrix, inversely weighted by the \eqn{X} and
\eqn{Y} variances. A biplot of the canonical scores of one matrix, available
as supplementary points if returned by \code{\link[=cancor_ord]{cancor_ord()}}, and the coefficients
of the other can be interpreted as .... Both biplots can have inertia
conferred to either or both factors. Greenacre (1984) describes both biplots,
though ter Braak (1990) recommends against the former.
}
\examples{
# data frame of life-cycle savings across countries
class(LifeCycleSavings)
head(LifeCycleSavings)
# canonical correlation analysis of age distributions and financial factors
savings_cancor <- cancor_ord(
  LifeCycleSavings[, c("pop15", "pop75")],
  LifeCycleSavings[, c("sr", "dpi", "ddpi")]
)
# wrap as a 'tbl_ord' object
(savings_cancor <- as_tbl_ord(savings_cancor))
# recover canonical weights
get_rows(savings_cancor)
get_cols(savings_cancor)
# augment canonical weights with row names and centers
(savings_cancor <- augment_ord(savings_cancor))
# row-standard biplot of structure correlations
savings_cancor \%>\%
  confer_inertia("cols") \%>\%
  ggbiplot(aes(label = .name, color = .matrix)) +
  theme_bw() +
  geom_origin() +
  geom_unit_circle(linetype = "dotted") +
  geom_rows_vector() +
  geom_rows_text_radiate() +
  geom_cols_point() +
  geom_cols_text_repel() +
  scale_color_brewer(limits = c("rows", "cols"), type = "qual") +
  expand_limits(x = c(-1, 1), y = c(-1, 1))
# column-standard biplot of structure correlations
savings_cancor \%>\%
  confer_inertia("rows") \%>\%
  ggbiplot(aes(label = .name, color = .matrix)) +
  theme_bw() +
  geom_origin() +
  geom_unit_circle(linetype = "dotted") +
  geom_cols_vector() +
  geom_cols_text_radiate() +
  geom_rows_point() +
  geom_rows_text_repel() +
  scale_color_brewer(limits = c("rows", "cols"), type = "qual") +
  expand_limits(x = c(-1, 1), y = c(-1, 1))
# symmetric biplot of structure correlations
savings_cancor \%>\%
  confer_inertia("symmetric") \%>\%
  ggbiplot(aes(label = .name, color = .matrix), elements = "active") +
  theme_bw() +
  geom_origin() +
  geom_unit_circle(linetype = "dotted") +
  geom_cols_vector() +
  geom_cols_text_radiate() +
  geom_rows_vector() +
  geom_rows_text_radiate() +
  scale_color_brewer(limits = c("rows", "cols"), type = "qual") +
  expand_limits(x = c(-1, 1), y = c(-1, 1))
}
\references{
Greenacre MJ (1984) \emph{Theory and applications of correspondence analysis}.
London: Academic Press, ISBN 0-12-299050-1.
\url{http://www.carme-n.org/?sec=books5}

ter Braak CJF (1990) "Interpreting canonical correlation analysis through
biplots of structure correlations and weights". \emph{Psychometrika} 55(3),
519--531. \doi{10.1007/BF02294765}
}
