% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-stats-cmds.r
\name{methods-cmds}
\alias{methods-cmds}
\alias{as_tbl_ord.cmds_ord}
\title{Functionality for classical multidimensional scaling objects}
\usage{
\method{as_tbl_ord}{cmds_ord}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"cmds_ord"}. This is a class introduced in this package
to identify objects returned by \code{\link[=cmdscale_ord]{cmdscale_ord()}}, which wraps
\code{\link[stats:cmdscale]{stats::cmdscale()}}.
}
\examples{
# 'dist' object (matrix of road distances) of large American cities
class(UScitiesD)
print(UScitiesD)
# use multidimensional scaling to infer artificial planar coordinates
UScitiesD \%>\%
  cmdscale_ord(k = 2) \%>\%
  as_tbl_ord() \%>\%
  print() -> usa_mds
# recover (equivalent) matrices of row and column artificial coordinates
get_rows(usa_mds)
get_cols(usa_mds)
# augment ordination with point names
(usa_mds <- augment_ord(usa_mds))
# reorient biplot to conventional compass
usa_mds \%>\%
  ggbiplot() +
  scale_x_reverse() +
  scale_y_reverse() +
  geom_cols_text(aes(label = .name), size = 3) +
  ggtitle("MDS biplot of distances between U.S. cities")
}
\seealso{
Other methods for eigen-decomposition-based techniques: 
\code{\link{methods-eigen}},
\code{\link{methods-factanal}}
}
\concept{methods for eigen-decomposition-based techniques}
