% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-stats-cmds.r
\name{methods-cmds}
\alias{methods-cmds}
\alias{as_tbl_ord.cmds_ord}
\alias{recover_rows.cmds_ord}
\alias{recover_cols.cmds_ord}
\alias{recover_inertia.cmds_ord}
\alias{recover_coord.cmds_ord}
\alias{recover_conference.cmds_ord}
\alias{augmentation_rows.cmds_ord}
\alias{augmentation_cols.cmds_ord}
\alias{augmentation_coord.cmds_ord}
\title{Functionality for classical multidimensional scaling objects}
\usage{
\method{as_tbl_ord}{cmds_ord}(x)

\method{recover_rows}{cmds_ord}(x)

\method{recover_cols}{cmds_ord}(x)

\method{recover_inertia}{cmds_ord}(x)

\method{recover_coord}{cmds_ord}(x)

\method{recover_conference}{cmds_ord}(x)

\method{augmentation_rows}{cmds_ord}(x)

\method{augmentation_cols}{cmds_ord}(x)

\method{augmentation_coord}{cmds_ord}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"cmds_ord"}. This is a class introduced in this package
to identify objects returned by \code{\link[=cmdscale_ord]{cmdscale_ord()}}, which wraps
\code{\link[stats:cmdscale]{stats::cmdscale()}}.
}
\examples{
# 'dist' object (matrix of road distances) of large American cities
class(UScitiesD)
print(UScitiesD)
# use multidimensional scaling to infer artificial planar coordinates
UScitiesD \%>\%
  cmdscale_ord(k = 2) \%>\%
  as_tbl_ord() \%>\%
  print() -> usa_mds
# summarize ordination
glance(usa_mds)
# recover (equivalent) matrices of row and column artificial coordinates
get_rows(usa_mds)
get_cols(usa_mds)
# augment ordination with point names
augment_ord(usa_mds)
# summarize artifical coordinates
tidy(usa_mds)
# scree plot of inertia
ggplot(tidy(usa_mds), aes(x = .name, y = .inertia)) +
  theme_bw() +
  geom_col() +
  labs(x = "", y = "Inertia")
# fortification automatically augments artificial coordinates
fortify(usa_mds)
# reorient biplot to conventional compass
usa_mds \%>\%
  ggbiplot() +
  scale_x_reverse() +
  scale_y_reverse() +
  geom_cols_text(aes(label = .name), size = 3) +
  ggtitle("MDS biplot of distances between U.S. cities")
}
