% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-cmds.r
\name{methods-cmds}
\alias{methods-cmds}
\alias{as_tbl_ord.cmds_ord}
\alias{recover_rows.cmds_ord}
\alias{recover_cols.cmds_ord}
\alias{recover_inertia.cmds_ord}
\alias{recover_coord.cmds_ord}
\alias{recover_conference.cmds_ord}
\alias{augmentation_rows.cmds_ord}
\alias{augmentation_cols.cmds_ord}
\alias{augmentation_coord.cmds_ord}
\title{Functionality for classical multidimensional scaling objects}
\usage{
\method{as_tbl_ord}{cmds_ord}(x)

\method{recover_rows}{cmds_ord}(x)

\method{recover_cols}{cmds_ord}(x)

\method{recover_inertia}{cmds_ord}(x)

\method{recover_coord}{cmds_ord}(x)

\method{recover_conference}{cmds_ord}(x)

\method{augmentation_rows}{cmds_ord}(x)

\method{augmentation_cols}{cmds_ord}(x)

\method{augmentation_coord}{cmds_ord}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"cmds_ord"}. This is a class introduced in this package
to identify objects returned by \code{\link[=cmdscale_ord]{cmdscale_ord()}}, which wraps
\code{\link[stats:cmdscale]{stats::cmdscale()}}.
}
\examples{
# Multidimensional scaling of country differences and regression of attributes
# Reproduce Exhibit 4.2 in Greenacre (2010)
data(country_differences)
country_differences \%>\%
  cmdscale_ord(k = 2) \%>\%
  as_tbl_ord() \%>\%
  print() -> differences_cmds
differences_plot <- differences_cmds \%>\%
  ggbiplot(aes(x = 1, y = 2, label = .name)) +
  geom_cols_text()
differences_plot
# Reproduce Exhibit 4.5 in Greenacre (2010)
data(country_attributes)
lm(country_attributes ~ get_rows(differences_cmds)) \%>\%
  as_tbl_ord() \%>\%
  print() -> attributes_fit
differences_plot +
  geom_cols_vector(data = attributes_fit) +
  geom_cols_text_radiate(data = attributes_fit, hjust = .3)
# Compare an MDS to a PCA
data(country_differences)
country_differences \%>\%
  cmdscale_ord(k = 2) \%>\%
  as_tbl_ord() \%>\%
  print() -> differences_cmds
country_differences \%>\%
  prcomp() \%>\%
  as_tbl_ord() \%>\%
  print() -> differences_pca
# negate PCA axis to match MDS axis
plot(gridExtra::arrangeGrob(grobs = list(
  ggbiplot(differences_cmds) + geom_rows_point(),
  ggbiplot(differences_pca) + scale_y_reverse() + geom_rows_point()
), ncol = 2))
}
