% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom-tidiers.r
\name{methods-eigen}
\alias{methods-eigen}
\alias{tidy.eigen}
\title{Tidy an 'eigen' object}
\usage{
\method{tidy}{eigen}(x, matrix = "vectors", ...)
}
\arguments{
\item{x}{An 'eigen' object, as returned by \code{\link[base:eigen]{base::eigen()}}.}

\item{matrix}{Character specifying which component of the eigendecomposition
should be tidied.
\itemize{
\item \code{"Lambda"}, \code{"vectors"}: returns information about the eigenvectors of
\code{x}
\item \code{"lambda"}, \code{"values"}: returns information about the eigenvalues
}}

\item{...}{Additional parameters passed to other methods; currently ignored.}
}
\description{
This \code{\link[broom:reexports]{broom::tidy()}} method tidies either of the two outputs of
\code{\link[base:eigen]{base::eigen()}}, depending on the \code{matrix} parameter, analogously to
\code{\link[broom:tidy_svd]{broom::tidy_svd()}}.
}
