% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-base-eigen.r
\name{methods-eigen}
\alias{methods-eigen}
\alias{as_tbl_ord.eigen_ord}
\alias{recover_rows.eigen_ord}
\alias{recover_cols.eigen_ord}
\alias{recover_inertia.eigen_ord}
\alias{recover_coord.eigen_ord}
\alias{recover_conference.eigen_ord}
\alias{augmentation_rows.eigen_ord}
\alias{augmentation_cols.eigen_ord}
\alias{augmentation_coord.eigen_ord}
\title{Functionality for eigendecompositions}
\usage{
\method{as_tbl_ord}{eigen_ord}(x)

\method{recover_rows}{eigen_ord}(x)

\method{recover_cols}{eigen_ord}(x)

\method{recover_inertia}{eigen_ord}(x)

\method{recover_coord}{eigen_ord}(x)

\method{recover_conference}{eigen_ord}(x)

\method{augmentation_rows}{eigen_ord}(x)

\method{augmentation_cols}{eigen_ord}(x)

\method{augmentation_coord}{eigen_ord}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"eigen_ord"} returned by \code{\link[=eigen_ord]{eigen_ord()}}.
}
\examples{
# subset QS data to rank variables
qs_ranks <- subset(
  qswur_usa,
  complete.cases(qswur_usa),
  select = 8:13
)
head(qs_ranks)
# eigendecomposition of Kendall correlation matrix
qs_ranks \%>\%
  cor(method = "kendall") \%>\%
  eigen_ord() \%>\%
  print() -> qs_eigen
# recover eigenvectors
get_rows(qs_eigen)
identical(get_cols(qs_eigen), get_rows(qs_eigen))
# wrap as a 'tbl_ord' and augment with dimension names
augment_ord(as_tbl_ord(qs_eigen))
# decomposition returns pure eigenvectors
get_conference(qs_eigen)
}
\seealso{
Other methods for eigen-decomposition-based techniques: 
\code{\link{methods-cmds}},
\code{\link{methods-factanal}}
}
\concept{methods for eigen-decomposition-based techniques}
