% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-eigen.r
\name{methods-eigen}
\alias{methods-eigen}
\alias{as_tbl_ord.eigen_ord}
\alias{reconstruct.eigen_ord}
\alias{recover_u.eigen_ord}
\alias{recover_v.eigen_ord}
\alias{recover_inertia.eigen_ord}
\alias{recover_coord.eigen_ord}
\alias{recover_conference.eigen_ord}
\alias{augmentation_u.eigen_ord}
\alias{augmentation_v.eigen_ord}
\alias{augmentation_coord.eigen_ord}
\title{Functionality for eigendecompositions}
\usage{
\method{as_tbl_ord}{eigen_ord}(x)

\method{reconstruct}{eigen_ord}(x)

\method{recover_u}{eigen_ord}(x)

\method{recover_v}{eigen_ord}(x)

\method{recover_inertia}{eigen_ord}(x)

\method{recover_coord}{eigen_ord}(x)

\method{recover_conference}{eigen_ord}(x)

\method{augmentation_u}{eigen_ord}(x)

\method{augmentation_v}{eigen_ord}(x)

\method{augmentation_coord}{eigen_ord}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"eigen_ord"} returned by \code{\link[=eigen_ord]{eigen_ord()}}.
}
\examples{
# Graph Laplacian eigenvectors for spectral partitioning
data(karate, package = "igraphdata")
igraph::plot.igraph(
  karate,
  vertex.label.family = "sans",
  vertex.label.color = "black",
  vertex.label.cex = .75
)
# first eigenvector (centrality)
karate_eigen_cent <- igraph::eigen_centrality(karate)
# eigendecomposition
karate \%>\%
  igraph::as_adjacency_matrix(sparse = FALSE) \%>\%
  eigen_ord() \%>\% as_tbl_ord() \%>\%
  mutate_u(
    faction = igraph::vertex_attr(karate, "Faction"),
    name = igraph::vertex_attr(karate, "name"),
    label = igraph::vertex_attr(karate, "label")
  ) \%>\%
  print() -> karate_eigen
# corroborate eigencentralities
tibble::tibble(
  igraph = karate_eigen_cent$vector,
  eigen = karate_eigen$vectors[, 1] / max(karate_eigen$vectors[, 1])
) \%>\%
  ggplot(aes(x = eigen, y = igraph)) +
  coord_equal() +
  geom_point()
# first and second eigenvectors for centrality and connectivity / partitioning
karate_eigen \%>\%
  confer_inertia(1) \%>\%
  ggbiplot(aes(x = 2, y = 1)) +
  geom_u_vector(aes(color = as.factor(faction))) +
  geom_u_text_radiate(aes(label = ifelse(grepl("H|A", label), label, NA))) +
  guides(color = FALSE) +
  labs(x = "algebraic connectivity", y = "eigencentrality") +
  ggtitle(
    "First (centrality) and second (connectivity) eigenvectors",
    "Colors distinguish factions around John A. and Mr. Hi"
  )
# second and third eigenvectors
karate_eigen \%>\%
  confer_inertia(1) \%>\%
  ggbiplot(aes(x = 2, y = 3)) +
  scale_x_continuous(expand = expand_scale(mult = .3)) +
  scale_y_continuous(expand = expand_scale(mult = .2)) +
  geom_u_vector(aes(color = EV1)) +
  geom_u_text_radiate(stat = "chull", aes(label = label)) +
  labs(x = "algebraic connectivity", y = "loading onto third eigenvector") +
  ggtitle(
    "Second (horizontal) and third (vertical) eigencentralities",
    "Color increases in value with eigencentrality, labels along convex hull"
  )
}
