% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-eigen.r
\name{methods-eigen}
\alias{methods-eigen}
\alias{as_tbl_ord.eigen}
\alias{reconstruct.eigen}
\alias{recover_u.eigen}
\alias{recover_v.eigen}
\alias{recover_inertia.eigen}
\alias{recover_coord.eigen}
\alias{recover_conference.eigen}
\alias{augmentation_u.eigen}
\alias{augmentation_v.eigen}
\alias{augmentation_coord.eigen}
\alias{negate_to.eigen}
\title{Functionality for matrix spectral decompositions}
\usage{
\method{as_tbl_ord}{eigen}(x)

\method{reconstruct}{eigen}(x)

\method{recover_u}{eigen}(x)

\method{recover_v}{eigen}(x)

\method{recover_inertia}{eigen}(x)

\method{recover_coord}{eigen}(x)

\method{recover_conference}{eigen}(x)

\method{augmentation_u}{eigen}(x)

\method{augmentation_v}{eigen}(x)

\method{augmentation_coord}{eigen}(x)

\method{negate_to}{eigen}(x, y, ..., .matrix)
}
\arguments{
\item{x}{An ordination object.}

\item{y}{A matrix or an ordination object.}

\item{...}{Additional elements of a custom \code{tbl_ord}.}

\item{.matrix}{A character string matching one of several indicators for one
or both matrices in a matrix decomposition used for ordination. The
standard values are \code{"u"}, \code{"v"}, and \code{"uv"}.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"eigen"}. This is a class introduced in this package to
identify objects returned by \code{\link[base:eigen]{base::eigen()}}, which is masked by a wrapper
that adds the class attribute.
}
