% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-kmeans.r
\name{methods-kmeans}
\alias{methods-kmeans}
\alias{as_tbl_ord.kmeans}
\alias{reconstruct.kmeans}
\alias{recover_u.kmeans}
\alias{recover_v.kmeans}
\alias{recover_coord.kmeans}
\alias{augmentation_u.kmeans}
\alias{augmentation_v.kmeans}
\alias{augmentation_coord.kmeans}
\title{Functionality for k-means clustering ('kmeans') objects}
\usage{
\method{as_tbl_ord}{kmeans}(x)

\method{reconstruct}{kmeans}(x)

\method{recover_u}{kmeans}(x)

\method{recover_v}{kmeans}(x)

\method{recover_coord}{kmeans}(x)

\method{augmentation_u}{kmeans}(x)

\method{augmentation_v}{kmeans}(x)

\method{augmentation_coord}{kmeans}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"kmeans"} as returned by \code{\link[stats:kmeans]{stats::kmeans()}}.
}
\examples{
# Interpreting (k-means) clusters as ordinations on the Motor Trends data
mtcars \%>\%
  scale() \%>\%
  kmeans(centers = 3) \%>\%
  as_tbl_ord() \%>\%
  augment() \%>\%
  print() -> mtcars_kmeans
mtcars_kmeans \%>\%
  augment_coord() \%>\%
  transform(.sdev = sqrt(.withinss / .size)) \%>\%
  print() -> mtcars_coord
# discriminate between clusters 1 and 2
mtcars_kmeans \%>\%
  ggbiplot(color = factor(.cluster)) +
  geom_jitter(stat = "u", aes(shape = .cluster), width = .2, height = .2) +
  geom_v_vector(aes(color = `3`)) +
  scale_color_distiller(type = "div", limits = c(-1, 1)) +
  geom_v_text_radiate(aes(label = .name)) +
  ggtitle(
    "Performance and design variable loadings onto clusters 1 and 2",
    "Color indicates loadings onto cluster 3"
  )
}
