% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-kmeans.r
\name{methods-kmeans}
\alias{methods-kmeans}
\alias{as_tbl_ord.kmeans}
\alias{reconstruct.kmeans}
\alias{recover_u.kmeans}
\alias{recover_v.kmeans}
\alias{recover_coord.kmeans}
\alias{augmentation_u.kmeans}
\alias{augmentation_v.kmeans}
\alias{augmentation_coord.kmeans}
\title{Functionality for k-means clustering ('kmeans') objects}
\usage{
\method{as_tbl_ord}{kmeans}(x)

\method{reconstruct}{kmeans}(x)

\method{recover_u}{kmeans}(x)

\method{recover_v}{kmeans}(x)

\method{recover_coord}{kmeans}(x)

\method{augmentation_u}{kmeans}(x)

\method{augmentation_v}{kmeans}(x)

\method{augmentation_coord}{kmeans}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to, objects of class 
\code{"kmeans"}.
}
\examples{

x <- scale(mtcars)
(m <- kmeans(x, 3))
plot(x, col = m$cluster)
points(m$centers, col = 1:3, pch = 8, cex = 2)

(b <- as_tbl_ord(m))
(d <- fortify(b))

# biplot with noise extending (in each artificial dimension) to the
# within-cluster (i.e. within-dimension) standard deviation
# (not great but suggestive)
coord_sdev <- dplyr::mutate(
  augment_coord(b),
  .sdev = sqrt(.withinss / .size)
)$.sdev
ggbiplot(b, aes(x = 1, y = 2, color = factor(.cluster))) +
  geom_u_point(size = 3, shape = "triangle") +
  #geom_u_point(position = position_jitter(coord_sdev[1], coord_sdev[2])) +
  geom_v_vector() +
  geom_v_text_radiate(aes(label = .name))
}
