% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lda.r
\name{methods-lda}
\alias{methods-lda}
\alias{as_tbl_ord.lda}
\alias{as_tbl_ord.lda_ord}
\alias{recover_u.lda}
\alias{recover_u.lda_ord}
\alias{recover_v.lda}
\alias{recover_v.lda_ord}
\alias{recover_inertia.lda}
\alias{recover_inertia.lda_ord}
\alias{recover_coord.lda}
\alias{recover_coord.lda_ord}
\alias{recover_conference.lda}
\alias{recover_conference.lda_ord}
\alias{augmentation_u.lda}
\alias{augmentation_u.lda_ord}
\alias{augmentation_v.lda}
\alias{augmentation_v.lda_ord}
\alias{augmentation_coord.lda}
\alias{augmentation_coord.lda_ord}
\alias{supplementation_u.lda}
\alias{supplementation_u.lda_ord}
\alias{augmentation_supplement_u.lda}
\alias{augmentation_supplement_u.lda_ord}
\title{Functionality for linear discriminant analysis ('lda') objects}
\usage{
\method{as_tbl_ord}{lda}(x)

\method{as_tbl_ord}{lda_ord}(x)

\method{recover_u}{lda}(x)

\method{recover_u}{lda_ord}(x)

\method{recover_v}{lda}(x)

\method{recover_v}{lda_ord}(x)

\method{recover_inertia}{lda}(x)

\method{recover_inertia}{lda_ord}(x)

\method{recover_coord}{lda}(x)

\method{recover_coord}{lda_ord}(x)

\method{recover_conference}{lda}(x)

\method{recover_conference}{lda_ord}(x)

\method{augmentation_u}{lda}(x)

\method{augmentation_u}{lda_ord}(x)

\method{augmentation_v}{lda}(x)

\method{augmentation_v}{lda_ord}(x)

\method{augmentation_coord}{lda}(x)

\method{augmentation_coord}{lda_ord}(x)

\method{supplementation_u}{lda}(x)

\method{supplementation_u}{lda_ord}(x)

\method{augmentation_supplement_u}{lda}(x)

\method{augmentation_supplement_u}{lda_ord}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"lda"} and \code{"lda_ord"} as returned by \code{\link[MASS:lda]{MASS::lda()}} and
\code{\link[=lda_ord]{lda_ord()}}.
}
\examples{
# Linear discriminant analysis of Reaven & Miller diabetes data
head(heplots::Diabetes)
(diabetes_lda <- MASS::lda(group ~ ., heplots::Diabetes))
lapply(supplementation_u(diabetes_lda), head)
as_tbl_ord(diabetes_lda) \%>\%
  augment() \%>\%
  supplement_u() \%>\%
  augment_supplement_u() \%>\%
  print() -> diabetes_lda
ggbiplot(diabetes_lda) +
  theme_bw() +
  geom_u_point(aes(color = .grouping, size = is.na(.supplement)), alpha = .5) +
  geom_v_axis(color = "#888888") +
  geom_v_text_radiate(aes(label = .name)) +
  scale_color_brewer(type = "qual", palette = 2) +
  ggtitle(
    "Column-standardized LDA biplot of Reaven & Miller diabetes data"
  )
}
