% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-mass-lda.r
\name{methods-lda}
\alias{methods-lda}
\alias{as_tbl_ord.lda}
\alias{as_tbl_ord.lda_ord}
\title{Functionality for linear discriminant analysis ('lda') objects}
\usage{
\method{as_tbl_ord}{lda}(x)

\method{as_tbl_ord}{lda_ord}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"lda"} and \code{"lda_ord"} as returned by \code{\link[MASS:lda]{MASS::lda()}} and
\code{\link[=lda_ord]{lda_ord()}}.
}
\details{
See \link{lda-ord} for details.
}
\examples{
# data frame of Anderson iris species measurements
class(iris)
head(iris)
# default (unstandardized discriminant) coefficients
lda_ord(iris[, 1:4], iris[, 5]) \%>\%
  as_tbl_ord() \%>\%
  print() -> iris_lda
# recover centroid coordinates and measurement discriminant coefficients
get_rows(iris_lda, elements = "active")
get_cols(iris_lda)
# augment ordination with centroid and measurement names
augment_ord(iris_lda)
}
\seealso{
Other methods for singular value decomposition-based techniques: 
\code{\link{methods-cancor}},
\code{\link{methods-ca}},
\code{\link{methods-lra}},
\code{\link{methods-mjca}},
\code{\link{methods-prcomp}},
\code{\link{methods-princomp}},
\code{\link{methods-svd}}
}
\concept{methods for singular value decomposition-based techniques}
