% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lda.r
\name{methods-lda}
\alias{methods-lda}
\alias{as_tbl_ord.lda}
\alias{as_tbl_ord.lda_ord}
\alias{recover_rows.lda}
\alias{recover_rows.lda_ord}
\alias{recover_cols.lda}
\alias{recover_cols.lda_ord}
\alias{recover_inertia.lda}
\alias{recover_inertia.lda_ord}
\alias{recover_coord.lda}
\alias{recover_coord.lda_ord}
\alias{recover_conference.lda}
\alias{recover_conference.lda_ord}
\alias{augmentation_rows.lda}
\alias{augmentation_rows.lda_ord}
\alias{augmentation_cols.lda}
\alias{augmentation_cols.lda_ord}
\alias{augmentation_coord.lda}
\alias{augmentation_coord.lda_ord}
\alias{supplementation_rows.lda}
\alias{supplementation_rows.lda_ord}
\title{Functionality for linear discriminant analysis ('lda') objects}
\usage{
\method{as_tbl_ord}{lda}(x)

\method{as_tbl_ord}{lda_ord}(x)

\method{recover_rows}{lda}(x)

\method{recover_rows}{lda_ord}(x)

\method{recover_cols}{lda}(x)

\method{recover_cols}{lda_ord}(x)

\method{recover_inertia}{lda}(x)

\method{recover_inertia}{lda_ord}(x)

\method{recover_coord}{lda}(x)

\method{recover_coord}{lda_ord}(x)

\method{recover_conference}{lda}(x)

\method{recover_conference}{lda_ord}(x)

\method{augmentation_rows}{lda}(x)

\method{augmentation_rows}{lda_ord}(x)

\method{augmentation_cols}{lda}(x)

\method{augmentation_cols}{lda_ord}(x)

\method{augmentation_coord}{lda}(x)

\method{augmentation_coord}{lda_ord}(x)

\method{supplementation_rows}{lda}(x)

\method{supplementation_rows}{lda_ord}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"lda"} and \code{"lda_ord"} as returned by \code{\link[MASS:lda]{MASS::lda()}} and
\code{\link[=lda_ord]{lda_ord()}}.
}
\details{
See \link{lda-ord} for details.
}
\examples{
# data frame of Anderson iris species measurements
class(iris)
head(iris)
# default (unstandardized discriminant) coefficients
lda_ord(iris[, 1:4], iris[, 5]) \%>\%
  as_tbl_ord() \%>\%
  print() -> iris_lda
# summarize ordination
glance(iris_lda)
# recover centroid coordinates and measurement discriminant coefficients
get_rows(iris_lda, .supplement = FALSE)
get_cols(iris_lda)
# augment ordination with centroid and measurement names
augment_ord(iris_lda)
# summarize linear discriminant axes
tidy(iris_lda)
# fortification of artificial coordinates yields proportion of variance
fortify(iris_lda, .matrix = "coord")
# scree plot of inertia
ggplot(iris_lda, .matrix = "coord", aes(x = .name, y = .prop_var)) +
  theme_bw() +
  scale_y_continuous(labels = scales::percent) +
  geom_col() +
  labs(x = "", y = "Proportion of inertia")
# fortification (without supplement)
fortify(iris_lda, .supplement = FALSE)
# unstandardized coefficient LDA biplot (centroids only)
iris_lda \%>\%
  augment_ord() \%>\%
  mutate_rows(species = .grouping) \%>\%
  ggbiplot(.supplement = FALSE) +
  theme_bw() +
  geom_rows_point(aes(color = .grouping), size = 3) +
  geom_cols_vector(color = "#888888") +
  geom_cols_text_radiate(aes(label = .name), size = 3) +
  scale_color_brewer(type = "qual", palette = 2) +
  ggtitle(
    "Row-standardized biplot of iris LDA",
    "Unstandardized measurement coefficients and species centroids"
  ) +
  expand_limits(y = c(-2, 4.25))
# biplot with supplementary points for observations
iris_lda \%>\%
  augment_ord() \%>\%
  mutate_rows(
    species = .grouping,
    discriminant = ifelse(! .supplement, "centroid", "case")
  ) \%>\%
  ggbiplot() +
  theme_bw() +
  geom_rows_point(aes(
    color = .grouping,
    size = discriminant, alpha = discriminant
  )) +
  geom_cols_vector(color = "#888888") +
  geom_cols_text_radiate(aes(label = .name), size = 3) +
  scale_color_brewer(type = "qual", palette = 2) +
  ggtitle(
    "Row-standardized biplot of iris LDA",
    "Unstandardized measurement coefficients, observations, & species centroids"
  ) +
  expand_limits(y = c(-3, 5))
}
