% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lpca.r
\name{methods-lpca}
\alias{methods-lpca}
\alias{as_tbl_ord.lsvd}
\alias{reconstruct.lsvd}
\alias{recover_u.lsvd}
\alias{recover_v.lsvd}
\alias{recover_coord.lsvd}
\alias{augmentation_u.lsvd}
\alias{augmentation_v.lsvd}
\alias{augmentation_coord.lsvd}
\alias{as_tbl_ord.lpca}
\alias{recover_u.lpca}
\alias{recover_v.lpca}
\alias{recover_coord.lpca}
\alias{augmentation_u.lpca}
\alias{augmentation_v.lpca}
\alias{augmentation_coord.lpca}
\alias{as_tbl_ord.clpca}
\alias{recover_u.clpca}
\alias{recover_v.clpca}
\alias{recover_coord.clpca}
\alias{augmentation_u.clpca}
\alias{augmentation_v.clpca}
\alias{augmentation_coord.clpca}
\title{Functionality for logistic PCA and logistic SVD objects}
\usage{
\method{as_tbl_ord}{lsvd}(x)

\method{reconstruct}{lsvd}(x)

\method{recover_u}{lsvd}(x)

\method{recover_v}{lsvd}(x)

\method{recover_coord}{lsvd}(x)

\method{augmentation_u}{lsvd}(x)

\method{augmentation_v}{lsvd}(x)

\method{augmentation_coord}{lsvd}(x)

\method{as_tbl_ord}{lpca}(x)

\method{recover_u}{lpca}(x)

\method{recover_v}{lpca}(x)

\method{recover_coord}{lpca}(x)

\method{augmentation_u}{lpca}(x)

\method{augmentation_v}{lpca}(x)

\method{augmentation_coord}{lpca}(x)

\method{as_tbl_ord}{clpca}(x)

\method{recover_u}{clpca}(x)

\method{recover_v}{clpca}(x)

\method{recover_coord}{clpca}(x)

\method{augmentation_u}{clpca}(x)

\method{augmentation_v}{clpca}(x)

\method{augmentation_coord}{clpca}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"lpca"} and \code{"lsvd"} from the
\strong{\link[logisticPCA:logisticPCA-package]{logisticPCA}} package. The signature
functions \code{\link[logisticPCA:logisticPCA]{logisticPCA::logisticPCA()}} and \code{\link[logisticPCA:logisticSVD]{logisticPCA::logisticSVD()}} are
masked with wrappers that add row and column names from the input matrix to
the output matrices.
}
\examples{

data(finches)

finches_mat <- t(finches)

finches_lsvd <- as_tbl_ord(logisticSVD(finches_mat))
ggbiplot(finches_lsvd, aes(x = LSC1, y = LSC2)) +
  geom_u_vector(aes(x = LSC1 * .01, y = LSC2 * .01)) +
  geom_u_text(aes(x = LSC1 * .01, y = LSC2 * .01, label = .name), size = 3) +
  geom_v_label(aes(label = .name), size = 3, alpha = .5)

finches_lpca <- as_tbl_ord(logisticPCA(finches_mat))
ggbiplot(finches_lpca, aes(x = LPC1, y = LPC2)) +
  geom_u_vector(aes(x = LPC1 * .02, y = LPC2 * .02)) +
  geom_u_text_radiate(aes(x = LPC1 * .02, y = LPC2 * .02, label = .name),
                      size = 3) +
  geom_v_label(aes(label = .name), size = 3, alpha = .5)

finches_clpca <- as_tbl_ord(convexLogisticPCA(finches_mat[, -13, drop = FALSE]))
ggbiplot(finches_clpca, aes(x = LPC1, y = LPC2)) +
  geom_u_vector(aes(x = LPC1 * .02, y = LPC2 * .02)) +
  geom_u_text_radiate(aes(x = LPC1 * .02, y = LPC2 * .02, label = .name),
                      size = 3) +
  geom_v_label(aes(label = .name), size = 3, alpha = .5)
# Logistic SVD of Sanderson finches data
finches \%>\% t() \%>\%
  logisticSVD() \%>\%
  as_tbl_ord() \%>\%
  augment() \%>\%
  print() -> finches_lsvd
finches_lsvd \%>\%
  ggbiplot(
    aes(x = LSC1, y = LSC2, label = .name),
    sec.axes = "v", scale.factor = 100
  ) +
  geom_u_vector(alpha = .5, color = "darkred") +
  geom_u_text_radiate(size = 3, color = "darkred") +
  geom_v_label(size = 3, alpha = .5, color = "royalblue3") +
  ggtitle(
    "Logistic SVD of the Galapagos island finches",
    "Islands (finches) scaled to the primary (secondary) axes"
  )
# Logistic PCA of Sanderson finches data
finches \%>\% t() \%>\%
  logisticPCA() \%>\%
  as_tbl_ord() \%>\%
  augment() \%>\%
  print() -> finches_lpca
finches_lpca \%>\%
  ggbiplot(
    aes(x = LPC1, y = LPC2, label = .name),
    sec.axes = "v", scale.factor = 50
  ) +
  geom_u_vector(alpha = .5, color = "darkred") +
  geom_u_text_radiate(size = 3, color = "darkred") +
  geom_v_label(size = 3, alpha = .5, color = "royalblue3") +
  ggtitle(
    "Logistic PCA of the Galapagos island finches",
    "Islands (finches) scaled to the primary (secondary) axes"
  )
# Convex logistic PCA of Sanderson finches data
finches \%>\% t() \%>\%
  # omit finches present on every island
  {.[, ! apply(., 2, all), drop = FALSE]} \%>\%
  convexLogisticPCA() \%>\%
  as_tbl_ord() \%>\%
  augment() \%>\%
  print() -> finches_clpca
finches_clpca \%>\%
  ggbiplot(
    aes(x = LPC1, y = LPC2, label = .name),
    sec.axes = "v", scale.factor = 50
  ) +
  geom_u_vector(alpha = .5, color = "darkred") +
  geom_u_text_radiate(size = 3, color = "darkred") +
  geom_v_label(size = 3, alpha = .5, color = "royalblue3") +
  ggtitle(
    "Convex logistic PCA of the Galapagos island finches",
    "Islands (finches) scaled to the primary (secondary) axes"
  )
}
