% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lra.r
\name{methods-lra}
\alias{methods-lra}
\alias{as_tbl_ord.lra}
\alias{recover_u.lra}
\alias{recover_v.lra}
\alias{recover_inertia.lra}
\alias{recover_coord.lra}
\alias{recover_conference.lra}
\alias{augmentation_u.lra}
\alias{augmentation_v.lra}
\alias{augmentation_coord.lra}
\title{Functionality for log-ratio analysis ('lra') objects}
\usage{
\method{as_tbl_ord}{lra}(x)

\method{recover_u}{lra}(x)

\method{recover_v}{lra}(x)

\method{recover_inertia}{lra}(x)

\method{recover_coord}{lra}(x)

\method{recover_conference}{lra}(x)

\method{augmentation_u}{lra}(x)

\method{augmentation_v}{lra}(x)

\method{augmentation_coord}{lra}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"lra"}, a class introduced in this package to organize
the singular value decomposition of a double-centered log-transformed data
matrix output by \code{\link[=lra]{lra()}}.
}
\examples{
# Log-ratio analysis of 1973 violent crime arrests in the United States
USArrests[, -3] \%>\%
  lra() \%>\%
  as_tbl_ord() \%>\%
  print() -> arrests_lra
# Adapt Exhibit 7.1 in Greenacre (2010)
arrests_lra \%>\%
  confer_inertia(0) \%>\%
  ggbiplot(aes(label = .name)) +
  theme_bw() +
  geom_u_text(color = "darkgreen", size = 3, alpha = .5) +
  geom_v_polygon(fill = NA, linetype = "dashed", color = "brown4") +
  geom_v_point(color = "brown4") +
  geom_v_text(
    color = "brown4", fontface = "bold",
    hjust = "outward", vjust = "outward"
  ) +
  ggtitle(
    "Log-ratio analysis of violent crime arrest rates",
    "United States, 1973"
  )
}
