% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ordr-lra.r
\name{methods-lra}
\alias{methods-lra}
\alias{as_tbl_ord.lra}
\alias{recover_rows.lra}
\alias{recover_cols.lra}
\alias{recover_inertia.lra}
\alias{recover_coord.lra}
\alias{recover_conference.lra}
\alias{augmentation_rows.lra}
\alias{augmentation_cols.lra}
\alias{augmentation_coord.lra}
\title{Functionality for log-ratio analysis ('lra') objects}
\usage{
\method{as_tbl_ord}{lra}(x)

\method{recover_rows}{lra}(x)

\method{recover_cols}{lra}(x)

\method{recover_inertia}{lra}(x)

\method{recover_coord}{lra}(x)

\method{recover_conference}{lra}(x)

\method{augmentation_rows}{lra}(x)

\method{augmentation_cols}{lra}(x)

\method{augmentation_coord}{lra}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"lra"}, a class introduced in this package to organize
the singular value decomposition of a double-centered log-transformed data
matrix output by \code{\link[=lra]{lra()}}.
}
\examples{
# data frame of violent crime arrests in the United States
class(USArrests)
head(USArrests)
# get state abbreviation data
state <- data.frame(
  .name = state.name,
  .abb = state.abb
)
# compute (non-compositional, unweighted) log-ratio analysis
USArrests \%>\%
  subset(select = -UrbanPop) \%>\%
  lra() \%>\%
  as_tbl_ord() \%>\%
  print() -> arrests_lra
# augment log-ratio profiles with names and join state abbreviations
arrests_lra \%>\%
  augment_ord() \%>\%
  left_join_rows(state, by = ".name") \%>\%
  print() -> arrests_lra
# recover state and arrest profiles
head(get_rows(arrests_lra))
get_cols(arrests_lra)
# initially, inertia is conferred on neither factor
get_conference(arrests_lra)
}
\seealso{
Other methods for singular value decomposition-based techniques: 
\code{\link{methods-cancor}},
\code{\link{methods-candisc}},
\code{\link{methods-ca}},
\code{\link{methods-lda}},
\code{\link{methods-mjca}},
\code{\link{methods-prcomp}},
\code{\link{methods-princomp}},
\code{\link{methods-svd}}
}
\concept{methods for singular value decomposition-based techniques}
