% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lra.r
\name{methods-lra}
\alias{methods-lra}
\alias{as_tbl_ord.lra}
\alias{recover_rows.lra}
\alias{recover_cols.lra}
\alias{recover_inertia.lra}
\alias{recover_coord.lra}
\alias{recover_conference.lra}
\alias{augmentation_rows.lra}
\alias{augmentation_cols.lra}
\alias{augmentation_coord.lra}
\title{Functionality for log-ratio analysis ('lra') objects}
\usage{
\method{as_tbl_ord}{lra}(x)

\method{recover_rows}{lra}(x)

\method{recover_cols}{lra}(x)

\method{recover_inertia}{lra}(x)

\method{recover_coord}{lra}(x)

\method{recover_conference}{lra}(x)

\method{augmentation_rows}{lra}(x)

\method{augmentation_cols}{lra}(x)

\method{augmentation_coord}{lra}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"lra"}, a class introduced in this package to organize
the singular value decomposition of a double-centered log-transformed data
matrix output by \code{\link[=lra]{lra()}}.
}
\examples{
# state abbreviations
state <- data.frame(
  .name = state.name,
  .abb = state.abb
)
# Log-ratio analysis of 1973 violent crime arrests in the United States
USArrests \%>\%
  subset(select = -UrbanPop) \%>\%
  lra() \%>\%
  print() -> arrests_lra
arrests_lra \%>\%
  as_tbl_ord() \%>\%
  augment() \%>\%
  left_join_rows(state, by = ".name") \%>\%
  print() -> arrests_lra
# Adapt Exhibit 7.1 in Greenacre (2010)
arrests_lra \%>\%
  confer_inertia(1) \%>\%
  ggbiplot(sec.axes = "v", scale.factor = 1/20) +
  theme_bw() +
  geom_rows_text(
    aes(label = .abb),
    size = 3, color = "darkgreen", alpha = .5
  ) +
  geom_cols_polygon(fill = NA, linetype = "dashed", color = "brown4") +
  geom_cols_text(
    aes(label = .name),
    color = "brown4", fontface = "bold"
  ) +
  ggtitle(
    "Log-ratio analysis of violent crime arrest rates",
    "United States, 1973"
  ) +
  guides(color = FALSE, size = FALSE)
}
