% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ca-mjca.r
\name{methods-mjca}
\alias{methods-mjca}
\alias{as_tbl_ord.mjca}
\title{Functionality for multiple and joint correspondence analysis ('mjca')
objects}
\usage{
\method{as_tbl_ord}{mjca}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"mjca"} from the \strong{\link[ca:mjca]{ca}} package.
}
\examples{
# table of admissions and rejections from UC Berkeley
class(UCBAdmissions)
head(as.data.frame(UCBAdmissions))
# perform multiple correspondence analysis
UCBAdmissions \%>\%
  ca::mjca() \%>\%
  as_tbl_ord() \%>\%
  # augment profiles with names, masses, distances, and inertias
  augment_ord() \%>\%
  print() -> admissions_mca
# recover row and column profiles
head(get_rows(admissions_mca))
get_cols(admissions_mca)
# column-standard biplot of factor levels
admissions_mca \%>\%
  ggbiplot() +
  theme_bw() + theme_biplot() +
  geom_origin() +
  #geom_rows_point(stat = "unique") +
  geom_cols_point(aes(color = .factor, shape = .factor, size = .mass)) +
  geom_cols_text_repel(aes(label = .level, color = .factor),
                       show.legend = FALSE) +
  scale_color_brewer(palette = "Dark2") +
  scale_size_area(guide = "none") +
  labs(color = "Factor level", shape = "Factor level")
# column-principal biplot of factor levels
admissions_mca \%>\%
  confer_inertia("colprincipal") \%>\%
  ggbiplot() +
  theme_bw() + theme_biplot() +
  geom_origin() +
  #geom_rows_point(stat = "unique") +
  geom_cols_point(aes(color = .factor, shape = .factor, size = .mass)) +
  geom_cols_text_repel(aes(label = .level, color = .factor),
                       show.legend = FALSE) +
  scale_color_brewer(palette = "Dark2") +
  scale_size_area(guide = "none") +
  labs(color = "Factor level", shape = "Factor level")
}
\seealso{
Other methods for singular value decomposition-based techniques: 
\code{\link{methods-cancor}},
\code{\link{methods-ca}},
\code{\link{methods-lda}},
\code{\link{methods-lra}},
\code{\link{methods-prcomp}},
\code{\link{methods-princomp}},
\code{\link{methods-svd}}
}
\concept{methods for singular value decomposition-based techniques}
