% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ca-mjca.r
\name{methods-mjca}
\alias{methods-mjca}
\alias{as_tbl_ord.mjca}
\alias{recover_rows.mjca}
\alias{recover_cols.mjca}
\alias{recover_inertia.mjca}
\alias{recover_conference.mjca}
\alias{recover_coord.mjca}
\alias{augmentation_rows.mjca}
\alias{augmentation_cols.mjca}
\alias{augmentation_coord.mjca}
\title{Functionality for multiple and joint correspondence analysis ('mjca')
objects}
\usage{
\method{as_tbl_ord}{mjca}(x)

\method{recover_rows}{mjca}(x)

\method{recover_cols}{mjca}(x)

\method{recover_inertia}{mjca}(x)

\method{recover_conference}{mjca}(x)

\method{recover_coord}{mjca}(x)

\method{augmentation_rows}{mjca}(x)

\method{augmentation_cols}{mjca}(x)

\method{augmentation_coord}{mjca}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"mjca"} from the \strong{\link[ca:mjca]{ca}} package.
}
\examples{
# table of admissions and rejections from UC Berkeley
class(UCBAdmissions)
head(as.data.frame(UCBAdmissions))
# perform multiple correspondence analysis
UCBAdmissions \%>\%
  ca::mjca() \%>\%
  as_tbl_ord() \%>\%
  print() -> admissions_mca
# summarize ordination
glance(admissions_mca)
# recover row and column profiles
head(get_rows(admissions_mca))
get_cols(admissions_mca)
# augment profiles with names, masses, distances, and inertias
augment_ord(admissions_mca)
# summarize artificial coordinates
tidy(admissions_mca)
# scree plot of proportion of variance (inertia)
tidy(admissions_mca) \%>\%
  ggplot(aes(x = .name, y = .prop_var)) +
  theme_bw() +
  scale_y_continuous(labels = scales::percent) +
  geom_col() +
  labs(x = "", y = "Proportion of inertia")
# fortification adds all above columns
fortify(admissions_mca)
# column-standard biplot of factor levels
admissions_mca \%>\%
  ggbiplot() +
  theme_bw() + theme_biplot() +
  geom_origin() +
  #geom_rows_point(stat = "unique") +
  geom_cols_point(aes(color = .factor, shape = .factor, size = .mass)) +
  geom_cols_text_repel(aes(label = .level, color = .factor),
                       show.legend = FALSE) +
  scale_color_brewer(palette = "Dark2") +
  scale_size_area(guide = FALSE) +
  labs(color = "Factor level", shape = "Factor level")
# column-principal biplot of factor levels
admissions_mca \%>\%
  confer_inertia("colprincipal") \%>\%
  ggbiplot() +
  theme_bw() + theme_biplot() +
  geom_origin() +
  #geom_rows_point(stat = "unique") +
  geom_cols_point(aes(color = .factor, shape = .factor, size = .mass)) +
  geom_cols_text_repel(aes(label = .level, color = .factor),
                       show.legend = FALSE) +
  scale_color_brewer(palette = "Dark2") +
  scale_size_area(guide = FALSE) +
  labs(color = "Factor level", shape = "Factor level")
}
