% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-mjca.r
\name{methods-mjca}
\alias{methods-mjca}
\alias{as_tbl_ord.mjca}
\alias{recover_u.mjca}
\alias{recover_v.mjca}
\alias{recover_inertia.mjca}
\alias{recover_conference.mjca}
\alias{recover_coord.mjca}
\alias{augmentation_u.mjca}
\alias{augmentation_v.mjca}
\alias{augmentation_coord.mjca}
\title{Functionality for multiple and joint correspondence analysis ('mjca')
objects}
\usage{
\method{as_tbl_ord}{mjca}(x)

\method{recover_u}{mjca}(x)

\method{recover_v}{mjca}(x)

\method{recover_inertia}{mjca}(x)

\method{recover_conference}{mjca}(x)

\method{recover_coord}{mjca}(x)

\method{augmentation_u}{mjca}(x)

\method{augmentation_v}{mjca}(x)

\method{augmentation_coord}{mjca}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"mjca"} from the \strong{\link[ca:mjca]{ca}} package.
}
