% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-nipals.r
\name{methods-nipals}
\alias{methods-nipals}
\alias{as_tbl_ord.nipals}
\alias{reconstruct.nipals}
\alias{recover_rows.nipals}
\alias{recover_cols.nipals}
\alias{recover_inertia.nipals}
\alias{recover_coord.nipals}
\alias{recover_conference.nipals}
\alias{augmentation_rows.nipals}
\alias{augmentation_cols.nipals}
\alias{augmentation_coord.nipals}
\title{Functionality for non-linear iterative PLS ('nipals') objects}
\usage{
\method{as_tbl_ord}{nipals}(x)

reconstruct.nipals(x)

\method{recover_rows}{nipals}(x)

\method{recover_cols}{nipals}(x)

\method{recover_inertia}{nipals}(x)

\method{recover_coord}{nipals}(x)

\method{recover_conference}{nipals}(x)

\method{augmentation_rows}{nipals}(x)

\method{augmentation_cols}{nipals}(x)

\method{augmentation_coord}{nipals}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"nipals"} as returned by \code{\link[ade4:nipals]{ade4::nipals()}}.
}
\examples{
# example taken from `help(ade4::nipals)`
data(doubs, package = "ade4")
# NIPALS on sites and environmental variables
nip1 <- ade4::nipals(doubs$env)
as_tbl_ord(nip1)
# augment with row and column names
nip1 \%>\%
  as_tbl_ord() \%>\%
  augment_ord() \%>\%
  print() -> nip1ord

# geographic arrangement of sites
cbind(site = rownames(doubs$env), doubs$xy) \%>\%
  tibble::as_tibble() \%>\%
  ggplot(aes(x = x, y = y)) +
  coord_equal() +
  theme_bw() + theme(panel.grid = element_blank()) +
  geom_path() +
  #geom_text(aes(label = site)) +
  geom_label(aes(label = ifelse(as.integer(site) \%\% 4L == 1L,
                                site, NA_character_)))
# biplot of sites and environmental variables
nip1ord \%>\%
  ggbiplot(aes(x = 2, y = 1),
           sec.axes = "v", scale.factor = 10) +
  theme_bw() + theme(panel.grid = element_blank()) +
  geom_rows_path(alpha = .6) +
  #geom_rows_text(aes(label = .name), alpha = .8) +
  geom_label(aes(label = ifelse(as.integer(.name) \%\% 4L == 1L,
                                .name, NA_character_),
                 alpha = .6)) +
  geom_cols_vector(color = "darkred") +
  geom_cols_text_radiate(aes(label = .name), color = "darkred", alpha = 2/3) +
  scale_size_area(guide = "none") +
  guides(alpha = "none")
}
