% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-prcomp.r
\name{methods-prcomp}
\alias{methods-prcomp}
\alias{as_tbl_ord.prcomp}
\alias{reconstruct.prcomp}
\alias{recover_u.prcomp}
\alias{recover_v.prcomp}
\alias{recover_inertia.prcomp}
\alias{recover_coord.prcomp}
\alias{recover_conference.prcomp}
\alias{augment_u.prcomp}
\alias{augment_v.prcomp}
\alias{augment_coord.prcomp}
\title{Functionality for principal components analysis ('prcomp') objects}
\usage{
\method{as_tbl_ord}{prcomp}(x)

\method{reconstruct}{prcomp}(x)

\method{recover_u}{prcomp}(x)

\method{recover_v}{prcomp}(x)

\method{recover_inertia}{prcomp}(x)

\method{recover_coord}{prcomp}(x)

\method{recover_conference}{prcomp}(x)

\method{augment_u}{prcomp}(x)

\method{augment_v}{prcomp}(x)

\method{augment_coord}{prcomp}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to, objects of class 
\code{"prcomp"}.
}
\examples{
# principal components analysis ('prcomp') examples

# calculate several 'prcomp's
x <- USPersonalExpenditure
p <- prcomp(x, center = FALSE, scale = FALSE)
p2 <- prcomp(x, center = TRUE, scale = FALSE)
p3 <- prcomp(x, center = TRUE, scale = TRUE)

# access the 'U' and 'V' matrices
recover_u(p)
recover_v(p)
recover_u(p2)
recover_v(p2)
recover_u(p3)
recover_v(p3)

# reconstruct approximations of the original data
reconstruct(p)
reconstruct(p2)
reconstruct(p3)

# check that the distances between the original and recovered values are small
hist(as.matrix(x) - reconstruct(p))
hist(as.matrix(x) - reconstruct(p2))
hist(as.matrix(x) - reconstruct(p3))

# wrap 'p', 'p2', and 'p3' as 'tbl_ord' objects
b <- as_tbl_ord(p)
b2 <- as_tbl_ord(p2)
b3 <- as_tbl_ord(p3)

# access the names of the artificial coordinates
recover_coord(b)
recover_coord(b2)
recover_coord(b3)

# augment methods
augment_u(b)
augment_v(b)
augment_coord(b)
augment_u(b2)
augment_v(b2)
augment_coord(b2)
augment_u(b3)
augment_v(b3)
augment_coord(b3)

# data frames
fortify(b)
fortify(b2)
fortify(b3)

# biplot of scores and loadings
gg <- ggbiplot(b3) +
  geom_u_point() +
  geom_v_vector()

# view biplot
gg

# add radiating text for loadings
gg + geom_v_text_radiate(aes(label = .name))
}
