% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-prcomp.r
\name{methods-prcomp}
\alias{methods-prcomp}
\alias{as_tbl_ord.prcomp}
\alias{recover_rows.prcomp}
\alias{recover_cols.prcomp}
\alias{recover_inertia.prcomp}
\alias{recover_coord.prcomp}
\alias{recover_conference.prcomp}
\alias{augmentation_rows.prcomp}
\alias{augmentation_cols.prcomp}
\alias{augmentation_coord.prcomp}
\title{Functionality for principal components analysis ('prcomp') objects}
\usage{
\method{as_tbl_ord}{prcomp}(x)

\method{recover_rows}{prcomp}(x)

\method{recover_cols}{prcomp}(x)

\method{recover_inertia}{prcomp}(x)

\method{recover_coord}{prcomp}(x)

\method{recover_conference}{prcomp}(x)

\method{augmentation_rows}{prcomp}(x)

\method{augmentation_cols}{prcomp}(x)

\method{augmentation_coord}{prcomp}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"prcomp"} as returned by \code{\link[stats:prcomp]{stats::prcomp()}}.
}
\examples{
# data frame of Anderson iris species measurements
class(iris)
head(iris)
# compute scaled row-principal components of scaled measurements
iris[, -5] \%>\%
  prcomp(scale = TRUE) \%>\%
  as_tbl_ord() \%>\%
  print() -> iris_pca
# summarize ordination
glance(iris_pca)
# bind species classification to observation coordinates
(iris_pca <- mutate_rows(iris_pca, species = iris$Species))
# recover observation principal coordinates and measurement standard coordinates
head(get_rows(iris_pca))
get_cols(iris_pca)
# augment measurements with names and scaling parameters
augment_ord(iris_pca)
# summarize principal components
tidy(iris_pca)
# fortification of artificial coordinates yields proportion of variance measure
fortify(iris_pca, .matrix = "coord")
# scree plot of inertia
ggplot(iris_pca, .matrix = "coord", aes(x = .name, y = .inertia)) +
  theme_bw() +
  geom_col() +
  labs(x = "", y = "Inertia")
# scree plot of proportion of variance (inertia)
ggplot(iris_pca, .matrix = "coord", aes(x = .name, y = .prop_var)) +
  theme_bw() +
  scale_y_continuous(labels = scales::percent) +
  geom_col() +
  labs(x = "", y = "Proportion of inertia")
# fortification adds all above columns
fortify(iris_pca)
# row-principal biplot
iris_pca \%>\%
  ggbiplot(aes(color = species)) +
  theme_bw() +
  scale_color_brewer(type = "qual", palette = 2) +
  geom_rows_point(alpha = .5) +
  geom_cols_vector(color = "#444444") +
  geom_cols_text_radiate(aes(label = .name), color = "#444444") +
  ggtitle("Row-principal PCA biplot of Anderson iris measurements")
}
\author{
Emily Paul
}
