% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-princomp.r
\name{methods-princomp}
\alias{methods-princomp}
\alias{as_tbl_ord.princomp}
\alias{reconstruct.princomp}
\alias{recover_u.princomp}
\alias{recover_v.princomp}
\alias{recover_inertia.princomp}
\alias{recover_coord.princomp}
\alias{recover_conference.princomp}
\alias{augmentation_u.princomp}
\alias{augmentation_v.princomp}
\alias{augmentation_coord.princomp}
\title{Functionality for principal components analysis ('princomp') objects}
\usage{
\method{as_tbl_ord}{princomp}(x)

\method{reconstruct}{princomp}(x)

\method{recover_u}{princomp}(x)

\method{recover_v}{princomp}(x)

\method{recover_inertia}{princomp}(x)

\method{recover_coord}{princomp}(x)

\method{recover_conference}{princomp}(x)

\method{augmentation_u}{princomp}(x)

\method{augmentation_v}{princomp}(x)

\method{augmentation_coord}{princomp}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"princomp"} as returned by \code{\link[stats:princomp]{stats::princomp()}}.
}
\examples{
# principal components analysis (`princomp`) examples

# calculate a `princomp`
x <- USArrests
p <- princomp(USArrests)

# access the 'U' and 'V' matrices
recover_u(p)
recover_v(p)

# reconstruct an approximation of the original data
reconstruct(p)

# check that the distances between the original and recovered values are small
hist(as.matrix(x) - reconstruct(p))

# access the names of the artificial coordinates
recover_coord(p)

# augment methods
augment_u(p)
augment_v(p)
augmentation_coord(p)

# wrap `p` as a 'tbl_ord' object
b <- as_tbl_ord(p)
# print `b`
b
# pass b to `fortify`
fortify(b)

# biplot of scores and loadings
gg <- ggbiplot(b) +
  geom_u_point() +
  geom_v_vector(aes(x = 100 * Comp.1, y = 100 * Comp.2))
# view biplot
gg
# Unscaled PCA and two-scaled biplot of Anderson iris data
iris[, -5] \%>\%
  princomp() \%>\%
  as_tbl_ord() \%>\%
  confer_inertia(1) \%>\%
  mutate_u(species = iris$Species) \%>\%
  mutate_v(measure = gsub("\\\\.", " ", tolower(names(iris)[-5]))) \%>\%
  print() -> iris_pca
iris_pca \%>\%
  ggbiplot(aes(color = species), sec.axes = "v", scale.factor = 3) +
  theme_bw() +
  scale_color_brewer(type = "qual", palette = 2) +
  geom_u_point() +
  geom_v_vector(color = "#444444") +
  geom_v_text_radiate(aes(label = measure), color = "#444444") +
  ggtitle(
    "Row-principal unscaled PCA biplot of Anderson iris data",
    "Variable loadings scaled to secondary axes"
  )
}
