% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-stats-princomp.r
\name{methods-princomp}
\alias{methods-princomp}
\alias{as_tbl_ord.princomp}
\title{Functionality for principal components analysis ('princomp') objects}
\usage{
\method{as_tbl_ord}{princomp}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"princomp"} as returned by \code{\link[stats:princomp]{stats::princomp()}}.
}
\examples{
# data frame of Anderson iris species measurements
class(iris)
head(iris)
# compute unscaled row-principal components of scaled measurements
iris[, -5] \%>\%
  princomp() \%>\%
  as_tbl_ord() \%>\%
  print() -> iris_pca
# recover observation principal coordinates and measurement standard coordinates
head(get_rows(iris_pca))
get_cols(iris_pca)
# augment measurement coordinates with names and scaling parameters
(iris_pca <- augment_ord(iris_pca))
}
\seealso{
Other methods for singular value decomposition-based techniques: 
\code{\link{methods-cancor}},
\code{\link{methods-ca}},
\code{\link{methods-lda}},
\code{\link{methods-lra}},
\code{\link{methods-mjca}},
\code{\link{methods-prcomp}},
\code{\link{methods-svd}}
}
\author{
Emily Paul
}
\concept{methods for singular value decomposition-based techniques}
