% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-princomp.r
\name{methods-princomp}
\alias{methods-princomp}
\alias{as_tbl_ord.princomp}
\alias{recover_rows.princomp}
\alias{recover_cols.princomp}
\alias{recover_inertia.princomp}
\alias{recover_coord.princomp}
\alias{recover_conference.princomp}
\alias{augmentation_rows.princomp}
\alias{augmentation_cols.princomp}
\alias{augmentation_coord.princomp}
\title{Functionality for principal components analysis ('princomp') objects}
\usage{
\method{as_tbl_ord}{princomp}(x)

\method{recover_rows}{princomp}(x)

\method{recover_cols}{princomp}(x)

\method{recover_inertia}{princomp}(x)

\method{recover_coord}{princomp}(x)

\method{recover_conference}{princomp}(x)

\method{augmentation_rows}{princomp}(x)

\method{augmentation_cols}{princomp}(x)

\method{augmentation_coord}{princomp}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"princomp"} as returned by \code{\link[stats:princomp]{stats::princomp()}}.
}
\examples{
# data frame of Anderson iris species measurements
class(iris)
head(iris)
# compute unscaled row-principal components of scaled measurements
iris[, -5] \%>\%
  princomp() \%>\%
  as_tbl_ord() \%>\%
  print() -> iris_pca
# summarize ordination
glance(iris_pca)
# bind species classification to observation coordinates
(iris_pca <- mutate_rows(iris_pca, species = iris$Species))
# recover observation principal coordinates and measurement standard coordinates
head(get_rows(iris_pca))
get_cols(iris_pca)
# augment measurement coordinates with names and scaling parameters
augment_ord(iris_pca)
# summarize principal components
tidy(iris_pca)
# fortification of artificial coordinates yields proportion of variance measure
fortify(iris_pca, .matrix = "coord")
# scree plot of inertia
ggplot(iris_pca, .matrix = "coord", aes(x = .name, y = .inertia)) +
  theme_bw() +
  geom_col() +
  labs(x = "", y = "Inertia")
# scree plot of proportion of variance (inertia)
ggplot(iris_pca, .matrix = "coord", aes(x = .name, y = .prop_var)) +
  theme_bw() +
  scale_y_continuous(labels = scales::percent) +
  geom_col() +
  labs(x = "", y = "Proportion of inertia")
# fortification adds all above columns
fortify(iris_pca)
# row-principal biplot with coordinate-wise standard deviations
iris_pca \%>\%
  ggbiplot(aes(color = species)) +
  theme_bw() +
  scale_color_brewer(type = "qual", palette = 2) +
  geom_rows_point(alpha = .5) +
  geom_cols_vector(color = "#444444") +
  geom_cols_text_radiate(aes(label = .name), color = "#444444") +
  ggtitle("Row-principal unscaled PCA biplot of Anderson iris measurements") +
  expand_limits(y = c(NA, 2))
}
\author{
Emily Paul
}
