% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-svd.r
\name{methods-svd}
\alias{methods-svd}
\alias{as_tbl_ord.svd}
\alias{reconstruct.svd}
\alias{recover_u.svd}
\alias{recover_v.svd}
\alias{recover_inertia.svd}
\alias{recover_coord.svd}
\alias{recover_conference.svd}
\alias{augmentation_u.svd}
\alias{augmentation_v.svd}
\alias{augmentation_coord.svd}
\title{Functionality for singular value decompositions}
\usage{
\method{as_tbl_ord}{svd}(x)

\method{reconstruct}{svd}(x)

\method{recover_u}{svd}(x)

\method{recover_v}{svd}(x)

\method{recover_inertia}{svd}(x)

\method{recover_coord}{svd}(x)

\method{recover_conference}{svd}(x)

\method{augmentation_u}{svd}(x)

\method{augmentation_v}{svd}(x)

\method{augmentation_coord}{svd}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"svd"} returned by \code{\link[=svd_ord]{svd_ord()}}.
}
