% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-svd.r
\name{methods-svd}
\alias{methods-svd}
\alias{as_tbl_ord.svd}
\alias{recover_u.svd}
\alias{recover_v.svd}
\alias{recover_inertia.svd}
\alias{recover_coord.svd}
\alias{augment_u.svd}
\alias{augment_v.svd}
\alias{augment_coord.svd}
\alias{reconstruct.svd}
\title{Functionality for singular value decompositions}
\usage{
\method{as_tbl_ord}{svd}(x)

\method{recover_u}{svd}(x)

\method{recover_v}{svd}(x)

\method{recover_inertia}{svd}(x)

\method{recover_coord}{svd}(x)

\method{recover_inertia}{svd}(x)

\method{augment_u}{svd}(x)

\method{augment_v}{svd}(x)

\method{augment_coord}{svd}(x)

\method{reconstruct}{svd}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to, objects of class 
\code{"svd"}. (This is a class introduced in this package to identify 
objects returned by \code{\link{svd}}, which is masked by a wrapper that
adds the class attribute.)
}
