% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinate.r
\name{ordinate}
\alias{ordinate}
\title{Specify columns of a data frame to ordinate and augment}
\usage{
ordinate(data, cols, model, augment = NULL, ...)
}
\arguments{
\item{data}{A data frame.}

\item{cols}{<\code{\link[tidyr:tidyr_tidy_select]{tidy-select}}> Columns of \code{data} to
pass to \code{model}.}

\item{model}{An ordination function whose output is coercible to class
'\link{tbl_ord}'.}

\item{augment}{<\code{\link[=tidyr_tidy_select]{tidy-select}}> Columns of \code{data} to
augment to the row data of the ordination.}

\item{...}{Additional arguments passed to \code{model}.}
}
\description{
Specify columns of a data frame to ordinate and augment
}
\examples{
# PCA of iris data
ordinate(iris, 1:4, prcomp)
ordinate(iris, -Species, princomp, augment = c(Sepal.Width, Species))
# LRA of arrest data
ordinate(USArrests, cols = c(Murder, Rape, Assault), lra)
# CA of hair & eye color data
haireye <- as.data.frame(rowSums(HairEyeColor, dims = 2L))
ordinate(haireye, cols = everything(), model = ca::ca, augment = everything())
}
