% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinate.r
\name{ordinate}
\alias{ordinate}
\title{Fit an ordination model to columns of a data frame}
\usage{
ordinate(data, cols, model, augment, ...)
}
\arguments{
\item{data}{A data frame.}

\item{cols}{<\code{\link[tidyr:tidyr_tidy_select]{tidy-select}}> Columns of \code{data} to
pass to \code{model}. If missing, all columns are used.}

\item{model}{An ordination function whose output is coercible to class
'\link{tbl_ord}'. Alternatively, a formula \code{~ fun(., ...)} where \code{fun} is such a
function and other arguments are explicit, which will be evaluated with
\code{data} in place of \code{.}.}

\item{augment}{<\code{\link[=tidyr_tidy_select]{tidy-select}}> Columns of \code{data} to
augment to the row data of the ordination. If missing, all columns not
included in \code{cols} will be augmented.}

\item{...}{Additional arguments passed to \code{model}.}
}
\description{
This is a convenience function to fit an ordination model to a
specified subset of columns of a data frame and augment the tbl_ord object
with both its intrinsic diagnostics (via \verb{[augment_ord()]}) and any
additional columns of the data frame (via \verb{[cbind_rows()]}).
}
\examples{
# PCA of iris data
ordinate(iris, -Species, princomp, augment = c(Sepal.Width, Species))
ordinate(iris, 1:4, ~ prcomp(., center = TRUE, scale. = TRUE))

# LRA of arrest data
ordinate(USArrests, cols = c(Murder, Rape, Assault), lra)

# CA of hair & eye color data
haireye <- as.data.frame(rowSums(HairEyeColor, dims = 2L))
ordinate(haireye, cols = everything(), model = MASS::corresp)

# FA of Swiss social data
ordinate(swiss, model = factanal, factors = 2L, scores = "Bartlett")

# LDA of iris data
ordinate(iris, model = ~ lda_ord(.[, 1:4], .[, 5], ret.x = TRUE))

# CCA of savings data
ordinate(
  LifeCycleSavings[, c("pop15", "pop75")],
  # second data set must be handled as an additional parameter to `model`
  y = LifeCycleSavings[, c("sr", "dpi", "ddpi")],
  model = cancor_ord, scores = TRUE
)
}
