% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-.r, R/coord-rect.r, R/geom-axis.r,
%   R/geom-intervals.r, R/geom-isoline.r, R/geom-origin.r,
%   R/geom-text-radiate.r, R/geom-vector.r, R/stat-center.r, R/stat-chull.r,
%   R/stat-cone.r, R/stat-matrix.r, R/stat-rule.r, R/stat-scale.r,
%   R/stat-spantree.r, R/zzz-biplot-stats.r
\docType{data}
\name{ordr-ggproto}
\alias{ordr-ggproto}
\alias{CoordRect}
\alias{CoordBiplot}
\alias{GeomAxis}
\alias{GeomPointranges}
\alias{GeomLineranges}
\alias{GeomIsoline}
\alias{GeomOrigin}
\alias{GeomUnitCircle}
\alias{GeomTextRadiate}
\alias{GeomVector}
\alias{StatCenter}
\alias{StatStar}
\alias{StatChull}
\alias{StatCone}
\alias{StatRows}
\alias{StatCols}
\alias{StatRule}
\alias{StatRowsRule}
\alias{StatColsRule}
\alias{StatScale}
\alias{StatRowsScale}
\alias{StatColsScale}
\alias{StatSpantree}
\alias{StatRowsEllipse}
\alias{StatColsEllipse}
\alias{StatRowsCenter}
\alias{StatColsCenter}
\alias{StatRowsStar}
\alias{StatColsStar}
\alias{StatRowsChull}
\alias{StatColsChull}
\alias{StatRowsCone}
\alias{StatColsCone}
\alias{StatRowsSpantree}
\alias{StatColsSpantree}
\title{ggproto classes created and adapted for ordr}
\description{
In addition to geometric element layers (geoms) based on
base-\strong{ggplot2} layers like \code{geom_point()} but specified to matrix factors
as \code{geom_row_point()}, \strong{ordr} introduces \link[ggplot2:ggproto]{ggproto}
classes for some additional geometric elements commonly used in biplots.
The factor-specific geoms invoke the statistical transformation layers
(stats) \code{stat_rows()} and \code{stat_cols()}, which specify the matrix factor.
Because each ggplot layer consists of only one stat and one geom, this
necessitates that ggproto classes for new stats must also come in \verb{*Rows}
and \verb{*Cols} flavors.

Determine axis limits and offset vectors from reference data.
}
\section{Biplot layers}{

\code{\link[=ggbiplot]{ggbiplot()}} uses \code{\link[ggplot2:fortify]{ggplot2::fortify()}} internally to produce a single data
frame with a \code{.matrix} column distinguishing the subjects (\code{"rows"}) and
variables (\code{"cols"}). The stat layers \code{stat_rows()} and \code{stat_cols()} simply
filter the data frame to one of these two.

The geom layers \verb{geom_rows_*()} and \verb{geom_cols_*()} call the corresponding
stat in order to render plot elements for the corresponding factor matrix.
\verb{geom_dims_*()} selects a default matrix based on common practice, e.g.
points for rows and arrows for columns.
}

\seealso{
\code{\link[ggplot2:ggplot2-ggproto]{ggplot2::ggplot2-ggproto}} and \link[ggplot2:ggproto]{ggplot2::ggproto} for explanations
of base ggproto classes in \strong{ggplot2} and how to create new ones.
}
\keyword{datasets}
