% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-.r, R/coord-rect.r, R/geom-axis.r,
%   R/geom-bag.r, R/geom-interpolation.r, R/geom-intervals.r, R/geom-isoline.r,
%   R/geom-origin.r, R/geom-rule.r, R/geom-text-radiate.r, R/geom-vector.r,
%   R/stat-chull.r, R/stat-depth.r, R/stat-bag.r, R/stat-center.r,
%   R/stat-cone.r, R/stat-matrix.r, R/stat-referent.r, R/stat-projection.r,
%   R/stat-rule.r, R/stat-scale.r, R/stat-spantree.r, R/zzz-biplot-stats.r
\docType{data}
\name{ordr-ggproto}
\alias{ordr-ggproto}
\alias{CoordRect}
\alias{CoordBiplot}
\alias{GeomAxis}
\alias{GeomBag}
\alias{GeomInterpolation}
\alias{GeomPointranges}
\alias{GeomLineranges}
\alias{GeomIsoline}
\alias{GeomOrigin}
\alias{GeomUnitCircle}
\alias{GeomRule}
\alias{GeomTextRadiate}
\alias{GeomVector}
\alias{StatChull}
\alias{StatPeel}
\alias{StatDepth}
\alias{StatDepthFilled}
\alias{StatBag}
\alias{StatCenter}
\alias{StatStar}
\alias{StatCone}
\alias{StatRows}
\alias{StatCols}
\alias{StatReferent}
\alias{StatProjection}
\alias{StatRule}
\alias{StatScale}
\alias{StatSpantree}
\alias{StatRowsDensity2d}
\alias{StatColsDensity2d}
\alias{StatRowsDensity2dFilled}
\alias{StatColsDensity2dFilled}
\alias{StatRowsEllipse}
\alias{StatColsEllipse}
\alias{StatRowsBag}
\alias{StatColsBag}
\alias{StatRowsCenter}
\alias{StatColsCenter}
\alias{StatRowsStar}
\alias{StatColsStar}
\alias{StatRowsChull}
\alias{StatColsChull}
\alias{StatRowsPeel}
\alias{StatColsPeel}
\alias{StatRowsCone}
\alias{StatColsCone}
\alias{StatRowsDepth}
\alias{StatColsDepth}
\alias{StatRowsDepthFilled}
\alias{StatColsDepthFilled}
\alias{StatRowsProjection}
\alias{StatColsProjection}
\alias{StatRowsRule}
\alias{StatColsRule}
\alias{StatRowsScale}
\alias{StatColsScale}
\alias{StatRowsSpantree}
\alias{StatColsSpantree}
\title{ggproto classes created and adapted for ordr}
\description{
In addition to geometric element layers (geoms) based on
base-\strong{ggplot2} layers like \code{geom_point()} but specified to matrix factors
as \code{geom_row_point()}, \strong{ordr} introduces \link[ggplot2:ggproto]{ggproto}
classes for some additional geometric elements commonly used in biplots.
The factor-specific geoms invoke the statistical transformation layers
(stats) \code{stat_rows()} and \code{stat_cols()}, which specify the matrix factor.
Because each ggplot layer consists of only one stat and one geom, this
necessitates that ggproto classes for new stats must also come in \verb{*Rows}
and \verb{*Cols} flavors.
}
\seealso{
\code{\link[ggplot2:ggplot2-ggproto]{ggplot2::ggplot2-ggproto}} and \link[ggplot2:ggproto]{ggplot2::ggproto} for explanations
of base ggproto classes in \strong{ggplot2} and how to create new ones.
}
\keyword{datasets}
