% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-.r, R/geom-vector.r, R/geom-addition.r,
%   R/geom-axis.r, R/geom-intervals.r, R/geom-isoline.r, R/geom-origin.r,
%   R/geom-text-radiate.r, R/geom-unit-circle.r, R/stat-center.r,
%   R/stat-chull.r, R/stat-cone.r, R/stat-ellipse.r, R/stat-matrix.r,
%   R/stat-scale.r, R/stat-spantree.r
\docType{data}
\name{ordr-ggproto}
\alias{ordr-ggproto}
\alias{GeomVector}
\alias{GeomAddition}
\alias{GeomAxis}
\alias{GeomPointranges}
\alias{GeomLineranges}
\alias{GeomIsoline}
\alias{GeomOrigin}
\alias{GeomTextRadiate}
\alias{GeomUnitCircle}
\alias{StatCenter}
\alias{StatRowsCenter}
\alias{StatColsCenter}
\alias{StatStar}
\alias{StatRowsStar}
\alias{StatColsStar}
\alias{StatChull}
\alias{StatRowsChull}
\alias{StatColsChull}
\alias{StatCone}
\alias{StatRowsCone}
\alias{StatColsCone}
\alias{StatRowsEllipse}
\alias{StatColsEllipse}
\alias{StatRows}
\alias{StatCols}
\alias{StatScale}
\alias{StatRowsScale}
\alias{StatColsScale}
\alias{StatSpantree}
\alias{StatRowsSpantree}
\alias{StatColsSpantree}
\title{ggproto classes created and adapted for ordr}
\description{
In addition to geometric element layers (geoms) based on
base-\strong{ggplot2} layers like \code{geom_point()} but specified to matrix factors
as \code{geom_row_point()}, \strong{ordr} introduces \link[ggplot2:ggproto]{ggproto}
classes for some additional geometric elements commonly used in biplots.
The factor-specific geoms invoke the statistical transformation layers
(stats) \code{stat_rows()} and \code{stat_cols()}, which specify the matrix factor.
Because each ggplot layer consists of only one stat and one geom, this
necessitates that ggproto classes for new stats must also come in \verb{*_Rows}
and \verb{*_Cols} flavors.
}
\seealso{
\code{\link[ggplot2:ggplot2-ggproto]{ggplot2::ggplot2-ggproto}} and \link[ggplot2:ggproto]{ggplot2::ggproto} for explanations
of base ggproto classes in \strong{ggplot2} and how to create new ones.
}
\keyword{datasets}
