% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordr.r
\docType{package}
\name{ordr}
\alias{ordr}
\alias{ordr-package}
\title{\strong{ordr} package}
\description{
This is a \strong{\link[tidyverse:tidyverse]{tidyverse}} extension for
handling, manipulating, and visualizing ordination models with consistent
conventions and in a tidy workflow.
}
\details{
This package is designed to integrate ordination analysis and biplot
visualization into a \strong{tidyverse} workflow. It is inspired in particular by
the extensions \strong{ggbiplot} and \strong{tidygraph}.

The package consists in several modules:
\itemize{
\item the \code{"tbl_ord"} class, a wrapper for various ordination object classes
\item extracting \link{augmentation} for the factors of an ordination
\item using \link{dplyr-verbs} to add \link{annotation} to the factors of an ordination
\item manipulating the coordinates of an ordination via \link{alignment} with other
objects and \link{conference} of inertia
\item methods of the above generics for several widely-used ordination object
classes
\item convenient \link{formatting} of ordination objects that respects the above
manipulations
\item \link{ggbiplot}, a \strong{ggplot2} extension for rendering biplots, including
several plot layers, integrated with the \code{"tbl_ord"} class
}
}
\references{
Podani J (2000) \emph{Introduction to the Exploration of Multivariate Biological
Data}. Backhuys Publishers, ISBN 90-5782-067-6.
\url{http://ramet.elte.hu/~podani/books.html}

Palmer M \emph{Ordination Methods for Ecologists}. Website, accessed 2019-07-12.
\url{http://ordination.okstate.edu/}

Gabriel KR (1971) "The biplot graphic display of matrices with application to
principal component analysis". \emph{Biometrika} 58(3), 453--467.
\doi{10.1093/biomet/58.3.453}

Greenacre MJ (2010) \emph{Biplots in Practice}. Fundacion BBVA, ISBN:
978-84-923846.
\url{https://www.fbbva.es/microsite/multivariate-statistics/biplots.html}

ttnphns (2015) \emph{Answer to} "PCA and Correspondence analysis in their relation
to Biplot". CrossValidated, accessed 2019-07-12.
\url{https://stats.stackexchange.com/a/141755/68743}
}
