% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{qswur_usa}
\alias{qswur_usa}
\title{U.S. university rankings}
\format{
A \link[tibble:tibble]{tibble}.
}
\source{
Quacquarelli Symonds (2021).
}
\usage{
data(qswur_usa)
}
\description{
Classifications and rankings of U.S. universities for the years
2017--2020.
}
\details{
Ranking data were obtained from the public QS website.
}
\references{
Quacquarelli Symonds (2021) "University Rankings". TopUniversities.com
\url{https://www.topuniversities.com/university-rankings}.
}
\keyword{datasets}
