% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{ratios}
\alias{ratios}
\title{Ingredient ratios for several basic recipes}
\format{A \link[tibble:tibble]{tibble}.}
\source{
Ruhlman (2009).
}
\description{
Ratios of ingredients for several basic recipes presented in a
popular cooking book.
}
\examples{
# Correspondence analysis of Ruhlman's recipe ratios
ratios \%>\%
  select(-chapter) \%>\%
  tibble::column_to_rownames("recipe") \%>\%
  ca::ca() \%>\%
  as_tbl_ord() \%>\%
  augment() \%>\%
  mutate_u(recipe = .name, chapter = ratios$chapter) \%>\%
  mutate_v(ingredient = .name) \%>\%
  print() -> ratios_ca
ratios_ca \%>\%
  confer_inertia("rowprincipal") \%>\%
  ggbiplot() +
  theme_bw() +
  geom_v_vector(alpha = .5) +
  geom_u_text(aes(label = recipe, color = chapter)) +
  ggtitle("Row-principal CA biplot of Ruhlman recipe ratios")
}
\references{
Ruhlman M (2009) \emph{Ratio: The Simple Codes Behind the Craft of Everyday
Cooking}. Scribner, ISBN-13: 978-1416571728.
}
\keyword{datasets}
