% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.r
\name{reconstruct}
\alias{reconstruct}
\alias{reconstruct.list}
\title{Reconstruct data from a tbl_ord}
\usage{
reconstruct(x)

\method{reconstruct}{list}(x)
}
\arguments{
\item{x}{A \link{tbl_ord} object, or an object coercible to one.}
}
\description{
This convenience function reconstructs an approximation to the
(possibly transformed) original data matrix from the ordination factors and
metadata. The methods are most useful as points of reference for
understanding how ordinations are obtained from data.
}
