% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.r
\name{reconstruct}
\alias{reconstruct}
\alias{reconstruct.list}
\title{Reconstruct data from an ordination}
\usage{
reconstruct(x)

\method{reconstruct}{list}(x)
}
\arguments{
\item{x}{A \code{\link{tbl_ord}}, or an object convertible to one.}
}
\description{
This convenience function reconstructs an approximation to the (possibly
transformed) original data matrix from the ordination factors and metadata.
The methods are most useful as points of reference for understanding how
ordinations are obtained from data.
}
