% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lm.r, R/methods-lpca.r, R/reconstruct.r
\name{reconstruct.lm}
\alias{reconstruct.lm}
\alias{reconstruct.lsvd}
\alias{reconstruct}
\alias{reconstruct.list}
\title{Reconstruct approximations to original data from ordination factors}
\usage{
\method{reconstruct}{lm}(x)

\method{reconstruct}{lsvd}(x)

reconstruct(x)

\method{reconstruct}{list}(x)
}
\arguments{
\item{x}{A bibble, or an object convertible to one.}
}
\description{
Reconstruct approximations to original data from ordination factors
}
