% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmentation.r, R/fortify.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{augment}
\alias{tidy}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{broom}{\code{\link[broom]{tidy}}}

  \item{generics}{\code{\link[generics]{augment}}}
}}

