% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify.r, R/geom-text.r, R/utils.r,
%   R/ord-augmentation.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tidy}
\alias{geom_text_repel}
\alias{geom_label_repel}
\alias{\%>\%}
\alias{augment}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{tidy}}, \code{\link[generics]{augment}}}

  \item{ggrepel}{\code{\link[ggrepel]{geom_text_repel}}, \code{\link[ggrepel]{geom_label_repel}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}

