% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r, R/geom-text.r, R/ord-tidiers.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{geom_text_repel}
\alias{geom_label_repel}
\alias{tidy}
\alias{glance}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{glance}}, \code{\link[generics]{tidy}}}

  \item{ggrepel}{\code{\link[ggrepel:geom_text_repel]{geom_label_repel}}, \code{\link[ggrepel]{geom_text_repel}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}
}}

