% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r, R/ord-tidiers.r, R/zzz-biplot-geoms.r,
%   R/zzz-biplot-stats.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{tidy}
\alias{glance}
\alias{geom_text_repel}
\alias{geom_label_repel}
\alias{geom_axis}
\alias{geom_pointranges}
\alias{geom_lineranges}
\alias{geom_isoline}
\alias{geom_text_radiate}
\alias{geom_vector}
\alias{geom_bagplot}
\alias{geom_rule}
\alias{position_nudge_repel}
\alias{minpp}
\alias{maxpp}
\alias{minabspp}
\alias{stat_center}
\alias{stat_star}
\alias{stat_chull}
\alias{stat_peel}
\alias{stat_cone}
\alias{stat_depth}
\alias{stat_depth_filled}
\alias{stat_scale}
\alias{stat_spantree}
\alias{stat_bagplot}
\alias{stat_rule}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{glance}}, \code{\link[generics]{tidy}}}

  \item{gggda}{\code{\link[gggda]{geom_axis}}, \code{\link[gggda]{geom_bagplot}}, \code{\link[gggda]{geom_isoline}}, \code{\link[gggda]{geom_lineranges}}, \code{\link[gggda:geom_lineranges]{geom_pointranges}}, \code{\link[gggda]{geom_rule}}, \code{\link[gggda]{geom_text_radiate}}, \code{\link[gggda]{geom_vector}}, \code{\link[gggda:stat_rule]{maxpp}}, \code{\link[gggda:stat_rule]{minabspp}}, \code{\link[gggda:stat_rule]{minpp}}, \code{\link[gggda]{stat_bagplot}}, \code{\link[gggda]{stat_center}}, \code{\link[gggda]{stat_chull}}, \code{\link[gggda]{stat_cone}}, \code{\link[gggda]{stat_depth}}, \code{\link[gggda:stat_depth]{stat_depth_filled}}, \code{\link[gggda:stat_chull]{stat_peel}}, \code{\link[gggda]{stat_rule}}, \code{\link[gggda]{stat_scale}}, \code{\link[gggda]{stat_spantree}}, \code{\link[gggda:stat_center]{stat_star}}}

  \item{ggrepel}{\code{\link[ggrepel:geom_text_repel]{geom_label_repel}}, \code{\link[ggrepel]{geom_text_repel}}, \code{\link[ggrepel]{position_nudge_repel}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}
}}

