% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-chull.r
\docType{data}
\name{stat-biplot-chull}
\alias{stat-biplot-chull}
\alias{StatChull}
\alias{stat_chull}
\alias{StatRowsChull}
\alias{StatColsChull}
\alias{stat_rows_chull}
\alias{stat_cols_chull}
\title{Restrict ordination data to the convex hulls of both matrix factors}
\format{
An object of class \code{StatChull} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.

An object of class \code{StatRowsChull} (inherits from \code{StatChull}, \code{Stat}, \code{ggproto}, \code{gg}) of length 2.

An object of class \code{StatColsChull} (inherits from \code{StatChull}, \code{Stat}, \code{ggproto}, \code{gg}) of length 2.
}
\usage{
stat_chull(
  mapping = NULL,
  data = NULL,
  geom = "polygon",
  position = "identity",
  conical = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)

stat_rows_chull(
  mapping = NULL,
  data = NULL,
  geom = "polygon",
  position = "identity",
  conical = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)

stat_cols_chull(
  mapping = NULL,
  data = NULL,
  geom = "polygon",
  position = "identity",
  conical = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{The geometric object to use display the data}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{conical}{Logical; whether to include the origin when calculating the
hull. Defaults to \code{FALSE}.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:layer]{ggplot2::layer()}}.}
}
\description{
As used in a \strong{\link[ggplot2:ggplot2-package]{ggplot2}} vignette, this stat
layer restricts a dataset with \code{x} and \code{y} variables to its convex hull.
The biplot extension restricts each matrix factor to its own hull. The
\code{conical} parameter controls whether the origin is included in the
calculation, though note that the result is not the (unbounded) \href{https://en.wikipedia.org/wiki/Conical_combination#Conical_hull}{conical hull}.
}
\section{Biplot layers}{

\code{\link[=ggbiplot]{ggbiplot()}} uses \code{\link[ggplot2:fortify]{ggplot2::fortify()}} internally to produce a single data
frame with a \code{.matrix} column distinguishing the subjects (\code{"rows"}) and
variables (\code{"cols"}). The stat layers \code{stat_rows()} and \code{stat_cols()} simply
filter the data frame to one of these two.

The geom layers \verb{geom_rows_*()} and \verb{geom_cols_*()} call the corresponding
stat in order to render plot elements for the corresponding factor matrix.
\verb{geom_dims_*()} selects a default matrix based on common practice, e.g.
points for rows and arrows for columns.
}

\examples{
# Confer inertia in CA between rows and columns of benthic sample data
data(benthos)
benthos \%>\%
  ca::ca() \%>\%
  as_tbl_ord() \%>\%
  print() -> benthos_ca
# Reproduce Exhibit 8.3
benthos_ca \%>\%
  confer_inertia("colprincipal") \%>\%
  ggbiplot(aes(label = .name)) +
  theme_bw() +
  geom_rows_point(color = "brown", size = 1) +
  geom_cols_text(color = "darkgreen")
# Reproduce Exhibit 8.4
benthos_ca \%>\%
  augment(.matrix = "rows") \%>\%
  transmute_rows(mass = .mass, sqrt_mass = sqrt(.mass)) \%>\%
  confer_inertia("colprincipal") \%>\%
  ggbiplot(aes(label = .name), scale_rows = "sqrt_mass") +
  theme_bw() +
  geom_rows_vector(color = "brown", arrow = NULL) +
  geom_rows_point(aes(size = mass), color = "brown", shape = 17) +
  scale_size_continuous(range = c(1, 4), guide = "none") +
  geom_rows_text(
    stat = "chull",
    color = "brown", hjust = "outward", vjust = "outward"
  ) +
  geom_cols_point(color = "darkgreen") +
  geom_cols_text_repel(color = "darkgreen", min.segment.length = 2)
}
\keyword{datasets}
