% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-bagplot.r
\name{stat_bagplot}
\alias{stat_bagplot}
\title{Bagplots}
\usage{
stat_bagplot(
  mapping = NULL,
  data = NULL,
  geom = "bagplot",
  position = "identity",
  bag_var = "depth",
  fraction = 0.5,
  median = TRUE,
  fence = TRUE,
  outliers = TRUE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{bag_var}{The stat to use to construct the bag; one of \code{"depth"},
\code{"density"}, and \code{"chull"}.}

\item{fraction}{Fraction of the data to include in the bag.}

\item{median, fence, outliers}{Logical indicators whether to include median,
fence, and outliers in the composite output.}

\item{...}{
  Arguments passed on to \code{\link[=stat_depth]{stat_depth}}, \code{\link[ggplot2:geom_density_2d]{ggplot2::stat_density_2d}}
  \describe{
    \item{\code{notion}}{Character; the name of the depth function (passed to
\code{\link[ddalpha:depth.]{ddalpha::depth.()}}).}
    \item{\code{n}}{Number of grid points in each direction.}
    \item{\code{h}}{Bandwidth (vector of length two). If \code{NULL}, estimated
using \code{\link[MASS:bandwidth.nrd]{MASS::bandwidth.nrd()}}.}
    \item{\code{adjust}}{A multiplicative bandwidth adjustment to be used if 'h' is
'NULL'. This makes it possible to adjust the bandwidth while still
using the a bandwidth estimator. For example, \code{adjust = 1/2} means
use half of the default bandwidth.}
  }}
}
\description{
Construct medians, bags, fences, and outlier specifications for
bagplots.
}
\details{
A bagplot comprises a single, often filled, contour (the "bag")
overlaid upon the \link[=stat_chull]{convex hull} (the "fence") and a
scatterplot of outliers (the "loop"). Reusseeuw &al (1999) also suggest the
term "bag-and-bolster plot".
}
\section{Biplot layers}{

\code{\link[=ggbiplot]{ggbiplot()}} uses \code{\link[ggplot2:fortify]{ggplot2::fortify()}} internally to produce a single data
frame with a \code{.matrix} column distinguishing the subjects (\code{"rows"}) and
variables (\code{"cols"}). The stat layers \code{stat_rows()} and \code{stat_cols()} simply
filter the data frame to one of these two.

The geom layers \verb{geom_rows_*()} and \verb{geom_cols_*()} call the corresponding
stat in order to render plot elements for the corresponding factor matrix.
\verb{geom_dims_*()} selects a default matrix based on common practice, e.g.
points for rows and arrows for columns.
}

\section{Ordination aesthetics}{

This statistical transformation is compatible with the convenience function
\code{\link[=ord_aes]{ord_aes()}}.

Some transformations (e.g. \code{\link[=stat_center]{stat_center()}}) commute with projection to the
lower (1 or 2)-dimensional biplot space. If they detect aesthetics of the
form \verb{..coord[0-9]+}, then \code{..coord1} and \code{..coord2} are converted to \code{x} and
\code{y} while any remaining are ignored.

Other transformations (e.g. \code{\link[=stat_spantree]{stat_spantree()}}) yield different results in a
lower-dimensional biplot when they are computed before versus after
projection. If the stat layer detects these aesthetics, then the
transformation is performed before projection, and the results in the first
two dimensions are returned as \code{x} and \code{y}.

A small number of transformations (\code{\link[=stat_rule]{stat_rule()}}) are incompatible with
ordination aesthetics but will accept \code{ord_aes()} without warning.
}

\examples{
judges_pca <- ordinate(USJudgeRatings, prcomp)
ggbiplot(judges_pca, sec.axes = "cols") +
  geom_rows_bagplot() +
  geom_cols_vector(aes(label = name))
iris_pca <- ordinate(iris, prcomp, cols = seq(4), scale. = TRUE)
ggbiplot(iris_pca) +
  geom_rows_point(aes(color = Species)) +
  stat_rows_bagplot(aes(color = Species, fill = Species)) +
  geom_cols_vector(aes(label = name))
}
\references{
Rousseeuw PJ, Ruts I, & Tukey JW (1999) "The Bagplot: A Bivariate Boxplot".
\emph{The American Statistician}, \strong{53}(4): 382--387.
\doi{10.1080/00031305.1999.10474494}
}
