% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-rule.r
\name{stat_rule}
\alias{stat_rule}
\alias{minpp}
\alias{maxpp}
\alias{minabspp}
\title{Construct limited rulers offset from the origin}
\usage{
stat_rule(
  mapping = NULL,
  data = NULL,
  geom = "axis",
  position = "identity",
  .referent = NULL,
  referent = NULL,
  fun.lower = "minpp",
  fun.upper = "maxpp",
  fun.offset = "minabspp",
  fun.args = list(),
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)

minpp(x, p = 0.1)

maxpp(x, p = 0.1)

minabspp(x, p = 0.1)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{The geometric object to use to display the data for this layer.
When using a \verb{stat_*()} function to construct a layer, the \code{geom} argument
can be used to override the default coupling between stats and geoms. The
\code{geom} argument accepts the following:
\itemize{
\item A \code{Geom} ggproto subclass, for example \code{GeomPoint}.
\item A string naming the geom. To give the geom as a string, strip the
function name of the \code{geom_} prefix. For example, to use \code{geom_point()},
give the geom as \code{"point"}.
\item For more information and other ways to specify the geom, see the
\link[ggplot2:layer_geoms]{layer geom} documentation.
}}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{.referent}{A character string indicating the matrix factor(s) of an
ordination model to include in \code{referent}, handled the same way as
\code{.matrix}.}

\item{referent}{The point cloud to rule; a data frame with \code{x} and \code{y}
columns.}

\item{fun.lower, fun.upper, fun.offset}{Functions used to determine the limits
of the rules and the translations of the axes from the projections of
\code{referent} onto the axes and onto their normal vectors.}

\item{fun.args}{Arguments passed to the \verb{fun.*}.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:layer]{ggplot2::layer()}}.}

\item{x}{A numeric vector.}

\item{p}{A numeric value; the proportion of a range used as a buffer.}
}
\value{
A ggproto \link[ggplot2:layer]{layer}.
}
\description{
Determine axis limits and offset vectors from reference data.
}
\details{
Biplots with several axes can become cluttered and illegible. When this
happens, Gower, Gardner--Lubbe, & le Roux (2011) recommend to translate the
axes to a new point of intersection away from the origin, adjusting the axis
markers accordingly. Then the axes converge in a region of the plot offset
from most position markers or other elements. An alternative solution,
implemented in the \strong{\link{bipl5}} package, is to translate each axis
orthogonally away from the origin, which preserves the axis markers. This is
the technique implemented here.

Separately, axes that fill the plotting window are uninformative when they
exceed the range of the plotted position markers projected onto them. They
may even be misinformative, suggesting that linear relationships extrapolate
outside the data range. In these cases, Gower and Harding (1988) recommend
using finite ranges determined by the data projection onto each axis.

Three functions control these operations: \code{fun.offset} computes the
orthogonal distance of each axis from the origin, and \code{fun.lower} and
\code{fun.upper} compute the distance along each axis of the endpoints to the
(offset) origin. Both functions depend on what position data is to be offset
from or limited to, which must be either passed manually to the \code{referent}
parameter or encoded as named matrix factors to the helper parameter
\code{.referent}.
}
\section{Biplot layers}{

\code{\link[=ggbiplot]{ggbiplot()}} uses \code{\link[ggplot2:fortify]{ggplot2::fortify()}} internally to produce a single data
frame with a \code{.matrix} column distinguishing the subjects (\code{"rows"}) and
variables (\code{"cols"}). The stat layers \code{stat_rows()} and \code{stat_cols()} simply
filter the data frame to one of these two.

The geom layers \verb{geom_rows_*()} and \verb{geom_cols_*()} call the corresponding
stat in order to render plot elements for the corresponding factor matrix.
\verb{geom_dims_*()} selects a default matrix based on common practice, e.g.
points for rows and arrows for columns.
}

\section{Ordination aesthetics}{

The convenience function \code{\link[=ord_aes]{ord_aes()}} can be used to incorporate all
coordinates of the ordination model into a statistical transformation. It
maps the coordinates to the custom aesthetics \code{..coord1}, \code{..coord2}, etc.

Some transformations, e.g. \code{\link[=stat_center]{stat_center()}}, are commutative with projection
to the 'x' and 'y' coordinates. If they detect aesthetics of the form
\verb{..coord[0-9]+}, then \code{..coord1} and \code{..coord2} are converted to \code{x} and \code{y}
while any remaining are ignored.

Other transformations, e.g. \code{\link[=stat_spantree]{stat_spantree()}}, yield different results in a
planar biplot when they are computer before or after projection. If such a
stat layer detects these aesthetics, then the lot of them are used in the
transformation.

In either case, the stat layer returns a data frame with position aesthetics
\code{x} and \code{y}.
}

\section{Computed variables}{
 These are calculated during the statistical
transformation and can be accessed with \link[ggplot2:aes_eval]{delayed evaluation}.
\describe{
\item{\verb{x,y}}{cartesian coordinates (if passed polar)}
\item{\verb{angle,radius}}{polar coordinates (if passed cartesian)}
\item{\code{axis}}{unique axis identifier (integer)}
\item{\verb{lower,upper}}{distances to endpoints from origin (before offset)}
\item{\verb{yintercept,xintercept}}{intercepts (possibly \code{Inf}) of offset axis}
}
}

\examples{
# Freestone primary glass measurements
print(glass)

# default (standardized) linear discriminant analysis of sites on measurements
glass_lda <- MASS::lda(Site ~ SiO2 + Al2O3 + FeO + MgO + CaO, glass)
# bestow 'tbl_ord' class & augment observation, centroid, and variable fields
as_tbl_ord(glass_lda) \%>\%
  augment_ord() \%>\%
  mutate_rows(discriminant =
                ifelse(.element == "active", "centroid", "case")) \%>\%
  print() -> glass_lda
# row-standard biplot
glass_lda \%>\%
  confer_inertia(1) \%>\%
  ggbiplot() +
  theme_bw() + theme_biplot() +
  geom_rows_point(aes(shape = grouping, size = discriminant), alpha = .5) +
  stat_cols_rule(aes(label = name), color = "#888888", num = 8L,
                 fun.offset = \(x) minabspp(x, p = .2),
                 text_size = 2.5, label_dodge = .02) +
  scale_shape_manual(values = c(2L, 3L, 0L, 5L)) +
  ggtitle(
    "LDA of Freestone glass measurements",
    "Row-standard biplot of standardized LDA"
  )

# contribution LDA of sites on measurements
glass_lda <-
  lda_ord(Site ~ SiO2 + Al2O3 + FeO + MgO + CaO, glass,
          axes.scale = "contribution")
# bestow 'tbl_ord' class & augment observation, centroid, and variable fields
as_tbl_ord(glass_lda) \%>\%
  augment_ord() \%>\%
  mutate_rows(discriminant =
                ifelse(.element == "active", "centroid", "case")) \%>\%
  print() -> glass_lda
# symmetric biplot
glass_lda \%>\%
  confer_inertia(.5) \%>\%
  ggbiplot() +
  theme_bw() + theme_biplot() +
  geom_rows_point(aes(shape = grouping, alpha = discriminant)) +
  stat_cols_rule(aes(label = name), color = "#888888", num = 8L,
                 fun.offset = \(x) minabspp(x, p = .1),
                 text_size = 2.5, text_dodge = .025) +
  scale_shape_manual(values = c(16L, 17L, 15L, 18L)) +
  ggtitle(
    "LDA of Freestone glass measurements",
    "Symmetric biplot of contribution LDA"
  )
}
\references{
Gower JC, Gardner--Lubbe S, & le Roux NJ (2011) \emph{Understanding Biplots}.
Wiley, ISBN: 978-0-470-01255-0. \url{https://www.wiley.com/go/biplots}

Gower JC & Harding SA (1988) "Nonlinear biplots". \emph{Biometrika} \strong{75}(3):
445--455. \doi{10.1093/biomet/75.3.445}
}
\seealso{
Other stat layers: 
\code{\link{stat_center}()},
\code{\link{stat_chull}()},
\code{\link{stat_cone}()},
\code{\link{stat_scale}()},
\code{\link{stat_spantree}()}
}
\concept{stat layers}
