% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ord-supplementation.r
\name{supplementation}
\alias{supplementation}
\alias{supplementation_u}
\alias{supplementation_v}
\alias{supplementation_factor}
\title{Supplement ordination objects with new data}
\usage{
supplementation_u(x)

supplementation_v(x)

supplementation_factor(x, .matrix)
}
\arguments{
\item{x}{A \code{[tbl_ord]}, or an ordination object coercible to one.}

\item{.matrix}{A character string matching one of several indicators for one
or both matrices in a matrix decomposition used for ordination. The
standard values are \code{"u"}, \code{"v"}, and \code{"uv"}.}
}
\description{
These functions attach supplementary rows or columns to an
ordination object.
}
\details{
The \code{supplementation_*()} methods produce matrices of supplemental rows or
columns (i.e. rows of \eqn{U} or of \eqn{V}) of a \code{tbl_ord} object from the
object itself. The motivating example is linear discriminant analysis, which
produces a natural biplot of class discriminant centroids and variable axes
but is usually supplemented with case discriminant scores. If no
supplementary rows of a factor are produced, the function returns \code{NULL}.

The \code{supplement_*()} functions return the ordination object with either or
both factors attributed with the result of \code{supplementation_*()}. The
supplementary values are augmented with a \code{.supplement} column whose value
indicates their source and can be incorporated into a \link[=fortify]{tidied form}.
}
\examples{
# Linear discriminant analysis of Reaven & Miller diabetes data
head(heplots::Diabetes)
(diabetes_lda <- MASS::lda(group ~ ., heplots::Diabetes))
dim(supplementation_u(diabetes_lda))
as_tbl_ord(diabetes_lda) \%>\%
  augment() \%>\%
  mutate_u(discriminant = ifelse(! .supplement, "centroid", "case")) \%>\%
  print() -> diabetes_lda
ggbiplot(diabetes_lda) +
  theme_bw() +
  geom_u_point(aes(color = .grouping, size = discriminant), alpha = .5) +
  geom_v_axis(color = "#888888") +
  geom_v_text_radiate(aes(label = .name)) +
  scale_color_brewer(type = "qual", palette = 2) +
  ggtitle(
    "Column-standardized LDA biplot of Reaven & Miller diabetes data"
  )
}
