% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ord-supplementation.r
\name{supplementation}
\alias{supplementation}
\alias{supplementation_rows}
\alias{supplementation_cols}
\alias{supplementation_factor}
\title{Supplement ordination objects with new data}
\usage{
supplementation_rows(x)

supplementation_cols(x)

supplementation_factor(x, .matrix)
}
\arguments{
\item{x}{An object of class '\link{tbl_ord}'.}

\item{.matrix}{A character string matching one of several indicators for one
or both matrices in a matrix decomposition used for ordination. The
standard values are \code{"rows"}, \code{"cols"}, and \code{"dims"} (for both).}
}
\description{
These functions attach supplementary rows or columns to an
ordination object.
}
\details{
The \verb{supplementation_*()} methods produce matrices of supplemental rows or
columns of a \link{tbl_ord} object from the object itself. The motivating example
is linear discriminant analysis, which produces a natural biplot of class
discriminant centroids and variable axes but is usually supplemented with
case discriminant scores. If no supplementary rows of a factor are produced,
the function returns \code{NULL}.

The \verb{supplement_*()} functions return the ordination object with either or
both factors attributed with the result of \verb{supplementation_*()}. The
supplementary values are augmented with a \code{.supplement} column whose value
indicates their source and can be incorporated into a \link[=fortify]{tidied form}.
}
