% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ord-tidiers.r
\name{tidiers}
\alias{tidiers}
\alias{tidy.tbl_ord}
\alias{glance.tbl_ord}
\alias{fortify.tbl_ord}
\title{Tidiers for 'tbl_ord' objects}
\usage{
\method{tidy}{tbl_ord}(x, ...)

\method{glance}{tbl_ord}(x, ...)

\method{fortify}{tbl_ord}(model, data, ..., .matrix = "dims", .supplement = TRUE)
}
\arguments{
\item{x, model}{An object of class '\link{tbl_ord}'.}

\item{...}{Additional arguments allowed by generics; currently ignored.}

\item{data}{Passed to generic methods; currently ignored.}

\item{.matrix}{A character string matching one of several indicators for one
or both matrices in a matrix decomposition used for ordination. The
standard values are \code{"rows"}, \code{"cols"}, and \code{"dims"} (for both).}

\item{.supplement}{Logical; whether to include
\link[=supplementation]{supplementary} points.}
}
\description{
These functions return \link[tibble:tibble]{tibbles} that summarize
an object of class '\link{tbl_ord}'. \code{tidy()} output contains one row per
artificial coordinate and \code{glance()} output contains one row for the whole
ordination.
}
\details{
Three generics popularized by the \strong{ggplot2} and \strong{broom} packages make use
of the \link{augmentation} methods:
\itemize{
\item The \code{\link[generics:tidy]{generics::tidy()}} method
summarizes information about model components, which here are the
artificial coordinates created by ordinations. The output can be passed to
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} to generate scree plots.
\item The \code{\link[generics:glance]{generics::glance()}} method
reports information about the entire model, here always treated as one of a
broader class of ordination models.
\item The \code{\link[ggplot2:fortify]{ggplot2::fortify()}} method
augments and collapses row and/or column data, depending on \code{.matrix}, into
a single tibble, in preparation for \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}. Its output
resembles that of \code{\link[generics:augment]{generics::augment()}}, though rows in the output may
correspond to rows, columns, or both of the original data. If \code{.matrix} is
passed \code{"rows"}, \code{"cols"}, or \code{"dims"} (for both), then \code{fortify()} returns
a tibble whose fields are obtained, in order, via \verb{get_*()},
\verb{augmentation_*()}, and \verb{annotation_*()}.
}

The tibble is assigned a \code{"coordinates"} attribute whose value is obtained
via \code{\link[=get_coord]{get_coord()}}. This facilitates some downstream functionality that relies
on more than those coordinates used as position aesthetics in a biplot, in
particular \code{\link[=stat_spantree]{stat_spantree()}}.
}
\seealso{
\link{augmentation} methods that must interface with tidiers.
}
