% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-wrap.r
\name{wrap-ord}
\alias{wrap-ord}
\alias{eigen_ord}
\alias{svd_ord}
\alias{cmdscale_ord}
\alias{logisticSVD_ord}
\alias{logisticPCA_ord}
\alias{convexLogisticPCA_ord}
\alias{nipals_ord}
\title{Wrappers for lossy ordination methods}
\usage{
eigen_ord(x, symmetric = isSymmetric.matrix(x))

svd_ord(x, nu = min(dim(x)), nv = min(dim(x)))

cmdscale_ord(d, k = 2, add = FALSE)

logisticSVD_ord(
  x,
  k = 2,
  quiet = TRUE,
  max_iters = 1000,
  conv_criteria = 1e-05,
  random_start = FALSE,
  ...,
  partial_decomp = TRUE,
  main_effects = TRUE
)

logisticPCA_ord(
  x,
  k = 2,
  m = 4,
  quiet = TRUE,
  partial_decomp = FALSE,
  max_iters = 1000,
  conv_criteria = 1e-05,
  random_start = FALSE,
  ...,
  main_effects = TRUE
)

convexLogisticPCA_ord(
  x,
  k = 2,
  m = 4,
  quiet = TRUE,
  partial_decomp = FALSE,
  max_iters = 1000,
  conv_criteria = 1e-06,
  random_start = FALSE,
  ...,
  main_effects = TRUE,
  ss_factor = 4
)

nipals_ord(
  x,
  ncomp = min(nrow(x), ncol(x)),
  center = TRUE,
  scale = TRUE,
  maxiter = 500,
  tol = 1e-06,
  startcol = 0,
  fitted = FALSE,
  force.na = FALSE,
  gramschmidt = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{x}{a numeric or complex matrix whose spectral decomposition is to
    be computed.  Logical matrices are coerced to numeric.}

\item{symmetric}{if \code{TRUE}, the matrix is assumed to be symmetric
    (or Hermitian if complex) and only its lower triangle (diagonal
    included) is used.  If \code{symmetric} is not specified,
    \code{\link[base]{isSymmetric}(x)} is used.}

\item{nu}{the number of left  singular vectors to be computed.
    This must between \code{0} and \code{n = nrow(x)}.}

\item{nv}{the number of right singular vectors to be computed.
    This must be between \code{0} and \code{p = ncol(x)}.}

\item{d}{a distance structure such as that returned by \code{dist}
    or a full symmetric matrix containing the dissimilarities.}

\item{k}{the maximum dimension of the space which the data are to be
    represented in; must be in \eqn{\{1, 2, \ldots, n-1\}}{{1, 2, \\dots, n-1}}.}

\item{add}{logical indicating if an additive constant \eqn{c*} should
    be computed, and added to the non-diagonal dissimilarities such that
    the modified dissimilarities are Euclidean.}

\item{quiet}{logical; whether the calculation should give feedback}

\item{max_iters}{number of maximum iterations}

\item{conv_criteria}{convergence criteria. The difference between average deviance
in successive iterations}

\item{random_start}{logical; whether to randomly inititalize the parameters. If \code{FALSE},
function will use an eigen-decomposition as starting value}

\item{...}{Additional parameters passed to original functions.}

\item{partial_decomp}{logical; if \code{TRUE}, the function uses the rARPACK package
to more quickly calculate the eigen-decomposition. This is usually faster than standard
eigen-decomponsition when \code{ncol(x) > 100} and \code{k} is small}

\item{main_effects}{logical; whether to include main effects in the model}

\item{m}{value to approximate the saturated model. If \code{m = 0}, m is solved for}

\item{ss_factor}{step size multiplier. Amount by which to multiply the step size. Quadratic
convergence rate can be proven for \code{ss_factor = 1}, but I have found higher values
sometimes work better. The default is \code{ss_factor = 4}.
If it is not converging, try \code{ss_factor = 1}.}

\item{ncomp}{Maximum number of principal components to extract from x.}

\item{center}{If TRUE, subtract the mean from each column of x before PCA.}

\item{scale}{if TRUE, divide the standard deviation from each column of x before PCA.}

\item{maxiter}{Maximum number of NIPALS iterations for each
principal component.}

\item{tol}{Default 1e-9 tolerance for testing convergence of the NIPALS
iterations for each principal component.}

\item{startcol}{Determine the starting column of x for the iterations
of each principal component.
If 0, use the column of x that has maximum absolute sum.
If a number, use that column of x.
If a function, apply the function to each column of x and choose the column
with the maximum value of the function.}

\item{fitted}{Default FALSE. If TRUE, return the fitted (reconstructed) value of x.}

\item{force.na}{Default FALSE. If TRUE, force the function to use the
method for missing values, even if there are no missing values in x.}

\item{gramschmidt}{Default TRUE. If TRUE, perform Gram-Schmidt 
orthogonalization at each iteration.}

\item{verbose}{Default FALSE. Use TRUE or 1 to show some diagnostics.}
}
\description{
These \verb{*_ord} functions wrap functions imported from other
packages with modifications for use with '\link{tbl_ord}' methods. Some
parameters are hidden from the user and set to settings required for these
methods, some matrix outputs are given row or column names to be used by
them, and new '*_ord' S3 class attributes are added to enable them.
}
\details{
The following table summarizes the wrapped functions:\tabular{llll}{
   Original function \tab Hide params \tab Add names \tab New class \cr
   \code{\link[base:eigen]{base::eigen()}} \tab Yes \tab Yes \tab Yes \cr
   \code{\link[base:svd]{base::svd()}} \tab Yes \tab Yes \tab Yes \cr
   \code{\link[stats:cmdscale]{stats::cmdscale()}} \tab Yes \tab No \tab Yes \cr
   \code{\link[logisticPCA:logisticSVD]{logisticPCA::logisticSVD()}} \tab No \tab Yes \tab No \cr
   \code{\link[logisticPCA:logisticPCA]{logisticPCA::logisticPCA()}} \tab No \tab Yes \tab No \cr
   \code{\link[logisticPCA:convexLogisticPCA]{logisticPCA::convexLogisticPCA()}} \tab No \tab Yes \tab No \cr
   \code{\link[nipals:nipals]{nipals::nipals()}} \tab No \tab No \tab Yes \cr
}
}
